Update flash_phy for outstanding transactions
If a transaction is waiting on a particular bank, accesses to other
banks are allowed to move ahead and complete.

However, flash_phy still responds in order, so transactions completed
ahead of time are held until prior transactions are done.

flash_phy only performs this at the bank level; if there are page
level optimizations it is assumed to be done inside prim_flash