Fix fusesoc dependency breakage using python-requirements.txt
In #302, azure-pipelines.yml was modified to ensure that attrs 19.1.0 is
installed. This is because 19.2.0 breaks FuseSoC via the simplesat
dependency (see https://github.com/enthought/sat-solver/issues/270).
While the edits to azure-pipelines.yml fixed CI, users following the
getting started guide are still going to run into this issue. This
commit addresses the problem in an alternative way - by specifying the
version of attrs to use in python-requirements.txt.

Introducing a constraints.txt could have advantages, but that is more
intrusive and would require further doc updates. I feel that would be
overkill for a quick fix workaround that we hope to remove quite
rapidly.