[prim] Restore ability to select default implementation
When we introduced the primitives library, we also had the possibility
to choose a per-design default implementation for primitives. To set a
default implementation, the synthesis script would set the
PRIM_DEFAULT_IMPL define to values like "xilinx" or "generic", which was
then used as default value for the Impl parameter, which is present in
all primitives modules:

```verilog
parameter Impl = `"`PRIM_DEFAULT_IMPL`"
```

Unfortunately, this syntax to "cast" a define to a string isn't
supported in Cadence LEC. As a stop-gap, we removed the functionality
altogether. This resulted in always using the "generic" implementation
for all targets, which produces suboptimal results e.g. on Xilinx FPGAs.

With this commit, the original functionality is restored in a
LEC-compatible way. Instead of using strings for PRIM_DEFAULT_IMPL, we
now use integers. The allowed values are listed in the
prim_pkg::impl_e enum.

Implementation notes on types:

* PRIM_DEFAULT_IMPL is a define passed through to tools as command line
  options. Most/all tools only support "primitive" types in this way,
  not enums; hence PRIM_DEFAULT_IMPL is an integer, and all assignments
  to it must be integer as well. This includes the default assignment:

  `ifndef PRIM_DEFAULT_IMPL
    `define PRIM_DEFAULT_IMPL integer'(prim_pkg::ImplGeneric)
  `endif

* Since we assign the PRIM_DEFAULT_IMPL define to the Impl parameter,
  this parameter needs to be either an integer, or cast the RHS to an
  enum.

* In this implementation we stayed with an integer type for the Impl to
  avoid potential problems with tools don't supporting enums as
  parameter type. (We don't know of any tool yet that doesn't support
  it, but it seems sufficiently exotic to run into problems down the
  road.)

Thanks to Tim for an initial version of this code.

Fixes #40