[prim] Use prim_pkg::impl_e instead of int for Impl
All target-dependent primitives are supplied with an Impl parameter,
which chooses the implementation target. The potential values for this
parameter are listed in the prim_pkg::impl_e enum, which is based on
int.

This commit switches Impl over to use prim_pkg::impl_e as type, instead
of int. Doing so avoids many casts over the code base, at the cost of
having to pass an enum type name to the tool in the PRIM_DEFAULT_IMPL
define. This approach is standards-compliant, but seems fragile.
Let's give it a try nonetheless, and see how tools react to it.

Tested with
- Verilator 4.019
- Xilinx Vivado synthesis 2018.3
- Synopsys VCS 2019.06

Note for the future: revert this commit if you run into tool issues.