[doc] Update FPGA quickstart instructions
We already had a quickstart.md with instructions to get quickly started
on FPGA and Verilator. When adding the FPGA quickstart instructions, I
didn't remember that and Tim added the instructions instead to a new
document at fpga/quick_start_fpga.md. This PR updates the existing
document with a modified version of the FPGA quickstart that Tim wrote,
in the "original" quickstart place.

This quickstart is currently a skeleton, it needs to be filled in once
we have a RC of the bitstream and tried out the process.