[hugo] Revise the heading style (Demote headers)
Revise the HTML/CSS style to make the title to be a title of the page.
Other markdown headings are demoted in style (h1 tag --> h2 style).
And added h3 style.

In addition to that, now `site/docs/layouts/_default/list.html` is
symlinked to `single.html`.

Markdown files which set the title for `hugo` do not need to have the
title set again. This would result in the header printed twice.

This is related to PR #868

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>