[hw/dv] Reuse existing GPIO FIFO
If the GPIO FIFO already exists in the file system do not abort creation
immediately but continue to try to open it.

When using GDB to debug the Verilator simulation the file is not deleted
at the exit. If the file is not deleted manually the next run will
result in an error. Reusing the file is in this case possible.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>