[hmac] Block the configuration change while processing
There were a few test failures on the sudden CFG register update during
SHA/HMAC process. A few bugs caused system hang. They blocked the
interconnect so that the software cannot release the errors.

Related commits: 9e2abd5 ce32b00

Latest one that @cindychip found at Issue #991 also caused hang
condition inside HMAC. Rather fixing every hang condition, it seems to
be better to block the CFG update in the middle of the process. If the
software wants to update CFG, it can update CFG but it only can be
visible to the hardware after the current HMAC/SHA process is completed.

This PR also blocks the secret key update while the hash engine is
active. Secret key (KEY0 .. KEY7) can only be updated after the engine
completes the computing of hash and before the new message is fed.

If the event occurs (secret key is updated in process), the error will
be reported to the software through the ERROR FIFO.

This is related to #991

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>