[sw] Fix -O3 Infinite Loop in libc functions
A classic problem with trying to implement libc is that, at higher
optimisation levels, compilers will recognise the body of your C library
function implementation as a function it knows about, and replace the
body of that function with a call to the same function. This results in
a recursive function, which does absolutely nothing until it causes a
stack overflow.

Thankfully, GCC and clang provide options so you can disable this
recognition and replacement of libc functions. In OpenTitan's case, we
only need to do this for the `mem_ot` library, as we don't mind if other
libraries have calls to these libc functions inserted into them.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>