Revert "[sw/device] Drop all linker sections not specified"
This reverts commit 0f530aec8d0d9eee9b53d66c039618c78133a643.

Dropping all nonspecified sections from the ELF was done in order to
prevent the .note.gnu.build-id section to appear in the final binary
that we create with objcopy. As a side effect, this also removed all
other sections, like debug sections, making the ELF files unusable for
debugging with GDB (something we document in our user guide).

It turns out, this patch is actually unnecessary. We already suppress
the generation of the .note.gnu.build-id section by passing
|--build-id=none| to the linker.

If (for whatever reason in the future) we want to use the build id, a
better fix would be one of:

- Don't dump the .note.gnu.build-id section into the binary (*.bin) file
  by passing |-R .note.gnu.build-id| to objcopy.
- Modify the linker script to discard this section, or place it
  somewhere more suitable, and ideally give it a label so it can be
  referenced from software. See
  https://interrupt.memfault.com/blog/gnu-build-id-for-firmware for
  write-up.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>