[doc] Ensure install instructions match apt requirements
The `apt-get install` command has, at some points, not been updated with
changes to "apt-requirements.txt" (as happened when CMake was
accidentally added to the dependencies then hastily documented).

This commit allows us to generate the `apt-get install` command based on
the contents of apt-requirements.txt, and include that file directly.
This has changed the listed order of packages, but ensures the
documentation stays up to date when apt-requirements.txt changes.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>