[doc] Clarify C Inline Function Style
This change promotes the guidance about C11's inline functions into its
own section, instead of being part of the section about builtins (which
now contains a link to the inline guidance).

The only addition I have made is to be explicit that we do not want to
use `inline` on `static` C functions.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>