[sw, dif_uart] Fix dif_uart_byte_send_polled() routine
UART_STATUS_TXEMPTY means that the HW took the last data element of
the FIFO for processing. UART_STATUS_TXIDLE is set when there is no
activity on the UART lines. Combination of these two flags can be used
to determine whether the HW has completely processed the last element.

NOTE: polling operation guarantees that the requested byte is
      processed completely on return from the function. This means
      that we must check whether the FIFO is drained to insure that
      this is the case.

Fixes issue in PR #1510.

Signed-off-by: Silvestrs Timofejevs <silvestrst@lowrisc.org>