[aes] Disable padding in crypto (OpenSSL/BoringSSL) interface
By default, padding is enabled in the used EVP functions of
OpenSSL/BoringSSL. Both the sizes of the NIST test data sets we are using
so far and the sizes of our tests are multiples of the block size. No
padding is actually required. However, to make the padding checks in the
EVP decryption functions succeed, workarounds are required.

This commit disables the padding and padding checks in the EVP functions
and removes the corresponding workarounds for the padding checks.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>