[dvsim] Generate seeds without the secrets module
The `secrets` Python module is currently not listed as dependency in the
`python_requirements.txt` file. It is also tricky to install with `pip`
as it contains binary components that depend on OpenSSL being present
on the system.

To simplify things, this commit switches to using the `random` core
Python module to generate random seeds. There is no requirements for
simulation seeds to be cryptographically secure, so this change
shouldn't affect the quality of our simulations.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>