[sw, tock] Reduced binary size
The Tock debug binary with the .apps section is ~200KB. This takes
upwards of 3m30s to load onto the FPGA using spiflash.

To reduce the size of the binary let's remove the .apps section (since
we're not loading any apps at the moment) and enable lto. This brings
the binary down to ~57KB and allows spiflash to finish in under a
minute.

Signed-off-by: Jon Flatley <jflat@google.com>