Add script to generate technology-dependent primitives on the fly
Some primitives are intended to be implemented in a technology-specific
way, e.g. differently for Xilinx or Altera FPGAs. Which implementation
is to be used can be specified at synthesis time. Target-specific
implementations are collected in "technology libraries", or "techlibs".
When writing code, only "abstract primitives" are used. These are then
implemented by a techlib.

Currently, all techlibs are part of the OpenTitan repository, and all
abstract primitives are hand-crafted. This commit adds a script to
auto-generate these abstract primitives. The script is to be used as
FuseSoC generator.

Note that the script currently is in a rather "rough" form. The
SystemVerilog "parsing" is done using a regex instead of a proper
parser, and speed isn't great either.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>