[tlul] Match tlul_adapter_sram Outstanding to value
Reduce the number of Oustanding to its parameter value. Previous design
increased the request outstanding capability by one to accept next
request when it processes current request. It creates assertion error
below:

    `ASSERT(rvalidHighWhenRspFifoFull, rvalid_i |-> rspfifo_wready)

It is debatable whether the second request should be allowed or not.
This commit is to limit the request when the number of outstanding
requests reaches to its max value. Or we could revise the assertion to
consider the case.

This is related to #1489

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>