[sw] Create a temporary location for a rewrite of boot_rom
The current boot_rom currently cannot provide many of the features the
OpenTitan mask ROM will eventually need. Rather than attempting to
refactor the current ROM and risk breaking everyone, we'll do a complete
rewrite (which isn't too bad, since the current boot_rom is tiny).

The code in the new location is everything necessary to get to a
hello-world at boot.

Signed-off-by: Miguel Young de la Sota <mcyoung@google.com>