[fpv/keccak] FPV wrapper for Keccak
This commit implements simple FPV testbench that checks keccak
permutation. It assumes the keccak logic is used in SHA3-256 algorithm
and the given input value is NULL (zero length message). So the
bitstring feeded into the keccak is "" || 01 ||10*(1084)1 || 0*(512),
which are respectively M, SHA3 domain, pad10*1, and capacity.

More concrete tests are necessary to confirm the logic works.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>