[sw, libbase] Create non-volatile bit manipulation library
This change was suggested by Miguel Young, and aims to create a set of
bit manipulation routines that can be used across the project.

The functionality added in this change can be used in mmio library,
which will simply wrap the non-volatile versions. This would make it
easier to create a set of expectations in the mock MMIO library to
mimic the mmio library exactly.

There could be other use-cases outside of DIF that could benefit from
non-volatile bit manipulation API.

Signed-off-by: Silvestrs Timofejevs <silvestrst@lowrisc.org>