Specify minimum versions of Python dependencies
We install the fusesoc and edalize Python dependencies from
Git repositories which contain OT-specific modifications (potentially;
currently, edalize is unmodified, only fusesoc is patched). The
requirements in python-requirements.txt were written under the
assumption that a call to `pip install -r python-requirements.txt` would
aways install the latest version as specified by the branch argument.
This assumption is wrong, as documented in
https://pip.readthedocs.io/en/latest/reference/pip_install/#git:

"For non-editable installs, the project is built locally in a temp dir
and then installed normally. Note that if a satisfactory version of the
package is already installed, the VCS source will not overwrite it
without an --upgrade flag. VCS requirements pin the package version
(specified in the setup.py file) of the target commit, not necessarily
the commit itself."

It's still hard to specify the exact version we require and users are
encouraged to always use the `--upgrade`/`-U` flag with pip to be on the
safe side. Nonetheless, we can be slightly more specific by adding the
minimum version we need for fusesoc and edalize.

Fixes #1858

Signed-off-by: Philipp Wagner <phw@lowrisc.org>