[sw] Fix Software Include Guards
Recently a bug was found in CI where the code formatting hooks were not
firing correctly. This bug was fixed in
lowrisc/opentitan#8736c3b307b2a5f30514ddfabc9342b35edb9adb.

This change ensures that the include guards on existing code are
correct, so that going forwards we know the software tree is and remains
correct.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>