[topgen] Generate PLIC Enums and Peripheral Mapping
Until now, the PLIC DIF has used a handwritten set of enums to produce
human readable names for interrupt ids, and handwritten arrays to map
these back to peripherals.

This commit now auto-generates this information, based on the interrupt
information in `top_earlgrey.hjson`, and adds it to `top_earlgrey.h`.

The following enums are generated:
- `top_earlgrey_plic_irq_id_t`: interrupt ID enum, from
  `top['interrupt']`;
- `top_earlgrey_plic_peripheral_t`: interrupt peripheral (source) enum,
  `from top['interrupt_module']`,
- `top_earlgrey_plic_target_t`: interrupt target enum, from
  `top['num_cores']`

This also generates a mapping from an interrupt ID to the peripheral
that it came from, called `top_earlgrey_plic_interrupt_for_peripheral`
and implemented using a const array. This maps from
`top_earlgrey_plic_irq_id_t` to `top_earlgrey_plic_peripheral_t`. This
is generated into a .c file (as its definition cannot be put into a
header), and a library is created called `top_earlgrey_ot` to contain
it.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>