[util] vendor.py now re-clones by default
This commit adds the `--update` command line flag to `util/vendor.py`.

Before this change, the vendor tool would always download the revision
specificed in the `.vendor.hjson` file, and if that pointed to a branch,
the repository would be updated with all the upstream changes.

After this change, users have to explicitly pass `--update` to get this
behaviour.

Now, by default, the vendor tool will re-clone the upstream repository
as specified in the `.lock.hjson` file, and apply any new patches or
file exclusions to what should be the existing version of the
repository. This means you can easily re-vendor existing versions of
repositories without having to integrate upstream changes.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>