[doc] Remove doc build instructions from README
These instructions feel a bit out of place in the README, which is the
first impression users get from the OpenTitan project if they go to
GitHub. The online documentation at docs.opentitan.org is up to date
with the master branch; self-building documentation is only necessary to
actually test documentation changes, which is rather rare.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>