[dvsim] Tidy up wave dumping logic
Before this patch, you had to run one of:

  dvsim.py ... -w -d fsdb   (Verdi FSDB dumping on VCS or xcelium)
  dvsim.py ... -w -d shm    (Native SHM dumping on xcelium)
  dvsim.py ... -w -d vpd    (Native VCD+ dumping on VCS or xcelium)

This seems like unnecessary work for the user (and a VCS simulation
just hangs if you use the default 'fsdb' dump type without a Verdi
licence).

With this patch, dvsim.py guesses the format to use by looking on
$PATH for verdi. On success, it uses fsdb. On failure, it uses vpd for
VCS and SHM for xcelium. If you want to override the automatic choice,
you can still pass -d, but most of the time you shouldn't have to.

The patch also adds some more error checking (you can no longer
specify --waves when running Verible, for example) and adds support
for vpd dumping on xcelium.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>