[sw] Move Interrupt Vector Definition into CRT
The interrupt vector is defined in `ibex_interrupt_vectors.S`, and
expects the declarations in `sw/device/lib/handler.h` to be defined
(those in `sw/device/lib/hander.c` may be overriden by individual
executables).

However, if `ibex_interrupt_vectors.S` is not a direct dependency of an
executable, its contents are lost during linking and an interrupt vector
is never defined, which causes runtime issues, as noted in #2162.

Given the CRT initialises the value of `mtvec` to the address of the
vector in `ibex_interrupt_vectors.S`, it would make sense for this
dependency not to be broken, and for `ibex_interrupt_vectors.S` to be
part of the `riscv_crt` meson dependency, rather than the interrupt
library. This still allows the interrupt handlers to be overriden.

In order to help diagnose issues with the interrupt vector being dropped
during linking, the interrupt vector is now referred to using the
assembler global symbol (rather than one from the linker script), and
all 32 entries are provided (most are the instruction `unimp`), so that
the `.text` section will not overlap with it.

As a consequence of this change, the `riscv_crt` meson dependency now
requires at least one set of interrupt handler definitions, regardless
of whether you want to use them in your executable. This commit has had
to add the `sw_lib_irq_handlers` library to some existing executables.

The assembler was also needlessly padding the `.vectors` section, so an
argument has been provided to prevent this behaviour.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>