[util] Change Hugo binary usage
If a Hugo extended binary is installed globally the current approach
will try to use this file.
With different Hugo versions not all behave in the same way and newer
versions might fail to build where the "local" downloaded version will
work.

Always download and use a local binary.

An option `--force-global` is added to allow the usage of a Hugo binary
which must be accessible from the environment.
The version check is not executed in this case.

Signed-off-by: Tobias Wölfel <tobias.woelfel@mailbox.org>