[util] Add support for multiple src/dst directories to vendor.py
This patch adds the supporting code and also some documentation in
vendor_in_tool.md. The first use case is actually in Ibex, whose DV
code needs various things that are currently in OpenTitan. Here is a
vendor file that can be used there:

    {
        name: "lowrisc_ip",
        target_dir: "lowrisc_ip",

        upstream: {
            url: "https://github.com/lowRISC/opentitan"
            rev: "master"
        }

        mapping: [
            {from: "hw/dv/sv/common_ifs", to: "common_ifs"},
            {from: "hw/dv/sv/csr_utils",  to: "csr_utils"},
            {from: "hw/dv/sv/dv_lib",     to: "dv_lib"},
            {from: "hw/dv/sv/dv_utils",   to: "dv_utils"},
            {from: "util/dvsim",          to: "dvsim"},
            {from: "util/uvmdvgen",       to: "uvmdvgen"}
        ]
    }

Note that we don't currently support mapping and patch_repo at the
same time. To do so, you would have to run the 'git format-patch'
calls in _export_patches() in each subdirectory. Probably not too
difficult, but complicated enough to punt on until we have a user.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>