[prim] Add a "clog2 width" function
`$clog2()` is commonly used to get the number of bits needed to
address a certain number of items, aka the width of a signal.

There's on special case that's often missed: `$clog2(value)` returns 0
if `value == 1`, leading synthesis failures for code like
`logic [$clog2(value)-1:0] signal`.

This commit adds a helper function to cover the `value == 1` case
better by returning `1` in this case.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>