[site] Create earlgrey structure
For a structured representation of the `top_earlgrey` documentation a
branch node is needed. This allows all contained documentation to be
grouped together, creating a cohesive structure.
The Github preview will not work anymore, but the content in this place
adds not much value.
Make doc content a leaf so it will not mistakenly be highlighted as
having a subsection.

Signed-off-by: Tobias Wölfel <tobias.woelfel@mailbox.org>