[prim_ram_1p_scr] Add a memory scrambling draft implementation
This adds a draft implementation of the scrambling device for the data
memory. The module is implemented as a primitive, in the same spirit as
similar wrappers such as prim_ram_1p_adv. Hence, it can be conveniently
instantiated in comportable IPs (e.g. in OTBN) or in top_earlgrey for
the main system memory.

Note that this design is not final, and its main purpose is to be
instantiated in Bronze for area and timing estimates.

Signed-off-by: Michael Schaffner <msf@opentitan.org>