Extract OTBN ISA documentation into a YAML file
The idea is that this file is more "machine readable". We'll put other
information about the instructions into it (which might not appear in
the documentation, but will affect things like the random instruction
generator).

As well as taking the current text from _index.md, this patch contains
code in insn_yaml.py to parse the YAML file, doing various validity
checks on the result. For example, it makes sure that any syntax
example for an instruction has the same operands as the list of
operands. These checks caught several errors in the
hand-written (hand-copy/pasted!) text.

The parsed file can then be transformed with yaml_to_doc.py to
generate Markdown content suitable for our documentation. This is now
run in build_docs.py and wired in to Hugo with a simple shortcode.

I've manually compared the generated documentation with what was there
before and it's basically the same. Intentional differences:

  - Enumeration operands now list their values with a nested table.

  - BN.MULQACC's qwsel operands are now listed as immediates, rather
    than enumerations. They were enumerations of the form ".0", ".1",
    ".2", ".3". Now the syntax contains the dot, so they can just be
    imm2.

  - The increment options for BN.LID and BN.SID give their required
    syntax.

There are probably further improvements we could make to formatting,
but I thought it made sense to get this in sooner rather than later as
a starting point.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>