[tests] Wrap up the test main() function into a library
This change makes it so the old `main.c` under `sw/device/tests` lives
alongside its header, `test_main.h`, forming a complete library that
pulls in all the relevant dependencies. This makes the linking process
for an on-device test more straightforwrd.

`main()` doesn't actually need to be in one of the main object files, as
far as our linker scripts care; our CRT requests a `main` symbol, and
that gets pulled in regardless of where it's defined.

Signed-off-by: Miguel Young de la Sota <mcyoung@google.com>