[aes] Clear all reg status trackers on writes to control reg
With this commit, any write to the control register will clear the status
tracking for the initial key, IV, data input and data output registers. As
a result, software has to provide a fresh key, IV, input data whenever the
control register is updated and the unit is operated in automatic mode.
Without this commit and unless software initiated a clear of these registers
with random data, it is possible to:
- re-use the key of a previous message,
- re-use the last content of the IV registers (can be the CTR value) of the
  previous message.
Note that the actual registers are not cleared upon a write to the control
register. This clearing with random data should be initiated by software.

This is related to lowRISC/OpenTitan#2913.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>