[topgen] Add a function to find other side ports
The function that finds the other side of port connected through
inter-module signal is useful.

The `find_otherside_modules()` receives top-level cfg Hjson and the
current module instance name, signal name. Then it searches every
connections and returns a list of module name, signal name pairs if
exists.

This commit is related to #3040 #2464

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>

Co-authored-by: Eunchan Kim <eunchan@google.com>
Co-authored-by: Rupert Swarbrick <rswarbrick@lowrisc.org>