[doc/security] Add Device Life Cycle Specification
This commit converts to markdown the Device Life Cycle specification
referenced in RFC lowrisc/opentitan#1294.

This document describes the OpenTitan device operational states supported for
manufacturing, provisioning, regular operation and return to
manufaturer.

The document was converted to markdown by a GDoc plugin and formatted by
a md-linter. Since the maximum line width is set to 80cols, in most
cases a sentence change will trigger a full paragraph reformat/diff.

The GDoc to Markdown pluging generates tables in HTML format, which in
this case is useful to retain the per-cell text formatting and the rows
and columns layout.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>