[dvsim] Add support for second-level indirection
This adds support for a second level of indirection with the key-value
pairs in the HJson.

For example: lets say we supply the dict:
`{var: '{{foo}_xyz{bar}}', foo: p, bar: q, p_xyz_q: baz}`

Then after the substitutions above: `{var: '{p_xyz_q}', ...}`
We need to now substitute {p_xyz_q}, so that the final value of
`{var: 'baz', ...}`.

We will use this mechanism to set compile time values to keys that need
to be differentiated for each build. An example of where this will be
required to be done is setting different coverage collection hierarchy
files to different builds for the same DUT. In a subsequent PR, CSR
tests will have its own build and the coverage collected on those tests
will be significantly reduced. That PR will make this change probably a
bit clearer.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>