[otbn] Infer operand widths from encoding if there is one
If an instruction has an encoding, we check that any operand width
either specified or implied in the list of operands matches the number
of bits in the encoding.

This patch propagates information the other way, so that every operand
of an instruction with an encoding ends up with a width.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>