[dv/top] Temp disable pinmux assertion
Pinmux has an assertion to ensure that `AonWkupReqKnownO` is not X.
However `spi_device.sdi_i` could be X if simply means no input data.
If `wkup_padsel` selects the SDI signal, which is connected to MIO input
`IO_DPS1` if `jtag_en=0`.
Since pinmux will be re-write soon, we are temp disable this assertion
in all the automation tests. It will be removed once the pinmux is
rewritten.

Signed-off-by: Cindy Chen <chencindy@google.com>