[otp_ctrl/doc] Initial commit of OTP spec
This is the initial commit of the OTP spec. Note that there are
still a couple of TODOs in the Markdown file that have to be
worked out in more detail (e.g., the Life Cycle interface, some FSM
details, as well as links to docs that do not exist yet).

Also, we are still in the process of getting more clarity on some
questions about the actual OTP macro used in production, and that may
affect error codes and the OTP wrapper open-source interface.

These updates will follow soon in subsequent commits.

Signed-off-by: Michael Schaffner <msf@opentitan.org>