[aes] Disable SVAs checking key length/mode in case of storage errors
In the case of storage errors in the shadowed control register, the AES
module might be confronted with invalid key length and mode values (such
as all-zero values or if multiple bits are set). Since such storage
errors need to be handled by the RTL, the SVAs checking the key length
and mode values seen by the module must be disabled in case of storage
errors.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>