[doc/security] Ownership Transfer
Introduce OpenTitan Ownership Transfer. The Silicon Owner is defined as a
logical entity or entities allowed to sign code for execution, as well as to
sign ownership management commands. This document define the ownership
assignment/unassignment flows.

The document was converted to markdown by a GDoc plugin and formatted by
a md-linter. Since the maximum line width is set to 80cols, in most
cases a sentence change will trigger a full paragraph reformat/diff.

The GDoc to Markdown pluging generates tables in HTML format, which in
this case is useful to retain the per-cell text formatting and the rows
and columns layout.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>