[otbn] Further tweaks to assembly syntax
Apparently, I didn't check commit 6b1eab2 carefully enough. We
definitely want to strip spaces from the right of the operand string,
but not from the left. Or at least, not without doing some other work.
The problem is the syntax for mulqacc, which looks something like
this:

    [<z>] <wrs1>.<wrs1_qwsel>, <wrs2>.<wrs2_qwsel>, <imm>

If you strip the input string completely, an instruction like this

    BN.MULQACC w28.1, w29.0, 64

means we try to match "w28.1, w29.0, 64". The problem is that the
current matching logic requires the space after the optional ".Z".

Maybe the proper fix is to change the matching regex to be cleverer
and ignore whitespace between optional and required syntax, but that's
not completely trivial. For example, this syntax:

   <foo> [<bar>] <baz>

probably should need a space between <foo> and <baz> if <bar> isn't
supplied.

While I looking at this, I've also fixed a few more syntaxes for the
bignum instructions, removing stray spaces. This is needed to parse
correctly if we strip on the right for a similar reason. If the syntax
is:

    <wrd>, <wrs>, <imm> [, FG<flag_group>]

and you use the form without a flag group, the matching code expects
some trailing whitespace. Before stripping the input, there was a
newline that matched nicely. Afterwards, it failed. But I think it's
perfectly reasonable for that to fail, and the syntax

    <wrd>, <wrs>, <imm>[, FG<flag_group>]

looks nicer in the docs anyway.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>