[otbn] Remove stale TODO messages
These got added by commit 5159a2e, which moved tracking of the "shift"
of an operand from the encoding scheme to the operand itself. At one
point, I had a plan where the encoding scheme and operand would both
have a "shift", and left this note to remind me to add a check that
they matched.

Of course, that was a terrible design(!), so I removed the shift from
the encoding scheme entirely. But I forgot to remove the (now-stale)
TODO messages.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>