[otbn] Teach tooling about PC-relative operands
The branch (beq, bne) and direct jump (jal) instructions from RISC-V
compute a destination based on the current PC. The encoding holds the
offset, so a jump to PC 128 starting at PC 64 would encode an offset
of 64. But the assembly syntax looks like the absolute destination, so
the instruction would just look like

    JAL  x0, 128

The tooling didn't really understand that before. It didn't really
matter, because both the assembler and objdump worked by punting to
binutils (which gets it right).

Of course, the random instruction generator does need to understand
what's going on. Before this patch, it correctly figured out that it
needed to encode the offset, but it printed out that offset in the
assembly rendering. So the example above would have appeared as

    JAL x0, 64

To fix this correctly, we either need to print out the first
line ("JAL x0, 128") or print out something that's explicitly
pc-relative, like "JAL x0, . + 64".

This patch goes with the first option: we know our PC anyway, and it's
a bit easier to read the generated assembly if everything is absolute.

Note that the patch changes otbn-objdump, but the calculated pc will
never actually be used because the instructions that would use
it (jal, beq, bne) are all disassembled by the binutils assembler
anyway.

Similarly, ImmOperandType prints a pc-relative operand in the ". + 64"
format if it doesn't know the current PC. Again, this doesn't actually
happen (because we happen to know that the only call site that might
use it is LI support in otbn-as), but it seemed more sensible to cope
with this case just in case we want to add another pseudo-op to the
assembler in future. Otherwise, you just end up with the same amount
of work in adding defensive assertions.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>