[dv/doc] update make command to dvsim
In the "getting_started_with_dv" doc, we are referring to run DV with
`make` command. This is out-of-dated and not supported anymore. This is
the initial effort to introduce `dvsim` and make sure outside users are
able to run DV with the latest dvsim flow.

This PR solves the issue tracking in issue #3306

Signed-off-by: Cindy Chen <chencindy@google.com>