[otbn] Fix TOC headings for OTBN ISA documentation
It seems that you can't use readFile in a shortcode to read in
Markdown contents and then have them appear in the top-level table of
contents. See [1] for another disappointed customer!

Fortunately, we don't really need it because we only have two headings
we care about. So now the yaml_to_doc.py script generates a Markdown
fragment for each instruction group, which gets included explicitly by
the documentation. It feels a little less magic, but hey ho.

[1] https://github.com/gohugoio/hugo/issues/6690

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>