[dv/common] TLUL same cycle response
As discussed a#3208, add 2 knobs and supports in TLUL agent
1. support_device_same_cycle_response, device may respond d_valid in the
same cycle as a_valid. This is turned off in opentitan
2. support_a_valid_high_d_ready_low, spec requires d_ready=1 when
a_valid=1, Add this knob as our design doesn't obey this rule. This knob
is randomized

This PR won't change current TULUL behavior but only increase the chance of
`a_valid=1 & d_ready=1`.

Signed-off-by: Weicai Yang <weicai@google.com>