[otbn] Better specify "result" when setting flags
The L, M and Z flags operate on the "result" of an operation. Clarify
that this means only the "lower" 256 bits as they are written back to
the destination register, not the overflow bit (which is present in the
carry flag).

Fixes #3336 (See also this issue for a long reason by Felix Miller why
this is the correct behavior.)

Signed-off-by: Philipp Wagner <phw@lowrisc.org>