[OTBN] Document behavior of CSRRS and CSRRW
These instructions were meant to be following the RISC-V behavior of
instructions of the same name, but this wasn't documented. Add this
missing piece of documentation, and also the operation description for
the instruction. (This is actually something we've not done for other
RV32-inspired instructions, but something we wanted to get done as we
move on, so let's start here.)

Signed-off-by: Philipp Wagner <phw@lowrisc.org>