[OTBN] Mark OTBN as being in the D1/Functional design stage
Work on OTBN has been ongoing for a while now and we're making rapid
progress in the specification and design efforts. To reflect that, we
propose to move the design into the D1 (Functional) design stage. See
https://docs.opentitan.org/doc/project/development_stages/#hardware-design-stages
for a detailed explanation of the design stages.

Importantly, the design stage only refers to the v0.1 version of the
design. This version is not ready for production silicon, but an
intermediate milestone that allows us to synchronize implementation,
verification, simulator, and software development work. There will be
future revisions to the specification. The Google Doc [OTBN Enhancement
Proposals](https://docs.google.com/document/u/3/d/1rgtPlXiJo21dq7gL_QZdV2XPISR78hkoLhJV8f1pt5M/edit)
lists high-level proposals to evolve the design, which will be addressed
in a subsequent version.

In short, the D1 design stage indicates that the design "does something,"
but isn't complete yet. Looking at the individual items in the hardware
design stages documentation:

* **Feature set finalized, spec complete.** We have identified a
  feature set for the v0.1 version which is a very solid starting point
  for fine-tuning the ISA in various ways. This feature set is
  documented well in the OTBN specification. We had success in running
  larger chunks of crypto code in a simulator implementing this
  specification, serving as a proof-point that the design is going in
  the right direction.

* **CSRs identified; RTL/DV/SW collateral generated**. All CSRs have
  been documented in the HJSON file, and regtool has generated
  appropriate bindings for it. UVM-based CSR tests run in CI and nightly
  regressions.

* **SW interface automation completed.** A DIF has been written and
  merged, which makes use of all auto-generated header files, e.g. to
  obtain information about the register layout.

* **Clock(s)/reset(s) connected to all sub modules.** This is the case.

* **Lint run setup.** Lint is setup and runs in CI and in nightly lint
  runs.

* **Ports Frozen.** OTBN has only a small set of top-level ports, which
  are fully hooked up into the earlgrey design.

The RTL implementation of the OTBN design has started, and we continue
to evolve it by implementing more and more instructions, as well as
improving the implementation of existing ones.

Fixes #3361

Signed-off-by: Philipp Wagner <phw@lowrisc.org>