[test] Add system tests running on a Nexys Video board
Run a basic set of sanity tests on FPGA: flash the FPGA with the
bitstream, load software over SPI with the spiflash tool, connect to the
UART device, and check if the software does what it should do.

For this first PR we run all self-checking software tests, as we do on
the Verilator simulation target (minus the usbdev one, which isn't
supported on FPGA, as it needs the USBDPI module).

To run the FPGA tests only, run

```
pytest -sv --log-cli-level=DEBUG test/systemtest/earlgrey/test_fpga_nexysvideo.py
```

Signed-off-by: Philipp Wagner <phw@lowrisc.org>