[otbn] Connect ECALL up to the external register model
This gets rid of the environment (since there isn't really an
environment in OTBN) and stops deriving from riscvmodel's Model, since
we aren't really using anything from there.

Our state now has a "running" flag, which is essentially the same as
~otbn.STATUS.busy, but a bit easier to keep track of. We also plumb
that into the stepping logic: now, stepping a model that has stopped
doesn't do anything (rather than executing rubbish after the ECALL
instruction).

To track whether we're running or not, we also rename "jump" to
"start" in the stepped model and make it set the PC and then kick off
the simulation. Maybe we should get the start PC from an external
register (and expose setting/getting the external registers over TLUL
on stepped.py's stdin/stdout interface), but that's probably a job for
a future commit.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>