[top] Provide top/target-specific parameters for AES
This commit provides top/target-specific parameters for the AES unit.
The selection of the S-Box implementation and Masking hast a high
impact on FPGA resource utilization/ASIC area, the trigger delay
should only be used when performing SCA analysis.

This is just a first step. Ultimately, some of these parameters will
be driven through fusesoc, which will allow us to consolidate some
of the top levels.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>