[otbn] Fix branch address for BEQ/BNE
The immediate operand is PC-relative, but the "+ pc" operation happens
as part of decoding, so we don't need to add it on again as part of
the instruction body.

Maybe we should rename the operand to something like "dest", since the
assembly code is written with the destination, rather than the offset.
But let's leave that for later, sticking with the RISC-V naming for
now.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>