[doc] Don't strip markdown headings from HW checklist
The checklist template is processed by Mako, which uses `##` as template
comment. Use syntax which avoids stripping the headings out.

(The Mako comment behavior cannot really be configured; we could write a
pre-processor which does the escaping, but since we're only having a
single template here and many invocations of Mako in different scripts,
doing that seems fragile. See also
https://github.com/sqlalchemy/mako/issues/268.)

Signed-off-by: Philipp Wagner <phw@lowrisc.org>