[memutil] Use exceptions to simplify error handling
Rather than printing to stderr and returning false, the internal
functions now throw an exception, which percolates up as far as
ParseCLIArguments.

We also get rid of the explicit advance check for whether we can open
a file: it's always racy to do that sort of thing and if we can't open
the file, we'll hit an error soon enough!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>