[otbn] Fix BN.LID / BN.SID in ISS
Now, BN.LID / BN.SID load and store 256-bit little-endian values (so
byte 0 is the LSB; byte 31 is the MSB). As before (and as expected by
the RISC-V base ISA), the LW / SW instructions load and store 32-bit
little-endian values. Before this change, the code was a mess of
reversed() calls. After lots of debug prints, I think it's finally
both consistent and right!

This patch also updates the post-increment behaviour as discussed in
issue #3587 and expands the documentation a bit.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>