[fpga] Fix clock constraint
For some reason, Vivado moves one of the clock dividers inside clkmgr
to the outside of clkmgr. As a result, the path used to find the
relevant cell is wrong and the divided clock cannot be properly
constrained. This leads to a bunch of critical warnings.

This commit fixes the path such that the clock can be constrained.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>