[spiflash] Insert additional delays for robustness, increase SPI clock
This commit adds two additional delays to the SPI Flash tool:
- Upon receiving the first frame, the device is expected to erase the
  internal Flash. The host should wait for this operation to complete
  before requesting the hash check.
- On the host side, we need to add a delay between receiving the data
  over SPI and actually using it.

With these two changes, the tool is much more robust and we can
raise the SPI clock frequency back to 1 MHz.

In addition, this commit fixes some comments and variable names (all
delays are given in microseconds instead of nanoseconds.)

This is related to lowRISC/OpenTitan#3606.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>