[dv/alert_handler] fix two errors in alert_handler
1. In alert_handler sender_driver, there could be a race condition to when
reset issued. So to guarantee the reset can release all semaphores, this
PR moves the release function after reset is deasserted.

2. In alert_handler_sanity_test, this PR changed the order to predict
`lock` condition right after issued `do_lock` task. In current sequence,
the prediction happened after triggering alert and escalation, this
could be too late if we have any reset or ping that disable the thread
in the middle.

Signed-off-by: Cindy Chen <chencindy@google.com>