[reggen] Improve names when unpacking field bits
No functional change, but this patch names the elements in the tuple
that comes back from the call to bitmask, which makes the code a bit
easier to follow.

As a plus, the check that field_msb fits in the register width doesn't
need the explanatory comment any more, because the variable names make
it clear what's going on.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>