[reggen] Avoid "enum too big for checking" warnings
The code was avoiding generating a large object (enum_mask has size
exponential in the number of bits in the field) by turning off the
check when the field is large. Instead of doing that, we can generate
an object that's linear in the number of things we've seen and avoid
printing the warning.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>