[dvsim] Create jobs with dependencies instead of sub-jobs
This looks a bit more like normal schedulers, where you have a big
load of jobs with dependencies between them. One reason to do this
sort of thing is that you simplify the process of actually running
things (nothing new to kick off).

The main visible change here, however, is that if a build fails then
we don't run every dependent job for a second before giving up on
it. (We could have retro-fitted that into the existing design, but I'm
trying to move things towards a more "standard" shape as I go). With
the new code, when we're about to dispatch a job, we check to see
whether its dependencies have all run successfully. If not, we kill
it.

Note that this logic is made easier because dvsim runs in
phases (called "targets" in the code). We ensure that dependencies are
always in an earlier phase, so know that they will have run to
completion or failed before any dependent job is started. In code,
this is the assertion that dep.status is P, F or K in
Scheduler.dispatch().

The only other change in this patch is to printing. Because we now
have jobs for future phases/targets, we don't want to print both a
"[build]: ..." and a "[run]: ..." line each time. To avoid that, we
skip targets where everything is still queued if we've printed
something for a previous target.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>