[mask_rom] Fix SP Initialization
I failed to initialize the stack pointer correctly in the initial
Mask ROM CRT, and only just noticed. This confusion comes from the fact
that I unified `_start` and `_end` to always go from lowest address to
highest, and forgot that in RISC-V the stack grows downwards.

In the other linker scripts, `_stack_start` is the highest address on
the stack, but not in Mask ROM (or ROM_EXT).

Signed-off-by: Sam Elliott <selliott@lowrisc.org>