[reggen] Enum Value Definition Fixes
Reggen has been generating enum values already, directly using the
definitions from the hjson. This causes issues when binary literals are
included in the string, as we don't allow those in C/C++ code, so this
change moves to emitting all values as hexadecimal values.

The other change this makes is to add '_VALUE_` between the register
field name and the enum name, so that there are no collisions between
the offset/mask definitions and the enum values.

Closes #3804.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>