[memutil] Add iterator and merging insertion interfaces to RangedMap
If you want to use a RangedMap to represent chunks of memory (say),
you will need to be able to merge chunks. Given a merging function,
the new Emplace method calls it as necessary to glob things together.

This also exposes an iterator interface to allow iteration over the
contents. Of course, once we have an iterator interface, the Find()
function looks a bit silly, and should probably look more like the C++
standard library find() functions (returning an iterator that might be
end()). So we do that too.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>