[dvsim] Fix HJson bugs
- The `exports` key is meant to have a list of key-value pairs (dicts).
Instead, currently they are all parsed as strings instead. This fixes
the exports key and thereby enables the removal of `FlowCfg::process_exports()` subroutine.
- The values in the exports var also contains variable substitutions
(what we are referring to as wildcards) encapsulated in braces without
quotes - this is not allowed in HJson. Values that begin and end in
braces must have quotes, otherwise the braces end up getting treated as
HJson punctuation syntax. It is working currently because it treats the
complete item in a list (separated by comma) as a string.
- As with the first bullet, the `process_exports()` is removed from the
code.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>