[dv] Tidy up use of get_normalized_addr
The get_normalized_addr function is defined in cip_base_scoreboard and
is used to round down word addresses and (seemingly) to convert byte
offsets (as reported by the TL agent) into absolute addresses matching
the RAL model.

However, it turns out that this isn't actually how we were using
it (which has had me *very* confused!). In fact, most of the addresses
floating around are absolute addresses (with occasional exceptions in
code like spi_device_env_cfg and some of the hmac code). The
definition of get_normalized_addr threw away the upper bits of the
address (converting to an offset), rounded down to a multiple of 4,
and then added the base address back on again.

This patch splits up that function into two functions:

  - align_to_word_addr  (rounds down to a multiple of 4)
  - offset_to_addr      (adds default map's base address to an offset)

and updates the code to use the right one in each situation.

Now (I think) everything called an "address" is an absolute address,
including the base address. Things relative to the start of the
register block are now called "offset"s.

The other part of this patch moves these two functions from
cip_base_scoreboard into dv_base_reg_block, which is probably a more
appropriate place for them to live. This was the original motivation
because I thought I needed the "offset_to_addr" functionality in
reactive sequences. It turns out this is wrong (we actually just need
to round addresses down!), but it can't hurt to put things somewhere
more sensible.

Finally, the patch tidies up how the base address for the register
block gets set (since I had to touch the code anyway). This includes
fixes for some integer overflows to do with sizes that are the whole
of the address space.

Note that it's a bit odd to put these functions in the register block
rather than the register map (since one register block can have
multiple register maps). However, we don't seem to actually
instantiate dv_base_reg_map when building our register blocks (indeed,
I'm not sure whether you can do so with UVM), so putting the functions
there doesn't work! Assuming this is right, I'll follow up with a
patch to get rid of that class later.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>