[reggen/topgen] Add support for generating random netlist constants
This extends the instantiation paramter feature of reggen/topgen such
that random netlist constants can be generated and assigned in a
systematic way.

To that end, each IP has to declare in the corresponding HJSON which
parameters are compile time random netlist constants. Topgen then
aggregates and uniquifies those parameters, assigns random data and puts
them into a dedicated top-level package which is only meant for random
netlist constants.

The supported random constant types are either "data" which just
corresponds to a randomized logic of a certain length, or "perm" which
corresponds to a random permutation of linear array indices up to a
certain length.

Topgen allows to specify a custom seed inside top_earlgrey.hjson or on
the commandline using the --seed switch in order to make the random
number generation repeatable, if needed.

Signed-off-by: Michael Schaffner <msf@opentitan.org>