[dv/kmac] Create barebones KMAC testbench
This PR creates the barebones KMAC testbench extended from cip_lib, and
as such mostly just consists of the skeleton code generated by uvmdvgen
tool.

With this PR, the testbench can compile successfully - no tests exist as
of now but will be added later.

This PR also adds the KMAC testbench to the DV lint flow.

Signed-off-by: Udi Jonnalagadda <udij@google.com>