[doc/security] Device Provisioning
Introduce the initial Device Provisioning specification for discrete
devices.

The document describes the OpenTitan provisioning flow which is
divided into two stages:

1. Personalization: Covers initialization of the device with an unique
   cryptographic identity known as the Creator Identity, its
   endorsement certificate, as well as additional secrets required to
   configure defensive mechanisms. This occurs only at manufacturing
   time.
2. Owner Personalization: Covers provisioning of owner secrets and
   endorsement of the Owner Identity by the Silicon Owner. This may
   occur either at manufacturing time or as part of a later in-field
   ownership transfer.

The document was converted to markdown by a GDoc plugin and formatted by
a md-linter. Since the maximum line width is set to 80cols, in most
cases a sentence change will trigger a full paragraph reformat/diff.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>