[fpga] Add Vivado pre-synthesis hook, also setup hooks for CW305 board
This commit adds a new pre-synthesis to hook to change the severity of
messages. Also, two warnings are changed to errors:
- If the boot ROM init file cannot be found, the tool should not proceed
  and generate a non-functional bitstream.
- If latches are automatically inferred during synthesis, the tool should
  abort. This helps to prevent that such code ends up in the repo.

Existing hooks are also enabled on the ChipWhisperer CW305 FPGA board
except for the placement check of the boot ROM (specific to the
NexysVideo, required to use the boot ROM splicing script).

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>