[dv/dvsim] Fix -c option compile error
Current self.exports uses `_process_exports` to convert it from a list
of dict to dict. But when enable coverage collection, in its `post_init`
function, current code tries to add `cov_db_dirs` to `self.exports`
before it converts to `dict`. So there is a compile error.

This fix calls the super._post_init first and avoid this error.

Signed-off-by: Cindy Chen <chencindy@google.com>