[sw] Add API Guidelines for DIF Parametization
During the implementation of the Reset Manager, there were queries about
how to expose hardware parameters to DIF users, especially where those
parameters are related to the size of buffers required by the interface.

This Change adds guidance about how we want to treat these parameters
going forwards, so there don't have to be major interface changes when
we switch to fully parameterizing the DIFs.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>