[otbn] Introduce tracer for RTL
This adds a tracer module and interface. Both the module and interface
should be bound into a top-level instantiation of otbn_core with the
interface instance passed to the module. The tracer collects records
using the trace interface to gather relevant information and passes it
to the simulation via a DPI call.

The new tracer setup is added to the OTBN standalone verilator
simulation. An optional log file (via --trace-log-file) can be specified
which causes the trace to be output there.

Fixes #2624

Signed-off-by: Greg Chadwick <gac@lowrisc.org>