[otbn] Allow core to fetch IMEM on start cycle
The fetch unit issues its first request on the same cycle as the start
signal goes high (so there's a combinatorial path from the TL
interface to imem_req_core) but imem_access_core was only turning on
at the following cycle.

The result is that the first instruction was silently getting
discarded.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>