[otbn] Rename otbn_sanity to otbn_single
This is slightly prompted by the ongoing sanity -> smoke rename, but
things are a little more complicated here because the sequence that
was called otbn_sanity is actually a bit more involved than the smoke
test we're going to end up with.

So this patch renames it to the correct eventual name. I'll follow up
with another commit to define the smoke test properly.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>