[prim/util] Fix parameter type when using prefixes
Previously, the optional prefix was used for the type definition only but
not when instantiating a parameter of this type. This caused compilation
errors.

Also, instead of `localparam` the utility now generates `parameters`.
In accordance with our style guide, this is the preferred choice as in most
cases these snippets end up in package files.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>