[otbn] Fix bugs when marking memory known in rig model.py
There were two bugs here. The first was a stupid variable shadowing
problem (the code used 'idx' for the byte address to touch and also
for the index of the range in the for loop). Since there's no real
reason to call an address "idx", I've renamed it to "addr".

The second bug was in the definition of "left", which is supposed to
be all elements up to (but not including) index last_idx_below. This
had an off-by-one error because I'd thought foo[a:b] included the
entry at b, rather than stopping one before.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>