Update COMMITTERS list
The Technical Committee has agreed to make Martin and Felix OpenTitan
committer. This commit updates the in-tree COMMITTER file to reflect
that.

Please join me in welcoming Martin and Felix to the group of
commmitters! Thank you for your great work on OpenTitan so far, we're
happy to have you in the team.

If you're interested in becoming a committer as well, please have a look
at https://docs.opentitan.org/doc/project/technical_committee/ and talk
to one of the TC members for more information.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>