[sw] Add Custom Target Dependencies
Many of our meson custom_targets depend on scripts inside the repo, but we
give them absolute paths when invoking them (usually via
`meson.source_root() / 'util/script.py'`), because meson does not cope
with relative paths including `..`. This broadly works fine,
until one of those scripts is changed, at which point meson doesn't
always realise the script will have to be re-run.

This change adds `depend_files:` configuration for each custom target
that uses an in-tree script. Where we have abstracted the definition of
the custom target, this defines a `_depend_files` variable, to go with
the `_command` variable.

One other major change is that `cargo` is now a disabler, so if meson
does not find `cargo`, you will not get any targets that depend on
cargo -- we also now ensure that we use the `cargo` that meson found,
rather than any other one that might be on the user's path.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>