[docs] Consolidate Mask ROM Docs
We're working on pulling together the documents that describe the Mask
ROM functionality, so that they can be reviewed together and signed off.

This change creates a specific folder for documents about the Mask ROM,
and moves the existing `boot.md` into it (including an alias for its
previous URL). Any other Mask ROM documents will go into this folder.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>