[otp/dv] support secret digest check
This PR support secret digest value check in scb.
Different from normal digest, secret digest's input data is not direct
write data. It needs to go through a scramble algo.
The challenge here is when we use backdoor, the data is already
scrmabled by OTP. So we need to descramble it to ensure the consistency
in scb.

Signed-off-by: Cindy Chen <chencindy@google.com>