[DV common] Fixes in sim.mk for Verilator
This commit fixes two of the pre-build steps - copying tool sources and
SV filelist generation. Verilator build is invoked from within FuseSoC,
so the filelist generation is skipped. Also, Verilator does not have
additional tool sources that need to be copied over. This fixes the
sim.mk in preparation of adding Verilator support to DVSim.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>