[otbn] Remove "err" interrupt
We're going to consolidate these interrupts, so the only one is the
"done" interrupt. To spot errors, the host software can use the
ERR_CODE status register.

This patch looks big because of auto-generated code, but the only manual
changes to hardware are:

  - otbn.hjson: Remove the interrupt

  - otbn.sv:    Remove the interrupt and tidy up the error code logic

  - tb.sv:      Remove a top-level connection

The manual changes to software are also small: although we've now just
got one interrupt type, the DIF style is to keep an explicit index, so
all we really do is remove an entry from the enum. There's also a
manually entered count of interrupts in dif_plic.c which needs
updating.

In software test code, we remove various entries from
dif_otbn_unittest.cc. The interrupt tests try the first and last
interrupts, which doesn't make sense when there's just one, so this
patch removes the "last" part each time. Finally, we have to change an
interrupt count in dif_plic_unittest.cc.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>