[kmac] Revise to release MsgFIFO access when Idle
Problem:
    When Idle or KeyMgr is in operation (KDF), Sw access to MsgFIFO
    hangs the interconnect.

`kmac_keymgr` drops the `sw_ready_o` if `mux_sel` is not `SelSw`. The
mux is switched to the software when KeyMgr is not operating and the
software issues `CmdStart`.

In this case when the mux is not switched to the software, if the
software writes data into the Msg FIFO, the keymgr interface module
never asserts the ready. So, there's no way for software to release this
condition, as the software cannot issue CmdStart.

This commit is to fix the behavior by asserting the ready always. Then
it also creates an error code for this scenario so that the software
knows what it did wrong.

This is related to issue #4409

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>