[otbn] Fix stall behaviour in ISS at start of time
The previous version of the code worked, in that it took the right
number of cycles. But it incorrectly started the first instruction
immediately and then stalled just after it. This doesn't matter unless
you're checking against the RTL and you reset the block immediately
after starting it. Which is exactly what we do in the Verilator
testbench!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>