[otp_ctrl] Update EDN interface and KDI logic
This updates the EDN interface to use the correct structs instead of the
mock-up types, and instantiates the prim_edn_req gadget to synchronize
the data and pack it to the correct internal width.

This also fixes the KDI FSM that previously did not correctly request
entropy data under certain circumstances.

Signed-off-by: Michael Schaffner <msf@opentitan.org>