[sw] Document and Fix Meson Argument Handling
The primary motivation for this change was to add functionality to error
on unprocessed arguments, for instance if someone passes an argument
without the right preceeding option.

As part of this work, I noticed we weren't using bash's `getopts` in a
conventional manner, especially around including `?` in the optstring,
which doesn't mean anything.

As part of this change, I have: documented exactly how optstring should
be written; corrected how error handling in `./meson_init.sh` works so
that unexpected options and options without arguments are handled
correctly; and reordered all the options to be in alphabetical order.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>