[otbn] Fix address range check for narrow loads and stores
Dividing by 32 here converts from "address" to "32-byte word". We
should add 3 to the address before dividing, so that we're getting the
word corresponding to the top byte of the load or store. (In practice,
this doesn't actually do anything because addr is 4-byte aligned, but
it can't hurt to be explicit).

Before this patch, we were adding 31 instead (copy-pasted from the
wide load and store methods above), which causes an error if you try
to load or store 32 bits to e.g. 0xff4.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>