[otbn] Fix instructions that both read and write x1
The code was writing to the wrong address (one too high). This patch
fixes things by defining stack_wr_idx to be the address to which we
should write. It's normally equal to stack_wr_ptr_q (one above the top
of the stack), but gets decremented by one if we're popping too.

The patch also renames stack_rd_ptr to stack_rd_idx so that things of
length StackDepthW are called "idx" and things of length StackDepthW+1
are called "ptr": hopefully this will make things a bit easier to
understand.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>