[aes] Add reset to data, IV, state and key registers
We originally removed the reset from those registers as part of
lowRISC/OpenTitan#1845 to avoid Hamming weight leakage of potentially
sensitive assests during targeted reset glitch attacks.

When discussing the pros and cons of non-reset vs. reset flops later on,
we have concluded that in the case of AES reset flops should be used for
the following reasons:
- In the hardened AES implementation, key and state values are stored in
  multiple shares. Neither the Hamming weights of the individual shares nor
  the summed Hamming weight are proportional to the Hamming weight of the
  secret asset.
- Input/ouput data and IV values are (currently) not stored in multiple
  shares but these are less critical as they are used only once. Further,
  they are stored in banks of 32 bits leaving a larger hypothesis space
  compared to when glitching e.g. an 8-bit register into reset. In
  addition, they could potentially also be extracted when being transferred
  over the TL-UL bus.

Therefore, this commit re-adds resets to data, IV, state and key registers.
Upon reset, we keep clearing the registers with random data.

See lowRISC/OpenTitan#2603 for more information on the non-reset vs. reset
flop rationale.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>