[otbn] Switch to using bitfield for errors
This replaces the `ERR_CODE` register with an `ERR_BITS` register. This
is a bitfield where a particular bit signals a particular error. The
commit includes all RTL, documentation and software updates needed.

For the ISS this commit alters the error codes it uses to match the
`ERR_BITS` encoding but further work is required so the ISS can produce
several errors at once.

For now the RTL continues to prioritise errors and will only signal a
single error. This is to ease the transition to the use of `ERR_BITS`
whilst the ISS changes are still needed. The RTL will drop the
prioritising and will be capable of signalling several errors at once
when the ISS is ready to deal with this.

Signed-off-by: Greg Chadwick <gac@lowrisc.org>