[systemtest] Ensure all produced data is matched
`utils.find_in_files()` is a helper used by most of the systemtest
framework to search through output such as STDOUT/STDERR of a
simulation, UART logs, etc. It searches for a pattern match in the
output. As an optimization, the function returns early if it "knows"
that no more data will be produced, e.g. because the producing
simulation has terminated.

Until now, we returned immediately if the producer terminated, causing
us to miss the last output from it shortly before shutting down. This
commit changes that to also include this last output in the matching.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>