[otbn] Add support for "const values" to RIG
This is going to be important for generating loop bodies. The trick is
that we want to say something like "don't touch x3, x4 or x10 in the
loop body" to make sure that they still have the same values each
iteration. This way, they can be used as base addresses for memory
accesses or jumps, without having to worry that they'll get trashed on
the following iteration.

This also adds support for "forgetting values" in the model. We'll
generate a loop body by iterating over all known registers and either
marking them const or forgetting their exact values. This guarantees
that we don't use a trashable register as a base address.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>