[dv/alert] Fix logic in cip_base_scoreboard to check alert
This PR fixes logic in cip_base_scoreboard that checks alert.
Main issue is the delay, we remove the flexibility for user to set their
own delay in alert check. We need to know the accurate predict for when
alert happens in order to know if the alert request is ignored due to
alert handshake in progress

Signed-off-by: Cindy Chen <chencindy@google.com>