[rv_timer DV] Reduce rv_timer smoke test runtime
`rv_timer_smoke` is the longest running test in the private CI<
consuming a whopping 6 minutes of runtime.

In this change, the existing smoke vseq is renamed to random vseq and a
`smoke_test` switch is added to it. If set, it significantly reduces the
scope and runtime of the CI test so that we can continue to stay within
the 20 minute timeout.

The original `rv_timer_smoke` test entry is duplicated to
`rv_timer_random`, with both running the same random vseq. For the smoke
test version, the above plusarg is set to 1 and it is also reseeded only
once, since there is no need to run it multiple times.

This brings the `rv_timer_smoke` test runtime down from 6 minutes to 52
seconds, which will enable us to add more DUTs to the regression list
without blowing the 20 minute timeout.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>