[otbn] Exclude non-synthesizable files from being compiled
Synthesis with DC currently fails as DC is unable to parse the code in
otbn_tracer. We include those files in the default set of files to
build, but never use them; still, the tool must be able to parse them,
which isn't the case here.

Extend the approach in 45d6bcc11002d2ce07186af4f24266cc773d4261 to the
remaining model-only files. See also
https://github.com/lowRISC/opentitan/issues/4144#issuecomment-726937662
for a more detailed discussion of the problem.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>