[sw/sca] Add Mersenne Twister PRNG library to improve capture rate
This change introduces a Mersenne Twister PRNG that can be used to generate
random plaintexts on the device. Generating random plaintexts on the device
eliminates the overhead of sending them from the host and can significantly
improve capture rate. The host must use the same PRNG to be able to compute
the plaintext and the ciphertext of each trace.

Signed-off-by: Alphan Ulusoy <alphan@google.com>