[prim] Properly fix widths in tlul_adapter_sram
In commit a78b877, I tried to sort out widths properly when SramDw !=
TL_DW but didn't quite get it right. In this module, we have two FIFOs
(u_reqfifo and u_rspfifo) for bus accesses. These should be sized for
the bus width, rather than the SRAM width.

Doing so gets rid of some nasty casts (and an out-of-date waiver),
which is nice, and should fix a remaining AscentLint width mismatch
error.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>