[rtl/sram] fix otp<->sram connection parameters
Currently the `prim_sync_reqack_data` module instantiated to connect the
OTP and SRAM is parameterized such that it expects data to be
transferred from the SRC side (SRAM) to the DST side (OTP).

However this needs to be reversed, as the OTP is sending scrambling seed
data over to the SRAM.

This PR adds the relevant parameters.

Signed-off-by: Udi Jonnalagadda <udij@google.com>