[dv] Gracefully kill simulation
When running sims remotely (LSF, etc), if we need to kill a simulation
that might be hung, then issuing a TERM signal to the process running
the simulation (or `bkill`ing the LSF job) causes the simulation to
terminate unexpectedly and uncleanly. If wavedump is enabled, it causes
the dump to get corrupted, preventing us from debugging the cause of the
hang in the first place.

In this patch, there is logic added to the simulation that checks for
the presence of a `dv.stop` file in the run directory periodically every
1ms. Whether to enable this logic AND the polling interval can be
changed via plusargs. This file can be easily added to the run directory
by running `touch <rundir>/dv.stop`. DVsim can use this method rather
then killing the process (more updates on that coming soon).

Signed-off-by: Srikrishna Iyer <sriyer@google.com>