[tlgen] Allow some keys that are generated by topgen
These keys appear when loading the xbar data from a generated
top-level json file.

clock_group and clock_srcs come from the original top-level and are
used to calculate clock_connections. Ideally, we'd not pass them
through, but that's a bigger change to quite gnarly code, so I'm
leaving them be for now.

domain is added to every object in topgen's check_power_domains
function, so needs to be allowed here too.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>