[padring/top] Carry over pinout from bronze and align sim and FPGA tops
Note that this is an intermediate step needed for the pinmux updates.
There are a couple of things that are still Bronze-specific and these
will be cleaned up over the coming weeks as we work towards Silver.

[top] Align englishbreakfast I/O configuration with earlgrey

Since both designs use the same padring, the MIO/DIO config should be
the same, even if not all IOs are connected to a pad.

[topgen] Topgen adjustments to keep MIO order stable

This changes the way the MIOs are enumerated and connected to the
pinmux. Before, topgen would connect all inouts from all modules first,
and then proceed with inputs and outputs.

This is inconsistent with how the ordering is implied by the
`mio_modules` key inside the top hjson.

This fix changes how the MIOs are enumerated. Within a module, topgen
still connects the inouts first, but globally it observes the order
specifiec by the `mio_modules list.

Note that this piece of topgen is going to be overhauled, so this is
just a temporary fix to ensure the MIO insel/outsel indices remain the
same among tops that have different amounts of MIO peripherals.

[top/dv] Alignments and workarounds to get chip simulation working

[top] Regenerate files

Signed-off-by: Michael Schaffner <msf@google.com>