[otbn] Initial rewriting support in extracted documentation
This generates a "pretty" version of GPR/WDR accesses. For example,
the ADD instruction looked like this before:

    val1 = state.gprs.get_reg(self.grs1).read_unsigned()
    val2 = state.gprs.get_reg(self.grs2).read_unsigned()
    result = (val1 + val2) & ((1 << 32) - 1)
    state.gprs.get_reg(self.grd).write_unsigned(result)

and now looks like this:

    val1 = GPRs[self.grs1]
    val2 = GPRs[self.grs2]
    result = (val1 + val2) & ((1 << 32) - 1)
    GPRs[self.grd] = result

Signed (2's complement) conversions are shown explicitly. For example,
here's SRA:

    val1 = from_2s_complement(GPRs[self.grs1])
    val2 = GPRs[self.grs2] & 0x1f
    result = val1 >> val2
    GPRs[self.grd] = to_2s_complement(result)

WDRs are also converted. For example, BN.ADD looks like this:

    a = WDRs[self.wrs1]
    b = WDRs[self.wrs2]
    b_shifted = logical_byte_shift(b, self.shift_type, self.shift_bytes)

    (result, flags) = state.add_with_carry(a, b_shifted, 0)
    WDRs[self.wrd] = result
    state.set_flags(self.flag_group, flags)

This is just an initial rewrite pass. If we go with this, we'll want
to do things for flags, at least. This is easy enough with the same
framework.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>