[kmac] Latch absorbed signal to delay Done command
kmac_keymgr interface asserts `CmdDone` when it sees `absorbed` signal.
The SHA3 module asserts absorbed signal when it is in `StAbsorb` but the
next state is `StSqueeze`. It creates assertion error as only
`StSqueeze` expects `CmdDone`.

To break the timing path, latch in absorbed_o is added.

This issue is reported by @udinator

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>