[dv/hmac] Fix hmac vector parsing
This PR fixes the regression failure on hmac NIST vector parsing. The
hmac vector does not have `KeyLen` keyword, but still has `Key`. The
current parsing logic will set the key to 0 in this case.

Also hmac NIST vector does not need to reserve key input, but kmac
vectors needs to. So this PR adds a flag for user to decide if they want
the key to be reserved.

Signed-off-by: Cindy Chen <chencindy@google.com>