[prim] Avoid an apparent combinatorial loop in prim_secded_*_dec.sv
Verilator tracks combinatorial loops by variable, not by bit. Thus, it
sees a loop with the code that was there and generates an UNOPTFLAT
warning. Pull the shared ^syndrome_o calculation out into a separate
variable to fix things.

Running secded_gen.py was a bit painful. For next time I need to
figure this out:

    git grep secded_gen.py -- '*dec.sv' | \
      grep -v hamming | \
      cut -d ' ' -f 3- | \
      xargs -n 8 util/design/secded_gen.py

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>