[otbn] Fix missing operation docs for bn.sel
The class name had a spare "r" in it (probably a copy-paste error from
when I copied it from BNRSHI above). This doesn't matter in any way
for the ISS, because decoding and tracing is driven by the "insn"
class field (which had the right name). But it *does* matter for the
connection with document extraction, which means that the BN.SEL entry
in the docs was missing its Operation section.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>