[aes] Interface CSRNG through EDN
This commit interfaces the AES unit with CSRNG/EDN to reseed the internal
PRNGs for clearing and masking. Since on the SCA platform, we don't have
enough resources to implement CSRNG and EDN, a new parameter
SecSkipPRNGReseeding is added to skip reseeding requests on the SCA
platform.

If the reseeding was done with deterministic input instead of entropy
provided by CSRNG, this would result in quickly repeating, deterministic
PRNG output which is not suitable for evaluating SCA resistance.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>