[lint_commits] Propose a command line which signs the commit
We do some basic checks on the commit author to check if it is likely a
"real name." When this check fails we give a git command line to fix it.
However in that command line we do not include the `--signoff` flag to add a
`Signed-off-by` line, meaning users following our guidance will then be
facing another lint failure. This is fixed in this commit, hopefully
making our developer experience slightly nicer.

Also replace all uses of `-s` in this doc with `--signoff` to make it
clearer to users what this flag actually means.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>