[otp_ctrl] Properly end cnsty/integ checks that found an error
This changes the behavior of the buffered partition checker logic such
that failing checks signal back to the LFSR timer that they have
finished performing the check.

That way, the LFSR timer can still continue running and triggering
checks in other partitions, despite the faulty partition being in the
terminal error state.

Signed-off-by: Michael Schaffner <msf@opentitan.org>