[aon_timer dv doc] Update DV doc
UVMDVGen generated DV doc appears to not be up to date with the change
in the terminology (Testplan -> DV plan, DV plan -> DV document, path:
hw/ip/<ip>/doc/dv_plan/index.md -> hw/ip/<ip>/doc/dv/index.md). This
update fixes that.

Also the DV doc link in .prj.hjson file is updated so that it gets
linked up in the HW dashboard.

The testplan is also added to the build_docs.py script so that it can
render the DV doc page correctly.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>