[reggen] Make Params class less reggen-specific
The Params class is a nice container for holding parameters, only two
methods make it specific to the use in reggen. Leave only the common
parts in Params, and move all reggen-specific things into a new class
ReggenParams.

(I did go this route instead of renaming Params -> BaseParams to
indicate in the class name that the code is actually specific to reggen.
ReggenParams isn't the greatest name either, but the usage scenarios for
the class are rather wide, so RegParams sounded wrong as well ...)

Signed-off-by: Philipp Wagner <phw@lowrisc.org>