[rom_ctrl] Initial design RTL for the ROM checker
This has the counter FSM to read from ROM, the mux to arbitrate
between the checker and bus access, and a "compare" block to check
against expected outputs.

The checks that run at the start of time are all disabled by
default (because of the SkipCheck parameter). This means that we can
integrate this code with the top-level and teach pwrmgr to sequence
things properly before turning on the check.

This commit does not yet contain ROM scrambling (which will need a bit
more tooling work).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>