[prim] Make prim:flop_2sync depend on prim:flop
This is a bit of a hack. The generic implementation in
prim_generic_flop_2sync uses prim_flop, but fusesoc doesn't support
running generators a second time to resolve dependencies for generated
cores, so adding the dependency to prim:generic_flop_2sync doesn't
work. Add it to the parent for now, on the basis that
prim:generic_flop_2sync only gets used if prim:flop_2sync does.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>