[ci] Be more informative if install-package-dependencies fails
Also, use a temporary directory to avoid polluting the checkout and
move the logic to a shell script (no real change from the yaml file,
but this script has got big enough that it doesn't feel like it should
be inlined any more).

Splitting out into a shell script means a bit of extra work, because
we have to pass Azure Pipeline variables through to the script. Use
proper command line arguments, rather than environment variables, to
make it a bit more obvious what's going on.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>