[dv/hmac] Fix illegal bin failure
This PR fixes hmac illegal bin error. The error is due to hmac scb
sampling issue.
When we do a intr_test write, and then intr_test write, the interrupt
could be cleared. Then if we read intr_test, the interrupt test still
expects the test bits to be set.
This PR fixes the issue by sampling it on intr_test.

Signed-off-by: Cindy Chen <chencindy@google.com>