[memutil] Change DpiMemUtil so that it no longer owns MemAreas
This makes the code a bit simpler, which is nice, but the main point
of this is that other code might want to access the MemAreas in
question. For example, otbn_model.cc currently duplicates the backdoor
access logic: it would be much nicer if it could just use an
OtbnMemUtil object to do the work.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>