[dv/otp_ctrl] Move lc_escalate to a separate sequence
This PR moves the logic to drive lc_escalate_en to a separate sequence.
The reason of doing this is mainly to simplify scb.

There are three cycles after lc_escalate_en is On that registers could
behavior different. Due to address cycle and data cycle delays, these
three cycles can be even longer. This makes SCB hard to predict the
correct values.

Signed-off-by: Cindy Chen <chencindy@google.com>