[dif_kmac] Add header file and checklist for KMAC DIF.
This API provides a streaming interface to the KMAC unit. The KMAC
unit will be initialized with information about how entropy will be
generated and the endianness of the message and digest. After that
a transaction can be started by passing the required information
(e.g. key) to a `dif_kmac_mode_{sha3,shake,cshake,kmac}_start`
function. Data is then passed into the hash function using the
`dif_kmac_absorb` function. Once the data to be hashed has been
completely absorbed the output (a.k.a. digest) can be written out
using the `dif_kmac_squeeze` function. The output may be a variable
or fixed length depending on the mode. Once the squeeze has been
started no further absorb operations can be performed. Finally,
`dif_kmac_end` is called to end the operation ready for a new
operation to be started.

Example:

```c
// Initial hardware configuration.
  dif_kmac_config_t config = (dif_kmac_config_t) {
    .entropy_mode = kDifKmacEntropyModeSoftware,
    .entropy_seed = generate_seed(),
    .entropy_fast_process = kDifKmacToggleEnabled,
    .message_endianness = kDifKmacEndiannessLittle,
    .output_state_endianness = kDifKmacEndiannessLittle,
  };
  CHECK(dif_kmac_configure(&kmac, config));

  // Example KMAC XOF calculation with empty customization string.
  CHECK(dif_kmac_mode_kmac_start(&kmac,
                                 0 /* L=0 (XOF) */,
                                 &key,
                                 nullptr /* S="" */));
  CHECK(dif_kmac_absorb(&kmac, msg, len, nullptr /* block */));
  CHECK(dif_kmac_squeeze(&kmac, out, len, nullptr /* block */));
  CHECK(dif_kmac_end(&kmac));
```

Signed-off-by: Michael Munday <mike.munday@lowrisc.org>