[spi_device] Add passthrough datapath
This commit introduces the SPI passthrough feature. It snoops the SPI
lines and intervenes if the incoming command from the attached host
system is not permitted (command filter) or the data of the command
should be swapped (address manipulation).

The details are given at the header of
hw/ip/spi_device/rtl/spi_passthrough.sv systemverilog code.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>