[otp_ctrl,lint] Write a range select explicitly, not using >>
The right shift in otp_ctrl_part_buf.sv was being used to convert
an (N+1)-bit long byte address (that we happen to know is even) to an
N-bit long halfword address. The right shift, combined with Verilog's
truncation rules, mean that it functioned as a range select to grab
just the bottom N bits.

Verilator complains about such things (it treats the assignment of N+1
bits to an N-bit variable as a width mismatch). Here, it's easy to
sort out: just write the range select explicitly.

We have an equivalent situation in otp_ctrl_part_unbuf.sv (although
this was inlining the right shift into the branches of a conditional
expression). Make this match too.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>