[sram_ctrl] Use the correct tl_instr_en_e
This enum type is defined in both sram_ctrl_pkg.sv (as an 8-bit type
with values EnSramIfetch and DisSramIfetch) and in tlul_pkg.sv (as a
3-bit type with values InstrEn and InstrDis).

Importing the tlul_pkg version into scope in sram_ctrl is actually a
spec violation (SystemVerilog doesn't do shadowing, so you're not
allowed to import two things that clash) and also happens to confuse
Verilator's width inference.

Fortunately, everything sort of worked because the external ports used
the narrower type and the enum values had the same bottom bits, but it
was all a bit magic and kind of worked by luck. This patch explicitly
qualifies the name when we use tlul_pkg::tl_instr_en_e, which should
sort everything out.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>