[keymgr] Define an intermediate signal for setting cfg_regwen
This avoids an UNOPTFLAT warning from Verilator, caused because
Verilator doesn't automatically split the hw2reg structure when
topologically sorting the processes that read to and write from it.
Another way to address this is to use a "split_var" Verilator-specific
hint, but I think the code is also probably clearer for humans when
rewritten this way.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>