[pinmux] Update strap sampling mode in DFT-enabled LC states
This slightly modifies the strap sampling behavior in DFT-enabled
LC states. In particular, we now continously sample the straps in
those states, instead of gating the continous sampling mode
on whether the initial sample was nonzero or not.

This puts less burden on the emulation and simulation environments
where it the device can run in a DFT-enabled mode, since the
strap values may be changed arbitrarily during the test sequence.

This also changes the default life cycle state to RMA instead of DEV
for simulation and emulation environments to ungate all debug
infrastructure and functional modes.

Signed-off-by: Michael Schaffner <msf@google.com>