[otbn] Standalone sim: only dump regs if asked for
The standalone OTBN simulation, `otbnsim/standalone.py` has an option
`--dump-regs` which controls the dumping of all registers to a file or
STDOUT after the operation. By default, the registers were dumped to
STDOUT, implicitly passing `--dump-regs=-` to the simulation. This makes
it hard/impossible to *not* dump registers at all.

Instead of dumping to STDOUT by default, this commit makes `--dump-dmem`
and `--dump-regs` consistent in behavior: don't dump anything unless the
option is passed. Passing `-` as file name will dump the register or
memory contents to STDOUT.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>