[dv/sram] add executable SRAM test
This PR implements the sram_ctrl_executable test as described by the
testplan.

In this test, we fully randomize all DUT config knobs used to set the
SRAM into "executable" mode, and then drive TL transactions that are
randomized between DataType and InstrType.

The scoreboard is updated to handle the new set of error cases that is
introduced with this testing.

The tl_monitor and tl_seq_item have also had minor updates made to them
to copy the a_user bits, as this is required to drive the right kind of
stimulus to the SRAM.

Signed-off-by: Udi Jonnalagadda <udij@google.com>