[otbn] Some initial notes about functional coverage
This text should eventually be the thing that gets signed off as
describing all the relevant coverage points. We'll either keep this
document, and connect it up to automatic coverage collection, or move
the guts of this document somewhere else when there's a better place
for it.

So far, there are a couple of general notes, plus some
specific (architectural) coverage points for the base instruction set.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>