[entropy_src hack] This commit will not be merged.
It is a temporary hack to allow the sha3 fsm to gracefully terminate.
The current issue is that when disabling, the fsm still requests csrng to halt.
However, the outgoing "halt" request is qualified with es_enable, this means the halt is never seen and as a result never ack'd.

The fsm then is unable to gracefully terminate.  This probably implies all the outgoing signals need to be looked at a bit more carefully to see which should be qualified, which should not.  Alternatively, the internal fsm checks could further account for es_enable and not do a halt request.

Signed-off-by: Timothy Chen <timothytim@google.com>