[reggen,dv] Split some long lines in generated UVM code
In order to get a proper "longlines" check in CI, it would be nice to
drop the current Verible configuration from 150 characters down to
100, to match AscentLint. Since we use that for DV code as well as
RTL, we there are a few other places that need tidying up.

This patch avoids the long lines in the generated UVM code, such as
chip_ral_pkg.sv.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>