[rom_ctrl] Change the meaning of SkipCheck
This flag originally disabled the initial ROM check. The point was to
allow pwrmgr integration to happen after the code was merged. This
integration is now done, but there are several other integration
changes that aren't yet done:

  - KMAC doesn't yet signal rdy on its app interface (this seems to
    expect a key input, which doesn't make sense for the cSHAKE
    computation we want to do: probably a bit of proper re-work needed
    before this will work.

  - We don't yet have a sensible expected hash stored in the ROM.

If SkipCheck is set, we fake up the connections to KMAC, so that it
always looks like KMAC is ready for ROM data (but don't actually pass
anything through). We then fake things up so that it looks like KMAC
has immediately responded.

Finally, we force the comparison with the expected hash to always look
like it succeeded.

Once KMAC integration is done, we can remove the hacks on the KMAC
connection and sort out the ROM scrambling script to put something
sensible in the top 8 words. When *that* is done, we can remove
SkipCheck entirely.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>