[rom_ctrl] Fix assertion about rel_addr
We subtract TopStartAddr (the address of the first of the "expected
digest" words) from counter_addr to get an index into the expected
digest words. This will always be in the range 0..7 when we care about
it (when reading_top is true) and we actually get the index by slicing
out those bottom bits.

We've got an "unused_" signal to waive the fact that we ignore the
other bits and the RelAddrWide_A assertion says that we don't care
about their values.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>