[otbn] Add ECC to register files and call stack
ECC SECDED is applied to both base and bignum register files. For the
base register file a single ECC is applied to whole 32-bit word. For the
bignum register file there is a seperate ECC per 32-bits (8 in all to
cover a 256-bit register).

ECC encode and decode is implemented by the register file. No error
correction is applied, ECCs are only used for error detection.

The register file interfaces are expanded to allow a direct ECC write.
This is used where whatever is supplying the data already has an ECC so
it doesn't need to be recalculated (e.g. during a bignum register move
or select).

Fixes #6186

Signed-off-by: Greg Chadwick <gac@lowrisc.org>