[aes/pre_syn] Add experimental Yosys synthesis flow
This commit adds an experimental Yosys synthesis flow for the AES
module under hw/ip/aes/pre_syn. The idea of this open flow is to
generate ASIC area estimates as well as to procude netlists that can
further be used for developing and verifying SCA and FI countermeasures,
for example by means of formal security verification using tools like
REBECCA/Coco-Alma. The flow itself is based on the experimental Yosys
synthesis flow developed for Ibex.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>