[checklists] Update all checklists for consistency
Noticed a few discrepancies in some of the checklists (missing items).
Updated all of them to be uniform and reflective of the newer addtions.

I also reverted DV plan to testplan (my apologies for the back and
forth) - after talking to some more folks, it became evident that
testplan is the more preferred term.

Finally, a new checklist item is added to V2: DV_DOC_TESTPLAN_REVIEWED.
The goal of this item is to have the DV document and/or the testplan
and/or the functional coverage model fully reviewed for any potential
holes that can be addressed in post-V2 timeframe.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>