[otbn,dv] Add extremal value coverage for immediate insn fields
This code adds coverpoints and crosses to track the generic coverage
targets we have for immediate fields of instructions.

The basic approach is that we map instruction names to encoding
schemes, then have a big case statement in otbn_env_cov::on_insn that
dispatches to the correct encoding scheme.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>