[sw/silicon_creator] Introduce the keymgr driver
The key manager is initialized with an entropy reseed interval provided
by the mask rom and advanced into the initialized state. It is later
configured with the software binding tag and maximum key version
information extracted from the ROM_EXT manifest. The mask rom checks the
status of the key manager before handing over execution to the ROM_EXT.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>