[dv/mem_partial_access] Fix fork join_none inside while loop issue
This PR fixes the following issue:
while (var > 0) begin
  fork
    do_something();
    if (read_mem) begin
      var --;
    end
  join_none
  wait_outstanding_access();
end

If var is small, it will cause var underflow. So I reverted the logic
back to use `repeat(var)`. Originally, the code was changed to while
loop with the concern that memory is too small and does not have enough
mem_read. So here I added a `min_accesses`.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>