Avoid premailer 3.9.0 due to API breakage
https://github.com/peterbe/premailer/pull/255 introduced a new, for our
use case mandatory, argument to the constructor. Since this argument
wasn't part of the constructor in 3.8, we need to runtime-detect if
we're running premailer 3.9+ and then set the option.

Let's avoid this version for now and hope upstream sorts it out in a
more backwards-compat way -- or we wait a bit longer and force our users
to use premailer 3.9+.

Fixes #6768

Signed-off-by: Philipp Wagner <phw@lowrisc.org>