[otbn,dv] Dump an "expected end address" symbol from the RIG
The random instruction generator currently tries hard not to generate
any invalid instructions or trigger any errors on the processor. We're
going to start generating errors soon, but we don't want to do so
"accidentally".

To help us avoid that, we add an "_expected_end_addr" symbol to the
generated ELF file. A followup commit will teach the simulator to spot
that address and generate an "oh no! the world is broken!" error if
simulation stops on the wrong address.

Note that this change won't give us any extra confidence in the RTL,
but it should make it easier to debug if we make a mistake in the RIG
and it's not generating quite the program paths we thought it was.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>