[dv, mem_bkdr_util] Implement mem_bkdr_if as class
There have been numerous pain points associated with the inflexibility
of not being able to override the base functionality of the existing
`mem_bkdr_if`, especially in the closed source repo. In this commit, the
existing `mem_bkdr_if` is implemented as a class.

The following are the key changes made:
- Implemented the backdoor interface as a class
- Removed the `MEM_ARRAY_PATH_SLICE` to directly manipulate the
hierarchical path to the memory array by binding the interface to a
sub-path.
- Instead, the class takes the string path to the memory hierarchy as a
runtime input, which is read / written using the DPI-C functions
`uvm_hdl_read` and `uvm_hdl_deposit`.
- There are no parameterizations - all memory attributes are treated as
runtime settings.
- Some optimizations in the overall code.

The class is not used in any of the testbenches in this commit, so that
the review is easy. The subsequent commits achieve that.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>