[otbn,dv] Teach otbn_memutil to track expected end address
RTL simulations need to get hold of the expected end address (a
symbol) to figure out whether the end address they've seen matches or
not.

This patch adds an OnElfLoaded() hook to DpiMemUtil and implements it
in OtbnMemUtil() to grab the relevant symbol. Nothing uses the result
yet, but it's accessible by calling OtbnMemUtil::GetExpEndAddr() (from
C++) or OtbnMemUtilGetExpEndAddr() (over DPI from SystemVerilog).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>