[aon_timer] Move registers to the AON domain
Rather than performing CDC synchronization for all registers
individually, this moves the whole register block into the AON domain
and performs the synchronization on the TLUL bus instead.

This should save a fair amount of registers compared to synchronizing
each register separately. It also means that software does not need to
wait/poll to check when writes have propagated since the response will
not be issued until the write has completed.

As a side-effect, reads and writes to aon_timer registers are
considerably slower than before since the CPU stalls waiting for the
response. This required the timeout to be extended on the
chip_dif_pwrmgr_smoketest.

Relates to #5805

Signed-off-by: Tom Roberts <tomroberts@lowrisc.org>