[aon_timer] General register cleanup
NOTE: No functional changes intended just tidy-up

Now that the register file itself is clocked on the AON domain, all of
the registers that were instantiated externally can be moved back into
the autogenerated files.

DV exclusions that were applied due to the propagation delays on
register reads and writes have also been removed.

The "core_if" in the aon_timer testbench has also been removed as there
isn't really anything left for it to bind on to.

Signed-off-by: Tom Roberts <tomroberts@lowrisc.org>