[dv, util] Make poll_for_stop() opt-in
The `poll_for_stop()` method in `dv_utils_pkg` provides a mechanism to
kill the simulation more gracefully on hangs. It looks for the presence
of a file called `dv.stop` every 1us in the rundir which the user can
'touch' if it needs to be killed.

This mechanism was added because historically, it was not always
reliable to kill a sim by hitting `CTRL-C`, especially with waves which
would get corrupt on `CTRL-C`. The other reason was to be able to kill a
specific sims if they are deemed broken already while allowing others to
continue to run.

In any case, by allowing that logic by default introduces a pretty big
bottleneck - it consumes a whopping 54% of simulation runtime,
especially for UART which has periods of idle due to slowness of the
baud rate. In this commit, the logic by default is turned off. If this
feature is needed, it can be turned on by passing `--run-opts
"+poll_for_stop=1"` on the command line.

The other change is to increase the polling interval to 10us.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>