[dv, flash_ctrl] Split flash_ctrl_rand_ops_vseq
This is a minor enhancement that "renames" the
`flash_ctrl_rand_ops_vseq` to `flash_ctrl_rand_ops_base_vseq` which is
not run directly, but extensions of it that configure the randomization
knobs are. It adds the `flash_ctrl_rand_ops_vseq` which we run as a
test.

The smoke test is now made to extend from the base rand ops class.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>