[reggen,dv] Don't specify rights for add_reg when constructing RAL
This access control is supposed to be for when you have two different
address maps that have different views on the same registers. Here, we
were calculating a (somewhat bogus) value based on the default access
specified for fields in the hjson.

The fields' specific access rights get reflected separately: no need
to mess around here.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>