Revert "[tlul] Drop the "quick fail" path from tlul_adapter_sram.sv"
This reverts commit 1c3d506a81961964941974c384a137c45845baa2. The
problem is that some blocks make gnt_i dependent on req_o so an
erroring access just stalls. This is probably not great behaviour, and
maybe we should fix that in the future, but doing so would mean
renaming the signals (to e.g. rdy/vld, rather than gnt/req).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>