[prim] Remove dependency of prim:esc on a hardware block
Ideally, primitives shouldn't depend on blocks (the dependency is
supposed to go the other way around).

This is a little awkward, because the prim_esc_receiver module has
signals whose width depends on the number of escalation severities and
the width of the ping counter in the alert handler. We can work around
this by passing the parameters explicitly. At the instantiation sites,
we *can* point at alert_handler_reg_pkg without introducing a "wrong
way" dependency.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>