[aon_timer] Always capture wakeup cause.
When aon_timer triggers JUST before we enter low power, the wakeup
request is missed because sleep_mode is not yet asserted.
This causes the system to go into low power and never wakeup.

Instead we should always latch the wakeup reason, and only show it
as a wakeup signal when we are in low power mode.

Signed-off-by: Timothy Chen <timothytim@google.com>