[mask_rom] Reduce the brittleness of shutdown_unittest
shutdown_unittest is brittle with respect to the number of alert
sources: any time the HW folks change the sources, the test will break.
This brittleness is unnecessary because the unittest doesn't care about
the specifics of the alert configuration, just that a given
configuration gets applied.

1. Define a full alert configuration up to the current limit of
OTP-defined alert configuration (80 alerts).
2. Refactor the test cases to check that the applied configuration
length matches the number of alert config registers rather than the size
of the test's alert configuration structure.

Signed-off-by: Chris Frantz <cfrantz@google.com>