[kmac] Error Checker to send valid cmd only
This commit revises kmac_errchk to send only valid commands to the rest
of the modules. So that the following modules does not have complex
sequence checker anymore.

Due to timing concern, this module latches the sw_cmd input signal. It
changes the cycle accurate model in the DV environment.

Co-Authored-by: Udi Jonnalagadda <udij@google.com>
Signed-off-by: Eunchan Kim <eunchan@opentitan.org>