[otbn] Add a "clear" input to otbn_stack
This doesn't do anything except zero the write pointer. We're
eventually going to add some sort of secure wipe, which will splat the
stack contents. It might also do a 2-stage wipe of the write pointer,
in which case this will take a bit more cleverness.

For now, this change is just to allow functional correctness (clear
the x1 stack on start) for testing.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>