[tl,dv] Add missing configuration in tl_agent_env_cfg
Both the host agent and the device agent have attached monitors, which
need to know when to sample the interface to get the right
transactions. In the tl_agent_smoke test, this fact that this was
missing meant that, depending on scheduling, we might sample D before
A on a clock-edge where both channels were for the same transaction.

The result is an error where the monitor says "hey, I'm seeing a reply
on the D channel for a transaction with source FOO but I haven't seen
the corresponding request yet".

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>