[otbn] Document randomness properties
Add two pieces of documentation:

* Document what kind of randomness software developers can expect when
  reading from RND or URND.
* Document what requirements OTBN has on the EDN to provide appropriate
  randomness.

For high-quality random numbers made available through RND we went for
PTG.3 for now in line with current thinking, but might need to update
that if our plans change.

The goal is still to specify what software developers get when using
RND. If we don't do that, software written for OTBN has an additional
dependency on the configuration of the EDN, which is has no insight to.

Also note that we don't give any guarantees at the moment for the URND
randomness drawn from a local LFSR. There are plans to remove the shift
register with another construction that provides stronger guarantees,
but that's for another day.

The HTML indentation is ugly, but required to make Hugo happy.

Written up with valuable input from Felix and Vladimir.

Fixes #6953

Signed-off-by: Philipp Wagner <phw@lowrisc.org>