[topgen] Pull the 'clocks' code into a separate class in topgen
This slightly replicates the changes we made in reggen earlier in the
year. The reason to do this here is that we have several different
places in the code where we figure out the list of aon clocks, hint
clocks etc. and they all have to stay in sync.

Rather than rely on this, let's move all the cleverness into a single
class whose code can be run multiple times (to give the same answer
each time!).

This patch doesn't really do much of this consolidating, but it does
all the infrastructure work, converting top['clocks'] from a load of
nested dictionaries into a class and teaching everything that reads
top['clocks'] to understand the new format.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>