Require Verilator 4.210
Pull request #7964 exposed a problem in Verilator that led to a segfault
of the verilated simulation during startup. We weren't able to fully
debug it, but know that the problem does not surface with Verilator 4.210
any more.

PR #7964 was updated to include a workaround that also works with
earlier Verilator versions. The update here is only a safety measure in
case we find another problem that we cannot work around easily.

Furthermore, newer Verilator versions have a fair amount of changes to
improve compile time and run time performance specificially for
OpenTitan. However, the changes don't yet make a significant impact on
Verilator runs in CI.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>