[topgen] Add missing memory attributes
This adds the byte_write memory attribute to the memory config section,
and changes the allowed swaccess attributes to be compatible with the
existing reggen environment. I.e., reggen (and the associated DV
classes) do not understand the linker syntax r, w, rwx etc, but use
the labels "ro", "wo", "rw". We therefore align the allowed values with
the ones understood by the reggen environment, and add a remapping
function to the linker script template, alongside with an additional key
"exec" in the memory config that indicates whether a region is executable.

Signed-off-by: Michael Schaffner <msf@google.com>