[otbn,dv] Double the number of otbn_single runs
Since lots of the interesting corner cases cause (architecturally
defined) errors, we have to do lots of OTBN runs to see them all. It
turns out that the setup/teardown costs of a simulation aren't too
significant compared to the time spent running instructions, so
there's no real benefit to packing multiple runs into a single
simulation. And, obviously, debugging is easier with just a single
binary.

So bump up the number of otbn_single runs, to see more of the weird
error cases :-)

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>