[topgen] Improve matching of prefix when renaming parameters
Topgen moves parameter prefixes like `Sec`, `RndCnst` and `MemSize` to
the beginning when generating the top-level parameter names. This messed
up the parameter `SecureIbex` which became `SecRvCoreIbexureIbex` in the
top level. With this commit, prefixes are only moved if they are
followed by an upper case letter. The parameter in the top level thus
becomes `RvCoreIbexSecureIbex`.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>