[dv] Add sv_flist_gen_flags HJson var for FuseSoc
This commit adds the `sv_flist_gen_flags` variable for FuseSoC which can
be used to add fileset flags. Previously, the recommendation was to use
`sv_flist_gen_opts` to add more flags. But that does not work given that
new flags added in imported HJSon files can result in FuseSoC command
args to appear out of order, which FuseSoC does not like, This makes the
fileset_flags appear at the right place.

This is added for DV, FPV as well as synthesis flows in DVSim.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>