[util] Tidy up how we call finst_gen in reg_top.sv.tpl
Now, we iterate over elements of rb.all_regs, then any subregisters (if
the element happened to be a MultiRegister) and then any fields.
Hopefully, the result is a little easier to follow. The various
special cases in how we generate signal names mean that this code is
still a little complicated, but the idea is that a reader should be
able to scan to the right place a bit quicker.

The end results are equivalent, but with slightly better controlled
whitespace. To see what's going on, run:

    git diff -w --ignore-blank-lines

which will show no differences except that we fix the subregister
indices in comments in spi_device_reg_top.sv.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>