[reggen] Ensure that registers are "either RC or not"
This was true across the design, except in the trial1 example, where
the "MIXTYPE" register had registers that were a mixture of types.

The point of this is to ensure that all the fields of a register are
updated together. This will allow us to hoist things like the QE
signal up to the register level, de-duplicating things. But that's
only possible if all fields get updated together.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>