[otbn,dv] Destroy verilator_top before end in otbn_top_sim.cc
There seems to be a strange race condition with file-scope verilated
objects. Defining verilator_top as a pointer and constructing the
object inside main() instead of in some pre-main initialisation fixed
a segfault when constructing things.

It seems that there's a corresponding segfault that can be triggered
in the destructor unless we also destruct the object in advance. We
should probably root-cause this properly but, for now, this gets rid
of the segfault.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>