[otbn] Change how the ISS checks jump/branch destinations
Put in an explicit check on the various places where we could specify
a bad next PC and then put an assertion in set_next_pc() to make sure
we don't use it wrongly.

This should improve the documentation for JAL, JALR, BEQ and BNE.
What's more, it means we can represent things like the current RTL
behaviour on a branch or jump to an invalid address that occurs as the
final instruction of a loop body. In these cases, the RTL signals both
ErrBits.LOOP and ErrBits.BAD_INSN_ADDR, which seems like a sensible
behaviour. Match it!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>