[otbn,dv] Alter ISS to match RTL for BN.LID/BN.SID double errors
There are a couple of changes here. Firstly, we change the behaviour
on double increments. An instruction like

    bn.lid x3++, 0(x2++)

is illegal but possible to encode. The RTL catches such things as part
of the decode stage, so this sort of error masks any others. Seems
reasonable! Let's do the same in the spec.

Secondly, we change things to handle cases like this:

  addi   x2, x0, 1
  bn.lid x1, 0(x2)

where we underflow the call stack when reading x1 for grd, but manage
to read an address from x2. Unfortunately, the address is bogus! Now,
we signal both errors (CALL_STACK for the underflow and BAD_DATA_ADDR
for the bad address). This works similarly the other way around (when
grd points at an invalid WDR and grs1 underflows the call stack:
checked in the bnlid-1.s test). This matches the RTL's behaviour and
probably makes more sense than what the ISS was doing before.

Also, mirror these changes for BN.SID (which has basically the same
error behaviour).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>