[otbn,rtl] Prevent illegal instructions causing call stack errors
The read and write enables going to the base register file now factor in
whether or not an illegal instruction encoding was used, or if the
instruction performs an illegal CSR/WSR access. This prevents a call
stack error occurring when such an illegal instruction is seen.

Errors caused by indirect register accesses also cause illegal
instruction errors but cannot be factored into the register enables to
avoid a combinational loop.

Signed-off-by: Greg Chadwick <gac@lowrisc.org>