[otbn,dv] Fix logic when picking operands for BN.SID
Model.regs_with_known_vals returns a list of pairs "(idx, value)".
Searching in known_wdrs[0] meant we were only able to pick reg_idx if
either reg_val was a known WDR that matched the first "idx" OR the
first known WDR happened to have a value that matched reg_val.

This caused a weird bug with the otbn_reset test where our behaviour
depended on the (architecturally unspecified) value of w0, accessed
through x0 when we happened to have another WDR (the first in
known_wdrs) that had a known value of zero.

While we're fixing things, weaken the check so we only require the
register to have an architecturally specified value: we don't actually
need the RIG to know what it is.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>