[prim_esc_receiver] Fix response toggling corner case
This should prevent spurious integrity alerts when a ping request is
interrupted by an escalation request.

In particular, instead of hardcoding the value of the response diff pair
in the FSM, we just keep on toggling the previous values whenever we are
in a response state (no matter whether we are responding to a ping or
escalation response).

The only time we reset the diff-pair state to an initial value is in the
idle state when no request is detected.

Fix #8139

Signed-off-by: Michael Schaffner <msf@google.com>