[otbn] Plan and track coverage for external CSRs
By testing the external CSRs, we test both the interface of OTBN to
Ibex, as well as ensuring that OTBN is behaving as expected depending on
the operational state it is in. For example, if OTBN is locked or busy,
writes to CMD should be ignored. The functional coverage ensures that we
see a write to CMD in every operational state.

The coverage for the error registers isn't exhaustive, as it doesn't
explicitly require multi-error cases to be covered. There is already
coverage for such errors as part of the per-instruction coverage;
replicating it seems unnecessary.

Fixes #8024

Signed-off-by: Philipp Wagner <phw@lowrisc.org>