[dif/gpio] added autogen'd GPIO IRQ DIFs
This partially addresses #8142.

Note, the GPIO IP can raise 32 different IRQs (one for each GPIO pin),
however, the HJSON lists this as a single interrupt with a width of 32.
As a result, the autogen DIF templates and script had to be updated  to
accomodate this subtle difference.

Additionally, in the process of accomplishing the above, a very minor
bug was found in the unit test template that was fixed, resulting in the
regeneration of the UART IRQ DIFs.

Signed-off-by: Timothy Trippel <ttrippel@google.com>