[clkmgr] Add clock-gated indication signals
This adds "clock gated" indication output signals that are
encoded using the `lc_tx_t`. These signals will be wired
over to the alert handler in a subsequent PR in order to
connect them to the alert low-power groups (LPGs).

Note that these signals will be passed through 2-stage
synchronizers on the alert handler side. The timing of
these signals does not have to be cycle accurate.

Note, this is dependent on #8417 (first commit will disappear).

Signed-off-by: Michael Schaffner <msf@google.com>