[rstmgr] Add reset asserted indication signals
This adds "reset asserted" indication output signals that are
encoded using the `lc_tx_t`. These signals will be wired over
to the alert handler in a subsequent PR in order to connect
them to the alert low-power groups (LPGs).

Note that these signals will be passed through 2-stage
synchronizers on the alert handler side. The timing of
these signals does not have to be cycle accurate.

Signed-off-by: Michael Schaffner <msf@google.com>