[englishbreakfast] Add missing dependencies on crossbar
c58567302a7caec71bfa9749fd7c78252d1915b2 changed the dependencies for
the crossbar to use unique names per toplevel. The required change was
done for the earl grey toplevel, but not for english breakfast. Adjust
the two.

This problem wasn't apparent before because the use of the topgen
generator created a dependency on all cores in the output folder, i.e.
everything that topgen produced.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>