[opentitantool] Support configuration files
This changes introduces the TransportWrapper struct, which as the name
suggests wrap the existing Transport trait, while containing extra
information about the target device, obtained from a set of
configuration files specified on the command line. This combined
struct is passed to each Command implementation, instead of the lone
Transport parameter.

I have also replaced the Gpio interface published by the Transport
trait with a GpioPin interface, such that commands would ask the
TransportWrapper for separate GpioPin objects for each pin they want
to manipulate. The advantage of this is that the alias handling can be
done in a single method on TransportWrapper, which every app uses for
getting their hands on an instance of GpioPin, while GpioPin can
declare new set_drive_strength() or similar methods on top of read(),
write() and set_direction() without the TransportWrapper needing to
know.  (With the previous code, each invocation of the aforementioned
methods would need to take pin aliases into account, either by each
transport using a common helper method from the wrapper (backwards
dependency), or by the wrapper declaring a wrapping method for read(),
write(), etc., which would do the alias handling before passing
control to the equivalent method on the underlying transport.)

Also, the two sample .json files demonstrate how I propose knowledge
of strappings be removed from the Transport trait, and coded into
configuration instead. The benefit would be that we can have configuration
files specific for the combination of chip under test (OT or Dauntless)
and the transport. (It would not be reasonable that each transport would
know about both OT strapping and Dauntless strapping.) I have not yet
implemented logic in TransportWrapper that would expose a way of
selecting one particular strapping, and would set up a sequence of
GPIO pins according to the particular section of the configuration, though.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>