[dif/usbdev] Decouple buffer pool from device handle.
The USB device DIF library maintains a buffer pool to store incoming and
outgoing packets. This buffer pool was coupled with DIF handle/context
struct: `dif_usbdev_t`.

This commit decouples the buffer pool from the device handle to enable
the device handle struct to be auto-generated, and therefore enable IRQ
DIFs to be auto-generated, achieving a longer-term goal set forth in #8142.

Signed-off-by: Timothy Trippel <ttrippel@google.com>