[chip,dv,sw] Add SW timeout functionality
This commit adds 2 inlined functions in `ibex.h` to support a timeout
detection capability. The timeout is set in microseconds.

In addition, it adds a convenience macro that spins on an external
expression, throwing an error and aborting immediately if the timeout
occurred waiting for that expression to be true.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>