[prim] Anchor buffers around register flip flops
This gives us a natual anchor point if it is ever required.
Ideally, we would directly swap the rtl code with prim_flop_en.

However, doing so creates complications for DV's backdoor loading,
as that does not work well across module hierarchies.

Instead we add anchor buffers both on the write enable and the output
of the flop, which are the two areas we would most likely be concerned
with.

This still gives us a close by anchor point and does not impact existin
DV.

Signed-off-by: Timothy Chen <timothytim@google.com>