[rv_timer] Make interrupt names more descriptive.
The RV timer can raise and IRQ for each timer within each hart. In the
current configuration of the RV timer, we implement one timer and one
hart, so the name of the (single) resulting IRQ signal is:
`timer_expired_0_0`, where the `0_0` portion signifies hart-0, timer-0.

This commit fixes #8681 by changing the IRQ name above to:
`timer_expired_hart0_timer0`.

Signed-off-by: Timothy Trippel <ttrippel@google.com>