[dif/alert_handler] Remove need to pass in HW params.
This partially addresses #8409, with respect to the Alert Handler.

Templated IPs (which have the `templated` attribute in the `<toplevel>.hjson`
file) **_may_** have DIFs that require extra bits of information related to
the specific toplevel instantiation of said IP for DIF arg-checking purposes.
This toplevel instantiation specific information was most recently
encapsulated in the `dif_<ip>_params_t` struct, which was manually defined
in the Alert Handler's DIF header file, and passed in as an argument to
various DIFs. However, the information contained in this struct is
already automatically generated in the `alert_handler_regs.h` header
file. To reduce usage complexity, the `dif_alert_handler_params_t`
struct was deprecated, and the required parameter information is used
directly from the `alert_handler_regs.h` header file throughout the
various Alert Handler DIFs that require this information.

Signed-off-by: Timothy Trippel <ttrippel@google.com>