[clkmgr] Add params to HJSON so they can be used by DIFs.
This fixes #3932 and partially addresses #8409.

Some clkmgr DIFs check if the user passed in a valid bit index for the
CLKMGR_CLK_ENABLES_REG and CLKMGR_CLK_HINTS_REG registers. These bit
indices are used to gate a specific clock, or hint a specific clock should
be gated, from SW (i.e., using a DIF). Currently these indices are
defined in both `top_<toplevel>.h` and in the auto-generated (at
build-time) `clkmgr_regs.h`. However, the **number** of gateable and
hintable clocks is only defined in the `top_<toplevel>.h`, which cannot
be used in DIFs since it is toplevel-specific. The current work-around
for this issue is to pass these bits of information in at run-time to
the DIFs that use them, as described in #8409.

The first step in addressing this issue is to add local parameters to the
clkmgr's HJSON, so that the **number** of gateable and hintable clocks
are defined in `clkmgr_regs.h`. This commit implements this first step
as a part of a larger set of changes to address #8409.

Signed-off-by: Timothy Trippel <ttrippel@google.com>