[dif/alert_handler] Remove need to pass in HW params.
This partially addresses #8409, with respect to the Clock Manager.

Templated IPs (which have the `templated` attribute in the `<toplevel>.hjson`
file) **_may_** have DIFs that require extra bits of information related to
the specific toplevel instantiation of said IP for DIF arg-checking purposes.
This toplevel instantiation specific information was most recently
encapsulated in the `dif_<ip>_params_t` struct, which was manually defined
in the Clock Manager's DIF header file, and passed in as an argument to
various DIFs. However, the information contained in this struct is
was added as an autogenerated param in the `alert_handler_regs.h` header
file in the previous commit. Therefore to reduce usage complexity, the
`dif_clkmgr_params_t` struct was deprecated, and the required parameter
information is used directly from the `clkmgr_regs.h` header file throughout
the various Clock Manager DIFs that require this information.

Signed-off-by: Timothy Trippel <ttrippel@google.com>