[otbn] Flop STATUS register
This ensures that the STATUS register switches from IDLE to
BUSY_EXECUTE at the start of an operation on the same cycle as we zero
out INSN_CNT.

The STATUS change is also delayed by a cycle at the other end of the
run. Thus we rename the derived "done" signal in the model from
done_r_o to done_rr_o, showing that it's now delayed by two cycles.

The only other change is to add an expected stall cycle to the count
in stats_test.py, since there really is one extra stalled cycle that
we count.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>