[rom_ctrl,dv] Fully disable the scoreboard for CSR tests
We were previously seeing occasional failures in tests like
rom_ctrl_csr_hw_reset. The problem is that the test does a few memory
transactions and then ends. But rom_ctrl might not have finished its
work!

The scoreboard is "disabled" by setting en_scb = 0 in the base
environment, but we don't actually disable it: everything is still
connected up and running. Thus the scoreboard itself needs to know to
drain and ignore all analysis fifos and not do any checking.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>