[ spi_host ] Flag non-allowed write events as errors
The SPI_HOST TXDATA register/memory window is one of the few
registers in opentitan which support byte enables. However
the IP is not meant to deal zero-byte events.

Though such events are not expected, they are also not supported
by the SPI_HOST programming model.

This commit monitors the bus for such events and records an error
if such an event is detected.

Furthermore, events that do NOT correspond to any possible software
access pattern (i.e., not single-byte, two contiguous bytes, or full-word)
are also flagged as errors. The intent here is to simplify coverage as well
as provide further checks against possible bus faults.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>