[aes/rtl] Avoid using prim_packer_fifo in masking PRNG
If the EntropyWidth parameter is equal to the width of the individual
LFSRs in the masking PRNG, it is not necessary to buffer the entropy
and then reseed the whole PRNG in one shot. Instead, the individual
LFSRs can be reseeded as the entropy is delivered by the EDN - one LFSR
at a time. This reduces gate count.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>