[otbn,dv] Fix timeout when we inject an IMEM error early in a run
Before commit c559797, we only signalled an expected alert when we saw
a change to STATUS. That didn't work once we'd flopped the STATUS
register because it changed at the same time as the alert going out.

Unfortunately, there's also a problem with the "fix" in that commit.
It changed to sending the "expect an alert now" signal as soon as we
injected an error. However, if we inject an IMEM error just after
starting OTBN then it might be a while before we get URND data and
start trying to fetch stuff (and see that IMEM is invalid). In that
case, it's a while before we see the alert.

Calling set_exp_alert has two effects. Firstly, it allows an alert to
come in without causing an error. Secondly, it sets a timeout,
requiring an alert to come in by then in order to avoid an error.
Here, we set max_delay to some large number of cycles (10k) to
essentially disable the second of these effects. However, our ISS will
stop soon afterwards, so we can tighten things up again by requiring
that an alert goes out at most 10 cycles after that happens.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>