[dv] Slightly generalise run_stress_all_with_rand_reset_vseq
This task was originally designed for running stress sequences,
selected by plusarg. More recently, we taught it to take an optional
sequence argument, which would be used instead of the stress sequence.

Tidy things up a bit, moving the generic stuff that takes a sequence
to a new task (run_seq_with_rand_reset_vseq) and putting the plusargs
handling in a wrapper task.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>