[ottf] Added test status reporting and teardown logic.
When an on-device test completes, the OTTF now reports the test result
over the same interface as the existing test_main.{h,c} framework to
automatically terminate the test. This was accomplished by wrapping the
test_main() test entry point function in a wrapper function that is the
entry point to the FreeRTOS task the test runs in. The wrapper function
is used to report the test status. Additionally by using a wrapper
function, we can maintain the exact same function prototype used for
existing on-device test to minimize overhead migrating chip-level tests
to the OTTF.

Signed-off-by: Timothy Trippel <ttrippel@google.com>