[otbn,sw] Combine input and output buffers for RSA code
The existing RSA code had 4 * 512 bytes = 2kb of buffers, plus 256
bytes of config data, all of which needed to be accessible over the
bus. This is a problem because we're trying to cut down the
bus-accessible portion of DMEM to 2kb.

Fortunately, two of these buffers are "input" and "output". Merge the
two into an "inout" buffer and add a (non-globally-visible) work_buf
buffer on the OTBN side. Now, OTBN encrypts or decrypts the contents
of inout, writing to work_buf. Once that's done, it copies that buffer
back to inout.

This will add ~53 cycles to each operation, which is a tiny fraction
of the total time.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>