[dv] Allow using memutil_dpi_scrambled even without prim_ram_1p_scr
This is not going to be very useful at runtime, but we want to be able
to build things that may or may not use prim_ram_1p_scr, depending on
a parameter. If we're using it, we'll want to use this library.

Since fusesoc can't do things like "include this file iff parameter P
has value V", we'll need to include the library unconditionally.
Unfortunately, that then means Verilator simulations don't
link (because prim_ram_1p_scr and its DPI helper functions get
discarded from the design). Making the symbol weak moves the
resolution check to runtime, fixing things.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>