[sw/test_rom] Skip SRAM request for new key and LFSR init in test ROM.
The test `boot_rom` initialization code was requesting and new SRAM
scrambing key and re-initializing SRAM to pseudorandom values in the
ROM init asm. This wastes sim cycles as this is not production mask ROM,
and is used solely for testing.

This commit removes this extra initialization, addressing a task in #9163.

Signed-off-by: Timothy Trippel <ttrippel@google.com>