[rom_ctrl,dv] Scramble when randomizing ROM contents
This ensures that the ECC bits actually match what we expect, which is
needed in the smoke sequence. There's a design bug which means that we
were discarding those bits and replacing them with ones that are
correct. Oops!

Putting the right checksum in place means that we can now fix the
design without everything breaking.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>