[entropy_src] adapted for KMAC/SHA3 sparse FSM encoding
WARNING: The state width of sha3 was changed.
Thus, the width of sha3_fsm_o was changed as well.
This results in a width missmatch between
 - sha3_fsm_o
 - sha3_fsm
 - hw2reg.debug_status_sha3_fsm.d
As a quick fix this signal was disconnected from the debug reg.

A proper fix requires a change in the register file. It should be
discussed if a sparse fsm encoding should be exposed at all.

Signed-off-by: Michael Tempelmeier <michael.tempelmeier@gi-de.com>