[otbn/util] Add information-flow data structures for OTBN instructions.
These consist of two basic structures: InsnInformationFlow represents
the information flow of an instruction as represented in the YAML, with
tests based on operand values and following information-flow rules.
InformationFlowGraph represents a fully instantiated information flow
with register names, which can be derived from InsnInformationFlow once
operand values (and the register contents for any indirect references)
are given.

Signed-off-by: Jade Philipoom <jadep@google.com>