[spi_device] Add Address 4B mode tracker
In this commit, I designed a module inside SPI_DEVICE to track the
changes of the address size in flash and passthrough modes.

address field in the SPI flash commands is 3B or 4B based on the command
received. EN4B changes the address mode to 4 bytes. EX4B sets the
address size to 3 bytes.

The module `spid_addr_4b` receives the EN4B/EX4B commands from the
cmdparse and changes the value. If the value is changed, it also creates
an event to update the address mode in CSR too.

SW may change the value by updating the CSR (CFG.addr_4b_en). The event
propagates to SPI clock domain when a new SPI transaction begins.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>