[otbn] Some reformatting changes suggested by verible formatter
These seem to match our style guide and to improve the general layout.
There are some other changes that verible suggests at the moment that
look a bit more suspect and we might want to tweak it for. Those
changes are not included in this commit.

These come from running

    find hw/ip/otbn/rtl -name '*.sv' | \
      xargs util/verible-format.py --inplace -f

with the most recent version of Verible (v0.0-1784-g9f0278ee).

I've only included 10 files in the commit because reviewing this sort
of thing is a bit painful, so I thought it might be better to do it in
batches!

All of the changes are shuffling around whitespace. To see this, run
something like `git diff -w HEAD^..HEAD`. That should give us
confidence that the commit won't introduce any RTL bugs.

Other than little typos, most of the changes here are to do with
formatting port lists, both in the declarations of modules, and in
their instantiations. For better or for worse, we've agreed on a
format in our style guide; these changes tweak our code to match that.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>