[otbn,dv] WSR cleanups
When starting an operation, not every register should be reset. We
were already special-casing RND, but we also need to allow the key
sideload registers to persist. Tidy things up a bit by calling an
on_start method instead of re-initialising. The only change to
behaviour here is that we don't bother re-initialising URND any more.
It will get a new value when we get a seed from the EDN, so there's no
need to do anything clever until then.

Also, tidy up the way we set sideload keys. Make the change take
effect immediately to match the zero-latency update in the RTL. And
slightly simplify the representation in the SideloadKey class, which
I'd made overly complicated.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>