[otbn,dv] Fix bug in fatal alert tracking
I fumbled the change in 756e5b5 that allowed a sequence to see whether
we were waiting for an alert. I correctly split fatal_alert_expected
into two signals ("allowed" and "expected"), but I forgot to use the
"allowed" flavour in wait_for_expected_alert.

The result is that a re-triggering fatal alert causes a test failure.

As well as fixing that, this commit also tidies up a comment next to
the logic that clears the fatal_alert_expected flag.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>