[dif/alert_handler] Add class configuration DIF.
This commit adds a DIF to configure the escalation protocol of a
specific alert class. It enables fine-grain configuration control for a
specific alert class, rather than requiring all classes be configured in
a single shot, which was previously the case.

Additionally, it resolves a bug in a the previous alert class
configuration DIF which did not program the phase --> escalation signal
mapping correctly.

This partially addresses a task in #9899.

Signed-off-by: Timothy Trippel <ttrippel@google.com>