[bazel] Add bazel to CI and containers
bazel and gcc-9 is added to azure
bazel is added to Docker
ran buildifier

Updated instructions to ask for at least gcc8 the behavior of which is
already depended on by source in this repo.

skip verilator in bazel builds and tests in CI
skip targets that depend on verilator.
add a target so bazel can link it properly in sigverify_dynamic_functest

Signed-off-by: Drew Macrae <drewmacrae@google.com>