[dv/clkmgr] Add measurement saturation and fix timeout
Add code to set the measurement cnt to a high value to test counter
saturation.
Fix measurement timeout test to stop the clock after the measurements
start to make sure the CSRs are updated, and to expect all io derived
clocks to timeout when the primary io clock is stopped.

Signed-off-by: Guillermo Maturana <maturana@google.com>