[spi_device] Connect SRAM arbiter grant signal
SW access port (TL-UL) can handle the grant signal. Other requesters
from the HW can't handle. The assertion catches the case but the better
approach is to lower the SW access priority in prim_sram_arbiter and
also block the requeset when SPI transaction is active.

In this commmit, the SW access has lowest priority. When HW (upload)
requests, the Sw loses the access.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>