[dif/alert_handler] Add ping timer enable DIF.
The previous commit adds a DIF to configure the ping timer. This DIF can
also be used to enable and lock the ping timer too. However, if a user
wants to only configure the ping timer, and enable it later, they would
have to write to the time out register again too.

This commit adds another DIF that just enables (and locks, if requested)
the ping timer. This enables more fine grain configuration of the alert
handler's ping timer.

This addresses a task in #9899.

Signed-off-by: Timothy Trippel <ttrippel@google.com>