[reggen] Add support for validation of RTL CM annotation
This adds support for extraction and validation of countermeasure IDs
that are embedded in RTL comments, following the format

  // SEC_CM: [INSTANCE].CM_ASSET.CM_TYPE[, <more countermeasures>]

If a block has countermeasures declared in the Hjson, reggen greps all
associated RTL files of this block. It then validates whether the annotations
follow the correct format, followed by a check against the Hjson
countermeasure list.

Signed-off-by: Michael Schaffner <msf@opentitan.org>