[otbn,dv] Add ISS support for the lc_ctrl escalate signal
This would be a boring matter of adding some wiring except that we now
have to handle events when the ISS doesn't expect to be running. For
example, we might get an escalation event that comes in after an
operation has finished.

Rather than try to get the timing exactly right for when we need to
step the ISS or needn't bother, this commit alters the ISS so that it
would work properly even if we stepped it on every cycle and then
coarsely approximates when to step the Python, using a simple counter
on the SV side.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>