[otbn,dv] Reinstate resetting DUT at end of err vseqs
It turns out that you need to do this! Otherwise the locked DUT will
keep sending fatal alerts which, in turn, cause the fatal alert
monitor never to drop its objection and the test times out. Oops!

This basically reverts commit 001dace but strengthens the dut_init
call slightly so that it forces an actual reset (needed if we are
messing with the do_apply_reset flag in a stress sequence).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>