[otbn,dv] Avoid spurious MOD read entries in trace
If OTBN reset in the middle of a BN.ADDM or BN.SUBM then the second
term of the expression for ispr_read[IsprMod] stayed true while we
were in reset (because it depends on a non-reset flop), causing a
spurious

    < MOD: 0x00000000_000...0000

line once we started OTBN again after the reset. The fix is simple:
just qualify that term by insn_fetch_resp_valid. (Other terms are
qualified with that already because it's part of any_ispr_read).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>