[kmac/pre_syn] Add experimental Yosys synthesis flow
This commit adds an experimental Yosys synthesis flow for KMAC similar
to what we have for AES. The main purpose of this flow is to enable
formal masking verification using tools like Alma. At the moment, the
flow only supports synthesizing the keccak_2share module, i.e., the most
critical module containing the S-Boxes.

Ideally, we would provide a shared Yosys synthesis flow for modules that
need it but this is left for future work for after the ES.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>