[otbn,dv] Represent trace entries as a map to merge them better
This is useful for the trace entries that come from secure wipe.
Before this change, each register had a line representing it being
written with pseudo-random wipe value and also an line representing it
being zeroed.

Since we unconditionally sorted all the lines, this meant we couldn't
tell which write came last. What's worse, we don't actually want to
model the pseudo-random wipes in the ISS at all (because you'd need to
get the timing right in order to figure out the value of URND to use).

This commit changes our representation from a sorted list of strings
to a map of OtbnTraceBodyLine objects. These are minimally parsed from
their string representation so that we know the target of each
register write. That means that when we merge two entries, we can
ensure that the latest write to a location wins. As such, the merged
'Y' entry at the end of a secure wipe now looks like this:

    Y
    > ACC: 0x00000000_00000000_00000000_00000000_00000000_00000000_00000000_00000000
    > FLAGS0: {C: 0, M: 0, L: 1, Z: 0}
    > FLAGS1: {C: 0, M: 0, L: 1, Z: 0}
    > MOD: 0x00000000_00000000_00000000_00000000_00000000_00000000_00000000_00000000
    > x02: 0x00000000
    > x03: 0x00000000
    > x04: 0x00000000
    ...

Phew, much better!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>