[hw/dv] Changed set_response_queue to be umbounded due to queue overflow issues in Questa
As discussed in issue 9514 Questa was running into queue overflow issues when running uart_stress_all_with_rand_reset tests.
This is due to the driver put happening earlier than the sequence item get. It was aggreed that the check could be changed to be umbounded. The test now passes.

Signed-off-by: David Pudner <davidpudner@protonmail.com>