[otbn,dv] Run a check at start of wipe
With this commit, we check ISS vs. RTL state just as we start the wipe
and then again just after it has finished. That way, we can check the
end result of the operation and then check that everything was indeed
wiped to zero.

It's a bit silly that we run the check twice when secure wipe is
disabled (on consecutive negedges of the clock), but it doesn't really
seem worth fiddling around to avoid it because we'll be getting rid of
that mode as soon as everything works when it's enabled.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>