[dvsim] Add support for tags in testplan
This change allows testpoints to be annotated with tags so that the
testplan can be filtered as needed.

For example, a known set of tests can be run in both, DV and in
Verilator. Or, some set of tests may need to be run in GLS  / PA /
post-Si vectors. Rather than freshly create a separate testplan to
capture these needs (which will end up replicating the existing aspects
of the testpoint), tags allow us to represent these needs in the same
existing testplan.

This is just a proof of concept - the implementation may change if it
does not fit well with our requirements.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>