[dv/otp] Fix tl_intg_err
This tl_intg_err is because there are CSRs that could affect memory
access. Once the csr is set, if memory has integrity error, it will
return FFs, if there isn't integrity error, due to the memory is locked,
it will return 0.
It is hard to auto-predict this in auto-generated CSR test, so I plan to
exclude this register from write.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>