[usbdev] Enable defaulting to NAK for OUT transactions
Add a bit to change the behavior for OUT transactions to default to NAK
without software intervention, for safe communication of responses to
the host.
Before this commit, the device could incorrectly communicate acceptance
of a packet to the host if the firmware did not take action in time.
Defaulting to NAK gives the firmware time, since this is a nonbinding
condition--The device is merely saying, "Please try again later."

Clear the rxenable_out bit in hardware when a packet is received. Update
software code to match.

This change could have deleterious effects on performance in some cases,
but priority is given to safe handling of responses. For interfaces that
may perform transactions when a buffer is available, do not set the
set_nak_out bit for the corresponding endpoints.

Signed-off-by: Alexander Williams <awill@google.com>