[opentitantool] Use Result<> on all Transport methods
As we plan to introduce a Transport implementation backed by a network
protocol, I/O errors could occur even for the simplest methods (such
as "get baud rate").

This change ensures that every single method in the Transport trait
and its delegates return their value through a transport::Result<...>
to accomodate this future.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I88b3a48a1910cb3a5211c3dc6e84a8553300c53b