[kmac,dv] Remove assertion about last signal
The app interface has a valid signal. When that signal is low, the
interface should be ignored, so there's no need for an assertion about
the value of the last signal in this case.

I've just checked and I think the RTL is fine here: the arbiter in
kmac_app.sv handles the valid bits properly. Also, the agent that DV
uses to model the app interface sends Xs on these signals when valid
is low, so X propagation would catch if we'd done anything silly.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>