[aes] Hold EDN request until acknowledgment in case of fatal alerts
Previously, AES would drop a potentially asserted EDN request signal in
case of a fatal alert which could leave the EDN interface including CDC
in a weird state. With this commit, the request signal is always kept
asserted until acknowledgment which is similar to what other entropy
consumers do.

This is related to lowRISC/OpenTitan#10991.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>