[sw/test_rom] Re-enable SW-gateable clocks after reset
As #11081 points out, if a SW-gateable clock is gated, and the chip is
only partially reset (i.e., reset is NOT a POR), peripherals that the
test ROM uses (i.e., UART for logging and HMAC for bootstrap) could be
rendered in-operable.

This commit partially addresses #11081 (w.r.t. to the test ROM) by
resetting all SW-gateable clocks to their default POR values.

Signed-off-by: Timothy Trippel <ttrippel@google.com>