[opentitantool] Include context in conf errors
Errors like "Permission denied" or "JSON parse error on line 5" ought
to be shown in the context of which file path was attempted loaded
or parsed, in order to help the operator figure out what is the issue.

This CL introduces two new error Enum values, each of which carry a
file path, and an underlying error, for the above purpose.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I721bab68dbf511aa0397c84be5e8a213638ebdd7