[opentitantool] Pass capabilities via proxy protocol
Until now, the Proxy implementation of Transport has claimed to
support all of GPIO, SPI, I2C, etc.  And if a user attempted to get
e.g. an SPI instance and the remote session process had an underlying
Transport which did not support SPI, a TransportError would be
returned.

This change adds a GetCapabilities request/response to the protocol,
so that the Proxy object can properly advertise what the caller can
expect to be able to use it for.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I3391f4dfb739c0ec45dbcb7754a3494d18dc2138