[otbn,rtl] Rewire how we pass around error bits
Before this change, we passed lots of signals up to
top-level, where they were assembled into an "err_bits" struct and
then passed back down to the core. This had lots of scope for
accidental combinational loops because as soon as you qualify one
error with another, you've got a loop all the way around.

This commit tries to simplify things a bit. Now, the controller and
core have their own cut-down versions of the err_bits_t struct that
they pass out. We also need signals going "the other way" to tell the
controller when there has been a fatal error. Since these are
all (morally) local escalation signals, I've standardised on that
naming. Now, for example, otbn.sv defines core_escalate_en which
contains all the "extra" error signals that are OR'd into err_bits and
didn't come from otbn_core. This signal gets passed down to tell the
core to stop.

The net result should be that the flow of information is a bit
clearer (and more one-way!).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>