[otbn,rtl] Make sure INSN_CNT gets cleared at the right time
This commit reverts fc7f540c3e (where I added the "clear INSN_CNT on
an escalation" logic) because that takes effect one cycle too early.
To be precise, we'd much prefer it if we changed STATUS and INSN_CNT
on the same cycle: it's probably more understandable, and it's also
easier to model for DV.

(I think the original behaviour was probably right and my first commit
just broke it: oops).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>