[otbn,dv] Add dummy instructions to pairwise covergroup
This covergroup is supposed to be tracking pairs of instructions. I'd
carefully avoided putting ECALL in the "last_cp" coverpoint (since if
you get an ECALL, you'll not have a following instruction).

Unfortunately, I hadn't looked properly at the list of mnemonics and
had forgotten "dummy-insn" and "??" which are used for invalid
instructions and failed integrity checks, respectively. Add them to
avoid illegal bin reports.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>