[lib] Add our own implementations of low-level math polyfills
This includes a 64-bit long division function, and local polyfills for
popcount and bswap, which RISC-V does not always have instructions for.

The former must be called via a special signature, since it's an
expensive operation triggered by operator/ in C. The latter can be
called via their libgcc names as usual, so those callsites do not need
to be updated.

Signed-off-by: Miguel Young de la Sota <mcyoung@google.com>