[aes,dv] Split A-channel write handling into separate functions
The only change in behaviour should be that these functions now all
only run when the transaction is a write (before, some of them didn't
have the check that was probably supposed to be there).

However the split means there's rather less indentation which, in
turn, means we don't have to fight so hard to avoid the 100 column
limit.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>