[otbn,rtl] Remove FF for DMEM/IMEM Sec Wipe
Requesting URND keys at the same time with scrambling key request
from OTP should not be a problem and that is the only place where
we use the flopped signals.

Removing the flops also makes it so that we would have the same
number of delays in the case of imem/dmem_sec_wipe operation with
execute operation.

Signed-off-by: Canberk Topal <ctopal@lowrisc.org>