[otbn,dv] Signal WIPE_START one cycle before done.
This is for making sure that we are not getting a thrashed DMEM
when we try to check memories before doing the secure wipes.

Last commit makes it so that the wipe request comes one cycle
before the done signal so we need to commit the setting of the
WIPE_START flag before commiting the other ext_regs that signals
end of the operation.

Signed-off-by: Canberk Topal <ctopal@lowrisc.org>