[dvsim] Introduce Verilog compile-time seeds
This commit adds a mechanism within DVSim to introduce a compile-time
randomization of the design using a seed value that can be controlled
via the command-line.

Compile time randomization is achieved by setting a pre-processor
macro with the seed value, and using the macro to mutate some constants
in the design.

This mechanism is particularly useful to change constants that do not
have side effects in the design (such as addition of deletion of logic),
but are merely used to indicate some fixed values in the design, such as
a part of the secret keys.

We do not have a clean mechanism to change these fixed values across
successive compilations - this commit is one way to achieve that.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>