[prim] Fix a bunch of Verilator lint errors in prim_packer.sv
These are all to do with width conversion. Almost all of the changes
are just putting explicit widening casts where Verilog was doing it
implicitly before.

The one awkward place was the expression used to compute shiftl_data /
shiftl_mask. The eventual size was wrong (should be ConcatW, not
Width). But it's also a bit difficult to cast shifts and the resulting
expression was rather noisy, so I've split the expression into two
stages, which seems to tidy things up nicely. As a plus, this also
gives us a place to add a comment to document the step in the
computation.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>
Signed-off-by: Michael Schaffner <msf@opentitan.org>