[rtl] Convert some non-ANSI parameters to localparams
It turns out that there's a slightly obscure corner of the SV spec
that says that if you have a module like this:

  module foo #(
    parameter int Foo = 0
  ) (
    // ...
  );
    parameter int Bar = 1;
    // ...

  endmodule

then Bar essentially becomes a localparam: the presence of the ANSI
parameter list means that all non-ANSI parameters get treated as
localparams. Furthermore, some tools (Vivado, at least) warn when this
happens.

Convert these parameters to localparams, which is how they will behave
anyway.

This commit switches all occurrences that I found with

    git ls-files -- '*.sv' | \
      grep -v vendor | \
      xargs -n1 awk '\
        /module.*#\(/ {if(x==0) x=1} \
        /\)\s*\(/ {if(x==1) x=2;} \
        /^[^/]*parameter/ {\
          if(x==2) printf("%s:%s: %s\n", FILENAME, FNR, $0);\
        }'

(Note: there are a few false positives that this script finds in
aes_sbox_dom.sv, which has more than one module in the file).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>