[sw/testutils] Update adc_ctrl ISR testutil to clear status CSRs
In addition to the standard IRQ status register that must be
acknowledged when an IRQ is fired, the ADC Controller has two internal
CSRs, the filter status and INTR status registers, that must be
acknowledged (cleared). This was discussed in #11354, and determined
that future generations of the IP will remove this complexity. However,
until then, this commit will mask this complexity in software, specifically
in the autogenerated ADC Controller ISR testutil.

This addresses #11354.

Signed-off-by: Timothy Trippel <ttrippel@google.com>