[otbn,rtl] Use err_bits, not err_bits_q for alerts[AlertFatal]
This fixes a bug introduced by 99a459c where we'd only generate fatal
alerts when an operation finished. One problem with this is that
secure wipe takes ~100 cycles, so delays the alert. Another (more
serious) problem is that it means alerts are squashed altogether when
OTBN isn't running.

Note that the timing of alerts[AlertFatal] now matches that of the
FATAL_ALERT_CAUSE register, which should make a bit more sense.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>