[bazel] add script to prep for offline bazel builds
In order to build SW artifacts with bazel in an airgapped environment,
a distdir and cache must be prepared to hold a copy of bazel (and its
dependencies) and the OpenTitan bazel workspace dependencies. This adds
a script to prepare a directory to store both, so that both can be
easily transferred to the airgapped machine.

Signed-off-by: Timothy Trippel <ttrippel@google.com>