[ast,lint] Silence a Verilator warning about nonexistent latch
It seems that the "hey, you forgot to put in a latch" logic runs after
some constant propagation so the fact that we set scan_mode_i = 1'b0
means Verilator thinks there is no latch here:

    always_latch begin
      if ( !scan_mode_i ) sys_clk_byp_sel <= sys_clk_byp;
      ...
    end

We're doing this on purpose: there would be a latch if we weren't
forcing scan_mode_i off. So waive the warning.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>