[bazel] remove `git_repository()` repo rules
In order to support offline bazel builds on airgapped machines, we must
be able to prepare a repository cache with all external dependencies on
a machine connected to the internet, and then move this cache to the
airgapped machine and build with `bazel build --repository_cache=<cache>
<label>`. However, according to the bazel documentation (see
https://docs.bazel.build/versions/3.4.0/external.html#repository-rules)
only `http_archive` works with the repository cache, not
`git_repository` rules. Therefore, this commit removes all
`git_repository` rules from the workspace, and uses `http_archive` rules
instead.

Signed-off-by: Timothy Trippel <ttrippel@google.com>