[otbn,rtl] Pass an escalation signal to start/stop controller
Otherwise, an escalation that happens when we're in e.g. state
OtbnStartStopStateUrndRefresh leaves us in a weird half-running state.
The rest of OTBN knows that we're stopping immediately in this case:
make sure that the start/stop controller knows too!

The other change with this PR is that we change the
OtbnStartStopStateError state to OtbnStartStopStateLocked. An FSM
glitch will still ping state_error_o for a cycle, causing a fatal
alert, but now we have a nice terminal state for local escalation.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>