[flash_ctrl] Add checks for unexpected acks
- d2s item

Unexpeted acks are checked in two places, the controller access
path and the host access path.

For the controller access path, the accesses are fsm controlled,
so whenever the fsm is idle it should not see an ack.

For the host access path, a small amount of logic is added to track
the number of outstanding host transactions. When there are no
outstanding transactions, a response is not expected and would constitute a
spurious ack.

Signed-off-by: Timothy Chen <timothytim@google.com>