[bazel] detect airgapped mode in `pip_wheel` repo rule
In #12091, a custom bazel repository rule was setup to pre-download
/ build python wheels to transport them to an airgapped machine so the
`pip_install` rule does not attempt to go to the internet. This commit
enables the custom bazel repository rule previously added to detect if
it is executing in airgapped mode (via an environment variable), and if
so, use the pre-built/local Python wheels repo.

This fixes #12090.

Signed-off-by: Timothy Trippel <ttrippel@google.com>