[spi_device] Add WE set/clear logic to Status module
This commit revises spid_status module to allow other modules to set/
clear WEL bit in STATUS register.

This commit also revises the SW updation logic to ensure the atomicity
of the STATUS register from the host system point of view. The STATUS
that the host system sees is updated when a SPI transaction is
completed.

Any SW update of STATUS will be applied to the STATUS register during a
SPI transaction after going through an asynchronous FIFO then visiable
to the host system in following transactions.

It means that at the very first transaction, the STATUS is always reset
value (in current design all zero).

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>