[mubi/lc_ctrl] Change MUBI / lc_tx_t encodings
This swaps the encodings so that lc_tx_t uses 0xA / 0x5 as True / False
values, and MUBI uses 0x6 / 0x9 as True / False values.
While both of them are complementary, it is not possible to transform
the MUBI values into one another with a shift operation, which has
advantages for SW hardening.

Signed-off-by: Michael Schaffner <msf@opentitan.org>