[spi_device] Define a buffer address width
Previous design defined `threshold` as `[BufferAw:0]`. It is
syntatically correct but confuses contextually. It reduces readability.

In this commit, `SramBufferAw` is defined. It is the required bit width
of a Read Buffer size in bytes rather than the entries in DPSRAM.

This commit does not alter the width of the threshold. It stays 10 bits
as a Read Buffer size is 1kB.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>