[bazel/bitstreams] error out on unitialized cache access
If the bitstream cache is unitialized, and an access is attempted for
the first time in "offline" mode, the is a "invalid key" error when
attempting to access the latest bitstream entry in the cache. This
causes the bistream caching script and associated bazel repository rule
to error out with useful messages.

Signed-off-by: Timothy Trippel <ttrippel@google.com>