[bazel/rules_rust] use forked/patched `rules_rust` repo
As described in #12515, the upstream `rules_rust` repository attempted
to download several dependencies in a manner that did not first check
the repository cache before going to the network. This made supporting
an airgapped build environment problematic. To fix this issue, the
`rules_rust` repository was forked and patched. Please follow this link
for the commit message that describes the patch in detail:
https://github.com/lowRISC/rules_rust/commit/be0d6ca492f64cc8d460f54f467925ef2753ed89

This commit update the bazel dependencies to use our patched
`rust_rules` instead of the upstream ones. Additionally, this commit
updates the `prep-bazel-airgapped-builds.sh` script to pre-fetch
additional dependencies to pre-populate the bazel repository cache with
additional required dependencies to support airgapped builds.

This fixes #12515.

Signed-off-by: Timothy Trippel <ttrippel@google.com>