[flash_ctrl/dv] Split out a base hjson without setting tool
We want to use xcelium in open-source flash, while using another tool in
close source. See #12483

But tool can't be overridden. As we need to import all hjson before we
do the override, but one of the hjson file is {tool}.hjson, it causes a
chicken-egg problem.
split out a base hjson file which doesn't specify tool, to solve this problem.

Signed-off-by: Weicai Yang <weicai@google.com>