[otbn,dv] Correct the boundary of a "critical section"
The running_ flag is used to handle process control in
start_running_otbn. The idea is that it gets set when we start
run_otbn() and then cleared just before we exit. That lets us get the
timing right before running a "disable fork" in start_running_otbn.

A series of commits starting with c908164 had added some extra stuff
that run after we cleared the flag, causing occasional weird errors
where a UVM sequencer spots that a process has been killed while
waiting to send a sequence item.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>