[dv/top] Adjust test parameters
The pause test is currently failing because the test_rom takes
too long to execute and the watchdog triggers sooner than the
test expects. As a result, the test sees two reset reasons at
the same time and fails.

While we could adjust the delays such that it is passing again,
such tests may be prone to failures in the future.

Instead, the test was converted to not rely on resets but instead
on the watchdog / wakeup counts.  The pause function causes the
watchdog to stop counting during low power, and thus creates an
inverse relationship with wakeup.  The test checks on that instead.

Signed-off-by: Timothy Chen <timothytim@google.com>