[otbn,dv] Speed up loading ELF files through front-door
Doing front-door loads of an ELF file can be a surprisingly large
fraction of the total simulation, especially for sequences like
otbn_single.

Speed this up (90% of the time) by configuring the TL agent to run at
full speed. This drops a typical load from ~45us to ~16us simulated
time.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>