[otbn,rtl] Drop pending RND prefetch at end of operation
Consider a situation where we have two operations, A and B, that run
back-to-back. Suppose operation A writes to RND_PREFETCH and then
finishes before the EDN replies. Now suppose that operation B also
writes to RND_PREFETCH. In this case, the RTL uses a signal called
rnd_req_queued_q to track the fact that it should send out another RND
request once the first (ignored) one comes back. But what if operation
B also finishes before the EDN has replied to the first request?

Once the result comes in, there was a slight mismatch between how the
ISS and RTL handled things (depending on the timing of the *next*
operation, the ISS might or might not have sent out another request).
But this whole situation is silly! We know we don't care about the
result of the second prefetch, so we don't need to do it at all.

This patch tweaks the RTL to drop the pending request in this
situation and adds a "forget" function in the ISS to do the same.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>