[otbn,rtl] Get rid of rnd_prefetch_pending_q
This wasn't quite computed correctly, because it wouldn't be cleared
if a prefetch completed when edn_rnd_data_ignore_q was true (since the
data was ignored, rnd_valid_q wouldn't go high).

It turns out that we don't actually need the signal anyway: it was
only used to spot the situation when we are starting an operation with
a pending request, but we can use edn_rnd_req_q there to do the same
job.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>