[otp_ctrl] Prevent broadcast of scrambler's input/intermediate values
The scrambler broadcasts initial inputs, intermediate scrambling
results, and final scrambling results to its data output.

This commit adds an additional condition that only enables the final
scrambling results or a safe default value (as recommended in the OT’s
Secure Hardware Design Guidelines) to be broadcast to the data output.

Signed-off-by: Andres Meza <anmeza@ucsd.edu>