[ci] Reduce the number of CW310 tests being run in CI
The communication with the CW310 board in CI fails intermittently. This
is happening more frequency after more tests were added during the Bazel
transition. This commit is a temporary measure to reduce the frequency
of these failures by running fewer tests (the same set that was
previously being run by systemtest).

Signed-off-by: Miles Dai <milesdai@google.com>