fix(otp_ctrl): Keymgr Key valid glitch
This PR fixes the CDC issue on otp_keymgr_key_t.valid signal. Also
there's chance the key_share0, key_share1 can be changed during the
operation but the chance is limited to the escalation stage, in which
case the CDC is not meaningful anymore.

So rather than latching the key_shares, this commit adds assertion to
check the stability of the key_shares during the normal operation.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>