[entropy_src/dv] bit_sel_en timing adjustment
In bit_sel_en the DUT has a timing idiosyncracy in which
the selected RNGs bits remain in the bit select pfifo until
one more RNG sample is acquired.  This does not violate the
spec in any way, but can lead to test failures if the
scoreboard is expecting the last sample to be included in
the health checks.

This commit adjusts the scoreboarding to mimic this feature
and properly predict which samples are included in the health
test statistics in bit select mode.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>