fix(kmac): Relax KeyLen Assertion conditions
@weicaiyang reported an issue
https://github.com/lowRISC/opentitan/issues/13594

When an Application Interface using KMAC initiates the hashing
operation, the `kmac_app` prepares the KMAC configs (strength, mode,
keylen, keydata) for the specific App. At the same time, the App FSM
asserts the `start` signal for KMAC core to run the SHA3 logic.

This violates the KeyLengthStable_M assumption as the kmac_core already
moved to `StKey` state when `key_len_i` is changed.

This commit relaxes the condition by checking the previous value of `st`
is `StKmacIdle`.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>