[otbn,dv] Remove D/Q separation of model-internal registers
Prior to this commit, `otbn_core_model` separated the next state and the
current state of model-internal registers.  In the reset section of an
`always_ff` process, the SV code reset the `_q` part to zero after
resetting the model.  In the non-reset section of that process, the SV
code let the model step on the `_d` part of the register, which was
*not* reset by the SV code.  This mechanism thus relied on the model not
using any of the `_d` signals as inputs.  A subsequent change, which
checks the status before deciding whether to execute a command, will use
the status signal as an input, though.  Without this commit, that change
would fail.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>