[otbn,dv] Reset model status in scoreboard when reset is released
`otbn_scoreboard` receives status transitions of the OTBN model from
`otbn_model_monitor`.  On a reset, the monitor waits until the reset is
released before sampling the status (as the status is not guaranteed to
have settled directly on the falling reset edge, see 320eb75250).  The
scoreboard thus on a reset only receives a transaction to the initial
status (Idle), and it cannot know if the model entered the Idle state
due to a reset or because OTBN completed execution.

This commit fills this knowledge gap of the scoreboard by adding the
value of the reset signal (before release) to the status transition.  If
the status transitions due to reset, the scoreboard resets its mirror of
the model status as well before using it for checks.

As a bonus side effect, this removes the need to keep the initial value
of the model status in the scoreboard in sync with RTL and model code.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>