[chip/conn] Move a DV testplan entry to connectivity testplan
This PR removes a DV testplan entry (EDN clock and reset check in OTP)
to a connectivity check. We discussed this topic in one of the DV
meeting and found that there aren't any check points in simulation due
to the clock and reset groups. So we move this to a connectivity check.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>