[entropy_src/dv] Scoreboarding (and stimuli) for bad Mubis
With this commit the scoreboard can now comprehend bad MuBi entries,
as well as bad alert count thresholds.  Meanwhile the DUT_CFG
structure contraints are expanded to randomly inject bad Mubi
configurations, while the RNG and Base vseqs can recover
from them.

The goal of this commit is to not just confirm that the DUT can flag
bad register settings in the recov_alert_sts register, but also
that the alert is triggered when expected.  Furthermore the relevant
coverpoints for these alerts are only sampled when the alerts
are expected and seen.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>