[otbn,dv] Initialize destination WDRs of `bn.mulqacc.so` in `smoke_test`
If OTBN's model does not know the value of a register after reset and
the initial secure wipe, which will be the case when registers are not
zeroed as part of the secure wipe anymore, the effect of the
`bn.mulqacc.so` instruction on the destination register currently can
not be modeled correctly if the destination register has not been
written before.  This is because the `bn.mulqacc.so` instruction writes
*half* of the destination register while leaving the other half
untouched.  The model does not know the value of the half that is
untouched though (and instead sets it to zero).  This results in a
mismatch in the traces.

This commit works around this limitation in OTBN's `smoke_test` by
initializing the destination WDRs of the `bn.mulqacc.so` instructions to
zero (with a `bn.xor` instruction).

Signed-off-by: Andreas Kurth <adk@lowrisc.org>