[otbn,dv] Remove check when simulation ends
A complete simulation (i.e., one not terminated prematurely by a failed
assertion or another error considered fatal) always ends with a secure
wipe.  After a secure wipe, we cannot check the register content because
it contains random values that the model does not know so the check
would fail.  This commit removes that check.

Register content is still checked for correctness *before* the secure
wipe directly after execution.

Removing the check in this commit does not mean that the model no longer
checks that the secure wipe happens: that check is still covered by
comparing the changed registers in the traces that are produced by the
model and RTL during secure wipe.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>