[otbn,dv] Allow model registers to be invalid
In an upcoming change, secure wipe will not finish with zeroing out most
registers but instead leave random values.  These random values are
currently not being reproduced in the model.  Thus, instead of
reproducing random values, we will set the registers of the model to
invalid after a secure wipe.

This commit adds functions to invalidate the model registers.  As with
writing a value, the `write_invalid()` functions produce a change that
appears in the trace.  `x`s are used to represent an invalid value in
the trace.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>