[dif/entropy_src] add DIF to check if FIFO is full
This adds a DIF (and corresponding unit tests) to check if the firmware
override FIFO is full and cannot be written too.

Additionally, this updates the FIFO write DIF to check if the FIFO is
full before attempting to write to it, which under the previous
implementation would fail silently.

Signed-off-by: Timothy Trippel <ttrippel@google.com>