[otbn,dv] Alternate between two seeds for RND/URND
Prior to this commit, `standalonesim.py` of the OTBN simulator and the
Verilator TB returned a fixed value as EDN data.  This implied that random
values were identical after every reseed.  An upcoming change will reseed
between two rounds of securely wiping with random values, and if both
rounds write the same values, the trace check fails (which is intended).

This commit alternates between two different seeds for RND and URND, so
the two rounds of secure wipe will write different random values.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>