[doc] Unlist dangling pages from menus.
This commit introduces an 'unlist' parameter which can be used to remove
dangling pages from menus to make the navigation experience more
intuitive.

Dangling pages make menus harder to follow as there is no intuitive
ordering. For example, the hardware section had the following menu:

hw
 - Design Verification
 - OpenTitan Assertions
 - IP Cores
 - Linting
 - Top Earlgrey

 After this change the menu is as follows:

 hw
  - Design Verification
  - IP Cores
  - Top Earlgrey

This change does not affect the siteContent fragment, so it can still be
used to get a full list of pages.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>