[sw/silicon_creator] Fix KMAC entropy config for keymgr functest
Despite a comment saying KMAC was configured to use software entropy,
it was actually using idle_mode (reset configuration). However, the sole
purpose of this mode is to enable ROM ctrl right after coming out of
reset. No reseeding is happening and the PRNG is not advanced in this
mode. It thus shouldn't be used after ROM ctrl has finished.

This commit changes the configuration to indeed reseed the PRNG with
entropy provided by software. In addition, message masking is switched
on to reduce potential SCA leakage.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>