[prim_count] This reworks the primitive to make it more generic
This reworks the counter primitive and makes it more generic.
In particular, it supports both up and down counting, and there is no
need to set the counter mode anymore (it will always use cross counter
mode). The only remaining parameter is the set-mode which determines
which of the two counters is being loaded (the other counter will be
adjusted accordingly).

All SVAs have been reformulated so that the counter primitive can be
verified completely in FPV.

The patch also updates all instantiations of prim_count throughout the
design in order to make the patch atomic.

Signed-off-by: Michael Schaffner <msf@opentitan.org>