[silicon_creator/spi_device] Fix fault injection bug in spi_device_cmd_get()
We discovered a fault injection bug in spi_device_cmd_get() in
sw/device/silicon_creator/lib/drivers/spi_device.c, in which the code
copies the number of bytes given in a 9-bit byte count from the
PAYLOAD_DEPTH field of the SPI device's STATUS2 register into an
output buffer of size 256.  While the hardware ensures always that
PAYLOAD_DEPTH <= 256, a fault injection causing a flip of one of the
PAYLOAD_DEPTH bits causes a stack buffer overflow.

We fixed this bug by (1) deriving the size of the
spi_device_cmd_t::payload buffer from the size of the SPI device's
payload buffer, and (2) performing a hardened bounds check on the
PAYLOAD_DEPTH field read by the SPI device driver.

Signed-off-by: Nicholas Mosier <nmosier@google.com>