[crypto, test] Add and apply new Bazel rules for crypto tests.
The new rules allow crypto tests to be read from both hardcoded test
vectors and external data sources. Now, the large wycheproof test set
can run in CI for sigverify (FPGA only because it's too slow in
simulation).

To get the test to pass in CI, the Python script computing the modular
inverse constant for sigverify needed to be changed (as it used a
construct only available in Python >= 3.8).

Signed-off-by: Jade Philipoom <jadep@google.com>