[bazel] cleanup `opentitan_ram_binary` macro
The `opentitan_*` bazel macros were refactored / cleaned up while
another PR was inflight with additional changes to these macros. This
cleans up the newly added `opentitan_ram_binary` binary macro to match
the existing macros. Additionally, it removes the unneeded filegroup of
targets since the outputs of the targets instaniated by this macro are
designed to be embedded in an `opentitan_functest`.

Signed-off-by: Timothy Trippel <ttrippel@google.com>