[chip dv] Add xbar tgl excl cfg to all build modes
This commit adds the xbar-autogenerated tgl excl
coverage cfg file to all build modes. This is technically
not needed if the vdb associated with the primary build
mode is supplied first to urg, when merging the coverage
db. However, this makes the coverage associated with
cover_reg_top only vdb more accurate. This is useful when
analyzing coverage data collected only from the non-functional
tests.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>