fix(spi_device): Invalid command in Flash mode
Issue is reported by @weicaiyang in
https://github.com/lowRISC/opentitan/issues/14057.

The cmdparse FSM supposes to move to `StWait` state if an invalid
command has been received. The condition was missed in the previous
design. If an invalid command is received, `cmd_info_d.valid` is 0.

In this commit, the behavior is fixed by checking `module_active &&
data_valid_i` in `StIdle`.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>