[otbn,dv] Fix timeout issues for ctrl_redun test
In some options we are sitting on a while loop waiting for a
specific condition to happen (e.g. ispr_rd_en). However, that
condition might not happen at the first place (because maybe
we don't generate the relevant instruction for that condition).

Using join_any in the fork solves that problem.

Signed-off-by: Canberk Topal <ctopal@lowrisc.org>