[ottf] enable OTTF to be loaded in any slot at silicon_creator stage
This refactors the OTTF linker scripts to enable the OTTF to be loaded
in any slot (A, B, and virtual) at the silicon_creator (ROM_EXT) boot
stage.

Additionally this makes the "manifest" field of the
`opentitan_flash_binary` macro mandatory when the `signed` field is set
to `True`, and fixes existing signed tests that did not provide a
manifest.

This partially addresses #10712.

Signed-off-by: Timothy Trippel <ttrippel@google.com>