[dvsim] Fix run timeout override in hjson
When recursively merging scalar attributes across modes
that enable sub-modes, dvsim requires the attribute to
be set uniquely once. For example, if run mode `foo`
enables run mode `bar`, then `foo` and `bar` cannot
both set a non-default value to a scalar attribute.
For example, the `run_timeout_mins` cannot be set to
40 in `foo` and `100` in `bar`. If that is found to be
the case, then dvsim throws an error and exits, because
it does not know which to pick. There is no notion
of priority across modes that are enabled recursively
into other modes.

`Modes.py` assumes `-1` as the default value for
integral attributes, because it is the "lowest common
denominator". The existing `Modes.py` set the
`*_timeout_mins` to 60 as the default value, if not set
in the hjson. The end result of this is - if `foo` does
not set `run_timeout_mins`, then for `foo` it gets set
to 60. If it is set differently in `bar`, then bam,
we end up having a problem.

The fix in this commit changes the default value to
`*_timeout_mins` to `None` instead, so that it can be
properly set to a desired value in modes, test
specifications and globally in the hjson. The Deploy
class then limits builds and runs to 60 minutes if
it was not set in the hjson.

This unfortunately does not completely fix the problem
because `build_timeout_mins` is not overridable in the
hjson at the moment. It will be fixed in a future PR
when the need / urgency arises. The whole `Modes.py`
code needs a significant refactor to ensure it works
as expected.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>