[flash_ctrl] Split hardware / software interface reads.
Fix issue identified in #14011

In the original design, the life cycle interface and software interface
shared the read data path. This works fine except in RMA entry.

Upon RMA entry, since the datapath is shared, the life cycle interface
clears out all the read FIFO contents so that it can take over the
FIFO.  However, this means any data deposited in the FIFO that software
was intending to read is now no longer valid.

Practically speaking, this is not a problem. As we really do not care if
a host can successfully read when RMA entry begins.  However, this
behavior creates a situation where it is difficult for DV to predict
hardware behavior around RMA entry without probing internal signals.

By splitting the read path (very little overhead), we allow the software
data already depositted to stay alive while the RMA wipe begins.

Signed-off-by: Timothy Chen <timothytim@google.com>