[chip dv xcelium] Fix macro recursion bug reported by Xcelium
This piece of logic was recently wrapped in `DV_SPINWAIT` loop.
This worked with VCS, but it did not with Xcelium. Xcelium reported
an interesting compile error stating that the first argument of this
macro `WAIT_` itself had a `DV_WAIT` invocation within, resulting in
infinite recursion of macro unpacking.

Looking at the code more, it does not seem like the wrapping is
needed in the first place, since the while loop internall has a
spinwait which will kill the sim if it exceeded our runtime.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>