[rom_ctrl] Switch padding scheme to zeroes
As elaborated in #5771, we decided to switch the padding scheme so that
the ROM is padded out with zeroes as opposed to pseudo-random data.
This has been deemed the better approach, since confidentiality is not
required, and padding out with zero makes sure that there are no
"gadget" instructions in invalid ROM regions, since zeroes are
interpreted as invalid instructions in Ibex.

Signed-off-by: Michael Schaffner <msf@google.com>