[sw] Update memory_perftest cycle count expectations
Unsurprisingly, memory_perftest is a bit fragile because it fails when
cycle counts are not equal to its hardcoded expectations. The test was
tagged as "flaky" and "manual" in PR #14044 because cfrantz@ saw
indirectly-related changes invalidated the test expectations.

This commit updates the expectations to match reality, but does not
remove the annotations.

Signed-off-by: Dan McArdle <dmcardle@google.com>