[bazel] add slot management to `opentitan_functest` macro
This adds support to the `opentitan_functest` macro to specify which
stage and slot an `opentitan_flash_binary` gets built for, defaulting to
the `silicon_creator` stage, slot A.

This partially addresses #10712.

This functionality will soon be expanded address building more complex
test cases with multiple stages / slots filled, to address #13511.

Signed-off-by: Timothy Trippel <ttrippel@google.com>