[otbn,dv] Unset `recov_alert_seen` on reset
Prior to this commit, the OTBN scoreboard would not reset the flag that
tracks if a recoverable alert has been seen (`recov_alert_seen`) on a
reset.  This caused errors (such as "Double recoverable alert seen") in
tests that reset before the recoverable alert has been handled (such as
`otbn_reset`).  This could be reproduced with `otbn_reset` and the seeds
`2209004860` and `542535943`.

This commit adds the missing reset.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>