[dv/alert_handler] Add a direct sequence to enhance ping mechanism
This PR is based on issue #14556 where it is hard to hit the timing
corner case in the current mask setting.
To be able to cover this case faster, in block level I create a direct
sequence test to simulate the chip level scenario but force the
`wait_cyc_mask_i` to a much smaller value.
This will increase the chance of finding the corner case like issue #14556.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>