[test, top_earlgrey] Adjust run options for otbn_ecdsa_op_irq_test
Now that these tests are setting up the entropy complex in auto mode,
RND and URND requests may take a longer time to be served which can lead
to frequent time outs. Therefore, this commit increases the
simulated-time time out and increases the RNG rate of AST for these
tests.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>