[dv/kmac] Change how scb fetches sideload key
The current scoreboard only fetches sideload key when the valid bit
dropped and reset to 1, but does not count the scenario where we change
the key value on the fly.
Changing the key value on the fly is not a valid operation, but we want
to test it to make sure it won't create any unexpected errors.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>