[sw,aon_timer_testutil] Add AON timer shutdown
The commit adds a method to shutdown the AON timer
peripheral, disabling the wakeup and watchdog
timers, clearing the wakeup cause register and
the interrupt status. This enables AON timer to
completely be quiesced before the simulation is
ready to exit. This is necessary becuase lingering
activity in the design may ocassionally lead to
assertion failures, if the simulation does not
finish at the right instant, which is not easy
to adjust.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>