[entropy_src/dv] Support complete backpressure
Testing the cs_aes_halt_req/ack port calls for improving backpressure
support in the entire entropy_src pipeline, as submitted in PR #14644.
These changes in turn actually simplify some of the timing subtleties
that have to be done in the DV scoreboard.  Some of the DV timing
hacks done prior to this commit no longer accurately predict
the timing of the entropy source.  Most notably, data samples
no longer linger in the RNG bit-select FIFO until another RNG
sample is recieved.

This commit also updates the assertion tweaks that need to be
done to support the backpressure changes.  This is because
there is another instance of prim_mubi4_sync that needs to be
relaxed when bad Mubi values are injected into the pipeline.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>