[dvsim] Install a SIGTERM handler
This commit adds a SIGTERM handler for dvsim. The kokoro CI runner
terminates the dvsim invocation if it exceeds the 24 hour runtime.
Without the SIGTERM handler, the jobs running on the cloud (GCP)
end up as zombies. With this change, dvsim will attempt to kill
the cloud jobs cleanly, generate and print the final reports before
exiting.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>