[bazel] Use ROM in SRAM program test instead of Test ROM
Prior to this commit, we relied on the startup routine in the Test ROM
that configures the PMP to enable RWX the entire SRAM region. The ROM
does not have the same "gadget" for us to reuse, so we needed to figure
out how to do the same thing with OpenOCD commands. This proved very
tricky due to implementation details of OpenTitan's debug module[^0],
but we did find a workaround.

The test is a little flaky now that we're using the ROM. I think there
are at least two sources of flakiness: (1) the ROM bootloops because it
cannot find ROM_EXT in either flash slot, and (2) the watchdog timer.

  * I think that establishing the OpenOCD connection races with the
    device rebooting, so if we're unlucky the connection will fail. If
    this is the case, we may be able to automatically retry OpenOCD
    until the connection succeeds.

  * I've addressed the timer issue by adding a sequence of GDB commands
    that disable the watchdog timer.

While trying to reduce flakiness, it's easy to succumb to confirmation
bias. To smooth out the noise a bit, I've been repeating the test 20
times with the command below. Anecdotally, I'm seeing somewhere between
0 and 7 flakes per 20 runs.

    ./bazelisk.sh test --runs_per_test=20 \
      --cache_test_results=no --test_output=streamed \
      //sw/device/examples/sram_program:sram_program_fpga_cw310_test

[^0]: https://github.com/lowRISC/opentitan/issues/14978

Fixes #13968

Signed-off-by: Dan McArdle <dmcardle@google.com>