[dv, uart_agent] Delay TX by 1ps
At the chip level, UART IOs are muxed with other IOs.
If multiple interfaces are connected to the same chip
IOs, with muxes selects selecting the function, the
IO may undergo 0->1->0 transition in the same timestep.

Since UART is an asynchronous protocol, the monitor
ends up picking up the glitch value erroneously as the
start of a new transaction. To prevent this, we delay
the incoming uart_tx by 1ps, which functionally has no
impact.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>