[chip dv] Fix errors due to use of invalid HIER macros
The *_HIER macros are now localized to chip_if. This is done to discourage
the referencing of absolute hierarchical paths in UVM code, which makes
the codebase unportable. To sample / force / release a signal in the design,
and interface must be used.

In this commit, a convenience macros is added to generate a signal probe
function in an interface. It accepts the function name and the path to the
signal as argument. The generated function takes probe action (sample /
force / release) of type signal_probe_e and the value to be forced as input.
It returns the sampled value of the signal as output (if sample action is
chosen).

The checks on LC broadcast signals are also moved from the base vseq (which used
hierarchical paths to the signals in the design with uvm_hdl_read()) to the
chip_if interface. This avoids the use of uvm_hdl_read as well as the use of
hierarchical paths in test sequences.

Tests passing due to this change:
- chip_sw_lc_ctrl_program_error
- chip_sw_alert_handler_escalation
- chip_sw_alert_handler_entropy
- chip_sw_usb_ast_clk_calib

Signed-off-by: Srikrishna Iyer <sriyer@google.com>