feat(kmac): Allow ENTROPY SEED update after REGWEN off
_Related Issue: https://github.com/lowRISC/opentitan/issues/14902 _

`ENTROPY_SEED[i]` is being used to issue LFSR seeds in the SW entropy
mode. The updating of the seeds is one-time event. After the first
update, SW is not permitted to change the seeds.

The CSRs were also blocked by `CFG_REGWEN`. If SW clears the
`CFG_REGWEN` before it updates the seeds, the SW is not able to update
the seeds again until reset.

This commit releases the restriction. It allows the `ENTROPY_SEED[i]`
can be updated afer the SW clearing of `CFG_REGWEN`.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>