[chip dv] Conditionally drive the default interfaces
The interfaces driven by default are the to the dedicated
IOs, external clocks source and the SW strap interface. All these
are removed. The external clock source interface and the
SW strap interface are driven in the base test sequence
class only if plusargs are set. The default pulls with `ios_if`
continue to remain (will be revisited in future).

The previous commit detangled the external clock source from a
legacy chip_rst_if instance by using a dedicated instance in
chip_if (called ext_clk_if). This new interface is now used to
drive the external clock source. Most tests do not need it, since
clocks are provided internally by the AST.

The ext_clk_if drove the chip pin IO6 be default, causing the
IO to get locked out of other usages. This commit fixes that
by only conditionally enabling it, via a plusarg (+chip_clk_source).
The side effect of this is - the legacy clk_rst_if which used
the pin IO6 as its (passive) clock port no longer had an active
clock to monitor, causing a widespread breakage. As it turns out,
the interface cfg.clk_rst_vif (i.e. the legacy clk_rst_if instance)
is used by several tests (including the base classes) to wait for
clocks. With an active clock no longer present, the wait_clk()
invocation ended up hanging.

To fix this, the clock port of the legacy clk_rst_if was left
unconnected, and the clock port was made active so that the
wait_clks() invocation worked again. Now the issue with that is
tests are waiting on clocks that is not even used. This is ok for
now becuase tests were waiting for posedges of the external clock
attached to IOC6, which was unused as well!

The other issue was with the XBAR UVM env which replaces the
chip env when `xbar_mode` is active. It requires the reset
to the chip to be driven using the legacy clk_rst_if instance.
This is fixed in the testbench.

This commit also fixes other interfaces that were driving the
chip IOs by default in the base sequence, including the SW strap
interface which is not condition on bootstrap being enabled (it
will also be refactored in a future commit).

In addition, this commit adds chip_if::disconnect_all_interfaces()
method as a convenience to disconnect everything from the chip
IOs in one shot.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>