[util/topgen] Update alert lpg generation
- fixes #14954

The alert lpg generation currently uses the clock group, clock domain
and power domain of a particular clock to uniquely identify lpg groups.

As a result, it lumps all the transactional clocks of the same domains
into one group.  This behavior however is not correct, as the
transactional clocks are unique and can be individually gated.

This commit updates the lpg generation to take this into account.

Signed-off-by: Timothy Chen <timothytim@google.com>