[i2c] fix how sda_interference is detected.
- fixes #15067
- instead of checking the input value immediately after changing
  sda, we instead wait through the rise time and sychronization delays.
- After waiting through that delay, if the input does not match the
  output, then assert sda interference.

In the process of making this fix, the fsm was simplified a tiny bit
on the host side.  The SetupBit states were removed since they do not
serve any function.  There is no need for the host to wait until the
setup window before driving its output.

Additioanlly, minor fixes were made to the testbench to correct some
of the ack and nak sampling behavior.

Signed-off-by: Timothy Chen <timothytim@google.com>