[otbn,dv] Fault injection test fixes
This commit changes both the model and the test sequence. Model now
waits until we finish initial secure wiping process to move into LOCKED
state, before it was missing the error injection completely in that
scenario.

Test sequence now checks if we inject a FSM fault into start stop
controller, if so we set the immediate lock bit for the model. Also
we also turn the checking of matching URND requests in the test now
because injecting FSM error in the test will definitely cause non
matching URND requests.

Signed-off-by: Canberk Topal <ctopal@lowrisc.org>