[sim_verilator] Start with the first flash bank erased
Add initialization to the first flash bank so that it appears to have an
erased state upon start-up. Subsequent loading may overwrite the state
to achieve the desired values.

Also fix up some uncontrolled behavior in rom_epmp_test:
When doing some flash checks for the rom_epmp_test, the test code relied
upon buggy behavior in the verilated model simulation. With flash now
being initialized to erased, explicitly target the erased area for
addresses and values, keeping in mind that the same image is used for
both the ROM and the flash.

Signed-off-by: Alexander Williams <awill@google.com>