fix(spid): Latch FIFOADDR at correct time
_Related Issue: https://github.com/lowRISC/opentitan/issues/15266_

This PR fixes a bug for unaligned access in TPM read commands.

fifoaddr is latched with the counter being `'h1F`. The counter, however,
runs on SPI_CLK, while the latching logic is running on inverted SPI_CLK
domain.

As a result, the latching logic captures the fifo address at a cycle
earlier with prematured address. So it always assume address[1] as
address[0].

This commit fixes the issue by:

1. Convert `sck_fifoaddr_latch` into iSCK domain.
2. To not shift the address one bit more, use `sck_cmdaddr_wdata_q`
   rather than `sck_cmdaddr_wdata_d`.

Signed-off-by: Eli Kim <eli@opentitan.org>