[top/dv] Fix rstmgr info test
Previously, this test assumed that escalation would always happen
within a fixed amount of time.  However, that is not necessarily
the case for ping timeouts.  The ping mechanism randomly selects a
peripheral to check.  However, since the selection vector is larger
than the number of peripherals we have, it does not always select a valid
peripheral. When the alert handler does not select a valid peripheral,
it simply moves on to the test the next ping. However the max wait time
until the next ping is checked is in the mS range.
Therefore, the test should not make that assumption and just wait in
place.

Signed-off-by: Timothy Chen <timothytim@google.com>