[opentitantool] Support multi-layer alias conf
We have an agreed way of connecting HyperDebug to CW310, which should be
declared in a JSON conf file, such that opentitan tool understands a
name such as A3 to be an alias of the HyperDebug native pin CN7_18.

Separately, the ChromeOS team has a particular idea of how to use the
OpenTitan chip pins, and wishes to use the name AP_FLASH_SEL as an alias
for A3.  This is mostly independent of CW310 and HyperDebug, and would
apply even to future ASIC development board.

Because of the above, I foresee that we would want to simultaneously
provide a CW310/HyperDebug configuration file, and a CromeOS/EarlGrey
configuration file to OpenTitan tool.

This PR allows the --conf option to take multiple arguments, and it
adapts the alias resolution logic to be transitive.

Change-Id: I6f8a328f269dfee246a19c4d3bae9fd60f056aec
Signed-off-by: Jes B. Klinke <jbk@chromium.org>