[entropy_src/rtl] Minor predictability fix
This commit fixes a minor timing hiccup that was previously missed when
in RNG-mode.

The precon packer FIFO should not be cleared on disable if it has a full
64-bit data word. This is because once a word is completed and presented
to the condititioner input, our scoreboarding assumes that it has been
added to the SHA internal state.  Not clearing this FIFO allows the
conditioner to absorb the word later once the module is re-enabled.

This subtlety has previously been properly applied in FW_OV mode. This
commit also applies it in RNG mode as well.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>