[chip dv] Exclude AST mem config ports
The ports `sprgf_rm_o` and `sprom_rm_o` from AST are
covered in connectivity tests. We missed excluding these
signals before.

Note that Verdi regenerated the exclusion file, so the
diff looks extensive due to design changes from the last
time the el file was saved. The change however is only
the addition of the aforementioned AST ports.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>