[fpga] Enable AES PRNG reseeding on CW310
So far, we have been skipping reseeding operations for AES on the CW310
to simplify SCA (the reseeding disturbs the actual measurement). In the
meantime we have added an alternative but safe way to skip the
reseeding in software if really desired. Thus we no longer need the
hardware modification on the CW310.

It is however still needed for the CW305 where the entropy complex
isn't implemented.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>