[dv/common] Fix common alert_test failure
In OTP_CTRL and FLASH_CTRL, there are `alert_test` failure due to
accessing `alert_test` register together.
This is because there are too many `fork join_none` inside the
alert_test.
It is not necessary to have these parallel tasks because we cannot write
two CSRs together in tlul bus. So I removed the forks and keep the for
loop instead.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>