[entropy_src/rtl] Clear window_cntr on disable
This commit reverts the window_cntr_clr behavior to the original design
where this counter is cleared on disable.   When health_test_clr was
changed to a pulse emitted on enable, this counter was left uncleared
after disable.

This has been observed to cause spurious `window_wrap_pulse`'s when
reconfiguring the module, particularly if the window_size is shrunk
while reconfiguring the DUT.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>