[rv_dm dv] Fix regression failures
- Quite a few fallouts with the DV_CHECK macros failing uncovered bugs in
the sba_access_utils_pkg.
- The stimulus in sba_access_utils_pkg and the sba_access_monitor work
together to infer indirectly SBA accesses that are expected to show up
on the TL interface. The required bug fix resulted in larger changes
that were needed in both, the stimulus and the monitor.
- The rv_dm scoreboard also required fixes because the sba_access_monitor
was not setting rdata correctly, causing read checks to vacuously
pass.
- The design was updated to support sberror=2,7 cases. The test sequence
was updated to fix that.
- Updated testplan as well to reflect testing improvements.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>