[entropy_src/rtl] Put main_sm state defs in pkg
In order to close FSM coverage it may be neccesary issue carefully timed
disable or alert commands.

This commit separates the definitions of the parse main_sm state enum
into a separate package so that the enumeration is visible to DV. This
allows DV to refer to the states by name, track future changes to this
spare FSM and generate meaningful debugging messages.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>