[bazel] Rewrite GDB test script, respect subprocess exit codes
To implement #14490, we'll want to assert that the value of PC is the
address of a particular function. GDB is in a perfect position to test
this kind of assertion. It knows the addresses of symbols, it knows the
values of registers, it supports conditional expressions, and it has a
`quit` command, which takes an exit code parameter.

When I tried inserting an unconditional `quit 123` into the existing GDB
script for sram_program_fpga_cw310_test, I was surprised that the test
did not fail! It turns out we have been silently ignoring exit codes of
background processes.

I decided to rewrite the test script in Python, if only for clarity. The
rewrite is nearly a drop-in replacement. The only intentional difference
is that it checks the exit codes of its subprocesses.

As a proof of concept, I inserted these lines into the GDB script for
sram_program_fpga_cw310_test and verified that it works as expected:

```
        if &sram_main == 0x10001fc5
          echo :::: Correct.\\n
        else
          echo :::: Surprise!\\n
          quit 123
        end
```

Signed-off-by: Dan McArdle <dmcardle@google.com>