[cleanup] Remove licence-checker.hjson.
This file is no longer used (its contents have been transferred to the
Bazel build rule `//quality:license_check`). Keeping it in the source
tree caused some recent confusion.

This change also updates a comment in the CI license-checking script
which references a file that no longer exists (found while I was looking
for any references to the removed file).

Signed-off-by: Jade Philipoom <jadep@google.com>