[spid] fix: TPM RdFIFO reset
In previous design, TPM CSb (after scan mux) directly connects to the
write port reset of TPM Read FIFO.

The issue here is that if TPM CSb is asserted, it may trigger
metastability of the write port of the FIFO as the timing of the release
is independent of the SYS_CLK (write port clock).

This commit fixes the issue by adding a reset synchronizer to the reset
path. As read FIFO is being used after 32 SPI_CLK cycles, the 2 SYS_CLK
delay on the reset path will not cause an issue.

Signed-off-by: Eli Kim <eli@opentitan.org>