[entropy_src/dv] Properly count coincident HT failures
If Windowed, Continuous and/or External HT's simultaneously fail in the same sample
the DUT's alert count only registers one failure, for the sample, which in generally
fine.  However the previous approach at counting these coincident failures
overestimated the overcounting factor and was unable to predict certain alerts.

This commit corrects the accounting for alert counts when simultaneous alerts occur.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>