[entropy_src/dv] Fine tune extht timing cases (DV)
This commit fixes DV corner cases to capture the last RNG data accepted
into the DUT before it is disabled.  Since the new EXTHT agent
introduces a slight synchronization delay this adjustment is needed to
make sure that all samples are counted.

In the worst case the window_wrap_pulse ends two cycles after the DUT
has been disabled. (One cycle if RNG bit selection is disabled two
if it is enabled) Rather than dumping the a partial HT window
immediately when the DUT is disabled, wait these extra two cycles to
catch the last sample.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>