[dv] Correct write size checker function
The RAL has one more level of reg blocks at the chip-level
than at the block level, which means that we have to figure
out which regblock a register belons to before calling
get_supports_sub_word_csr_writes(). I.e., calling it on the
top-most RAL (which is the chip RAL in this case) does not
return the property we are looking for.

Signed-off-by: Michael Schaffner <msf@google.com>