[entropy_src/dv] Synchonize ExtHT and other HT scoreboarding
Prior to this commit, the ExtHT vs. the Repcnt and windowed tests all ran
in  separate threads.  This commit makes the health test scoring driven
by ExtHT events (sample received, window closed), thus all health tests
can be scored in the same task.

ExtHT Req events are now checked to see that they match the RNG data and
that the window-close events are occuring at the correct time.

The original intent was to find a predict/access collision in the alert
count registers.   This commit does not fix that bug unfortunately. It
does however simplify the code, and remove one more thread from the
scoreboard.   It also supports better scoring of ExtHT events.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>