[jaspergold] Add build pass pattern
Ensure that the log has this string which appears at the end of the
proof. If for whatever reason, the tool failed to run, currently it
shows up as a false pass.

Note that this issue is still not yet fixed - there is one additional
fix needed (probably) in dvsim.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>