[entropy_src/dv] Functional Coverage Improvements
This commit modifies the regression parameters to deal with the recent
OTP coverpoint changes added by @andreaskurth (support for both OTP
inputs, coverage of Invalid Mubi8's as well as True and False)

In general the `entropy_src_rng` and `entropy_src_fw_ov` tests have been
halved in duration, but with 6x more tests.  The simulated time is now
roughly 3x of the original, to account for the fact that the size of the
largest cross-coverage points have grown by somewhere between 3x and
4.5x (depending on the crosspoint).

The tradeoff in performing more numerous shorter sims is important
because the OTP settings are fixed through out a run, and so any gaps or
imbalances in OTP coverage have to come from more runs.

This commit also adjusts the constraints so that (where-applicable) there are
equal numbers of all three types of MuBi values.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>