Revert "[csrng] Temporarily hack-in disable handling"
The reverted commit was originally introduced as part of
lowRISC/OpenTitan#14410 to work around a temporary issue until
lowRISC/OpenTitan#14534 got merged. It seems we forgot to remove that
temporary workaround before merging the PR.

The workaround is critical because it weakens a security countermeasure
inside CSRNG: errors injected into the command stage FSMs are no longer
fatal. By disabling and re-enabling the module these FSMs can be made to
leave the "terminal" error state.

For this reason, we need to remove this workaround. This reverts commit
6f75a7219215d675d22140224363414d3f26d988.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>