[doc] Specify enabling and disabling of entropy complex
This formalizes the current recommendation that the components of the
entropy complex should be enabled in the order entropy_src, then CSRNG,
and then EDNs into strict rules.  It additionally adds rules that the
components of the entropy complex must be disabled in the reverse order,
and that all modules must be disabled before they can be re-enabled
again.

These rules prevent inconsistent states between the modules of the
entropy complex, which may otherwise occur (see #15789 or #16355 for
examples).

Signed-off-by: Andreas Kurth <adk@lowrisc.org>