[ci] Add checks for empty test_suites (and more!)
This commit adds scripts that check for empty test_suite targets and
targets that contain "banned" characters. It also re-homes the check for
test_suite targets that are not tagged "manual".

I'm adding these scripts after making a few mistakes while cleaning up a
BUILD file in PR #16401.

My first mistake was forgetting to unpack a tuple. I wound up inserting
the tuple's string representation, rather than one of the tuple's
elements, into a target name via `string.format()`. Surprisingly, this
is not an analysis-time failure! Bazel has no problem inserting
parentheses and commas into the target name.

If weird characters are in a target name, for some definition of
"weird", it was probably an accident. The new script returns an error
when it finds targets with weird characters.

The second mistake I made was defining a test_suite with a non-empty
`tests` parameter, where none of the labels actually existed. The result
is a test_suite that appears non-empty, but actually contains zero
tests.

I believe an empty test_suite indicates an error. The new script returns
an error if it finds any test_suite targets that contain zero tests.

Signed-off-by: Dan McArdle <dmcardle@google.com>