[chip, dv] Update sysrst_ctrl CSR exclusion to fix a csr test failure
In chip-level, even if we keep all sysrst_ctrl pin all 0s, randomly writing
CSRs may still trigger `key_intr_status`.

In block-level, we didn't see this issue, as core clk is really fast compared to
the aon_clk, and this intr depends on aon_clk. In chip-level, the CSR tests take
ver long, so that we can occasionally see this failure.

Signed-off-by: Weicai Yang <weicai@google.com>