[top_earlgrey/dv] Fix waiting for rom_ctrl after reset
The previous method of waiting for `CFG_REGWEN` of KMAC no longer worked
reliably, because it's currently possible that the simulation is
hundreds of cycles before or after the deassertion of reset when the
`wait_rom_check_done` function is called, so the transition from 0 to 1
of `CFG_REGWEN` can no longer be reliably detected with CSR spinwaits.

Instead, we now wait for digest 0 in the rom_ctrl registers to become
non-zero, which happens only after the post-reset rom_ctrl checks have
completed.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>