[dv/cov] Exclude coverage of dv-only code
There is a number of places in the RTL where some code is created
exclusively to support SVA, and the code is not synthesized.
This code needs to be marked to avoid coverage collection so we
don't try to increase its coverage.
This PR handles code under `ifdef INC_ASSERT.

Partially addresses #15423

Signed-off-by: Guillermo Maturana <maturana@google.com>