[opentitantool] SPI port name aliases
We currently have `ott bootstrap` and `ott spi ...` commands which use a
SPI port to communicate with a bootlader.  In the future, there will be
`ott spi tpm` subcommand to communicate with code running on OpenTitna
via SPI using the TPM protocol.  For these different uses, we want to
default to using different SPI ports.  (Dauntless uses the same SPI pins
for both, but OpenTitan uses the dedicated QSPI lines for bootstrapping,
but pinmux'ed lines for the TPM SPI communication.)

In order to support such various setups, this PR introduces SPI port
alias_of, the same way as exists for UART ports.  In particular, rather
than referring to spi instance "0", all the existing commands are
modified to default to a SPI instance called "BOOTSTRAP", the future
`ott spi tpm` command will default to a SPI instance called "TPM".

Various json configuration files will declare "BOOTSTRAP" and "TPM" to
be aliases of the particular transport instances, for HyperDebug with
CW310 for instance "BOOTSTRAP" will map to the new "QSPI" port, and
"TPM" will map to the existing "SPI2" port.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I86b1873add01f5e935f0b5d584744844f5aad081