[bazel,doc,util] Update @bitstreams repository with Git hook
This PR updates the Bazel rule that implements the `@bitstreams`
repository. It sets `configure = True` to indicate that the rule depends
on something outside Bazel's dependency graph. It also removes `local =
True` because the rule actually does hit the network. I think this was
set to true originally as a hack to force re-evaluation of the repo.

Additionally, this PR adds a Git post-checkout hook that invokes
`./bazelisk.sh sync --configure` and updates the FPGA docs with
instructions for installing the hook.

This PR is an alternative to PR #16863, which worked by dirtying the
tree. This PR does not generate files in the source tree.

Fixes https://github.com/lowRISC/opentitan/issues/16832

Signed-off-by: Dan McArdle <dmcardle@google.com>