[dv/alert_test] Fix alert_test regression failure
PR #16896 uses agent to automatically drive alert response sequences for
each IP.
But alert_test needs to disable auto response. The current design
disabled at pre_start stage, but it is too late.
The auto-response sequence runs from the beginning of run_phase.
So this PR moves the disablement to the build_phase.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>