[top/dv] Fix handling of unpacked dimensions in Ibex lockstep test
The unpacked dimension of glitched I/O ports are defined by parameters
in ibex_pkg.sv and thus directly available to the test. Thus, there is
no need to probe the design to retrieve these values. While the probing
works for VCS it doesn't for Xcelium where the package parameters don't
show up as module parameter in the hierarchy.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>