[top/dv] Fix back propagation of non-lockstep inputs in lockstep test
The input ports of the ibex_core module are defined as `logic` without
an explicit net type. According to the standard, simulation tools are
thus supposed to model these inputs using a `var` type. However, it
turns out that some tools collapse input ports into a single object to
reduce the number of assignments to improve simulation performance. As a
result, glitches inserted to inputs of the non-lockstep core may
propagate back and also change the input of the lockstep core. If this
happens, both cores are glitched simultaneously without any alerts
firing.

To avoid this, we also glitch the corresponding input of the
ibex_lockstep instance to the opposite value. The ibex_lockstep
instance is embedded inside prim_buf cells across which glitches don't
progagate back. Also, the delay lines are embedded inside the
ibex_lockstep instance. It's thus fine to apply the glitch
simultaneously.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>