[test] move KMAC customization string load in power virus test
This moves the loading of the KMAC customization string to the
`crypto_data_load_task`, which happens later in the test. This was done
to facilitate debugging the IP configuration portion of the test in DV sims,
since an assertion fires if a KMAC operation is started but never
completed.

Signed-off-by: Timothy Trippel <ttrippel@google.com>