[entropy_src] Don't push to input FIFO unless indicating ready to AST
In lowRISC/OpenTitan#16732 we started registering the rng_enable output
to AST to make the signal glitch free for closing CDC. As a result of
the introduced register, it became now possible for entropy_src to
push entropy from AST into the input FIFO despite indicating not being
ready. While this might not have been a big issue for the actual chip,
it led to large amount of failures in block-level DV due to the
push-pull agent on the AST interface, the scoreboard and the RTL
becoming out of sync whenever re-enabling entropy_src while AST provided
valid input data.

This can be fixed by only pushing valid data from AST into the input
FIFO if entropy_src is indeed indicating ready to AST. The behavior
during disabling of entropy_src isn't affected by this fix.

This resolves lowRISC/OpenTitan#17236.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>