[opentitantool]: Proxy forwarding of strappings
When using `opentitansession`, often the session process will be
provided with configuration files on startup, and the individual
`opentitantool` invocation may or may not get any configuration file
arguments.

To handle all cases, `gpio apply` should check if either the session
process knows a strapping by the name, and in that case have the session
process apply it, or check if local files declare a strapping by the
name, and in that case instruct the session process to individually
configure every pin mentioned in the strapping.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>

Change-Id: Ieb80862234b8c1fb879db15778e6c69ff397f1c2