[sysrst_ctrl,dv] Combo detect key combinations cross consolidation
Instead of requiring each key combination to happen in each of the four
combo blocks, I create a new covergroup that aggregates the key
combinations we see in all the combo blocks into one. This is ok because
all of the combo blocks are essentially the same and it simplifies the
verification effort. I also added this new sample to the combo detection
test.

Signed-off-by: Marno van der Maas <mvdmaas+git@lowrisc.org>