[dif] Add DIFs to acknowledge a snapshot of IRQs
In issue #17349, alees24 identified a deficiency in the available
functions for acknowledging IRQs: The choices were a single IRQ or all
of them.

Add a set of DIFs for acknowledging the precise set of IRQs that were
retrieved at the time of calling dif_<IP>_irq_get_state() so multiple
IRQs may be acknowledged at once, but IRQs that were not observed do not
get inadvertently cleared.

Signed-off-by: Alexander Williams <awill@opentitan.org>