[crypto] Simplify DMEM interface for RSA.
Now that we have a .bss directive for OTBN, there's no need to rely on
indirect pointers or specific memory locations in RSA code. This also
will make it easier to use the same underlying libraries for keygen and
signing.

Also includes a minor improvement to the OTBN testutils, so that we
print the error bits if OTBN fails a test because it became locked.

Signed-off-by: Jade Philipoom <jadep@google.com>