[rom/e2e] refactor key handling in the BUILD file
Many key references are hard coded which is not great for future
work on the keys. This commits refactors the way keys are handled
in the BUILD file so that all changes can be done in one place.
It also documents/rename some of the variables which purpose was
not clear.
This commit does not change any generated rule or tests.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>