[opentitantool] add a command to export a public key to header file
The "show" command can be used to display a key but the output formats
supported are only json, json5, etc. It does not make sense to add
an output format just for the ROM since it is so specific to this
particular command. Instead I created another command that output
the header that can be used in the ROM/ROM_EXT.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>