[flash_ctrl] update `IPoly` parameter in flash scrambler
While working on #9322, I noticed the GF(2) irreducible polynomial used
for the flash scrambling scheme was not the Conway polynomial with
degree 64, but rather the Conway polynomial with degree 32.

After discussion in #17443, it was determined that while OK, as the
polynomial was still irreducible, a better, NIST recommended polynomial
should be used instead.

This updates the GF(2) irreducible polynomial used by the flash
scrambler to the NIST recommended polynomial for block ciphers:
`x^64 + x^4 + x^3 + x + 1`.

Signed-off-by: Timothy Trippel <ttrippel@google.com>