[util,bazel] integrate flash scrambling with bazel
This integrates the flash scrambling script with Bazel to enable
pre-scrambling flash images before they are backdoor loaded in DV and
Verilator simulations.

A key feature to enable such is to pass the OTP VMEM image to the flash
scrambling script so the flash scrambling key seeds, and enablement
flag, can be read out and decoded for use.

While this commit reads out the scrambling key seeds, they still must be
processed to produce the actual scrambling keys (as is done in HW).
Additionally, the scrambling enablement flag must be read to determine
whether or not to enabling scrambling within the pre-preprocessing
script. These will both happen in a follow up.

Signed-off-by: Timothy Trippel <ttrippel@google.com>