[dif] Do not program a dummy segment of length zero
When issuing dummy segments to the spi_host, avoid a segemnt of length
zero.  Programming a zero to the hardware results in a 512-bit long
dummy segment.

For convenience, it is simpler to be able to prepare a SPI transaction
list with a zero-length dummy segment and have the DIF skip such
segments.

Signed-off-by: Chris Frantz <cfrantz@google.com>