[opentitantool] Only do UART pacing when SW flow control enabled
In order for SW flow control to work, we cannot allow very many
characters to accumulate in buffers/FIFOs of the operating system or
hardware debuggers.  For that reason, the current code "paces" its
output, to ensure that there cannot be any built-up.

The overhead of such pacing may reduce the throughput, depending on
timer precision and system load.  The effect is large enough that the
Ti50 serial "rescue" bootstrapping takes 5 minutes instead of around 1
minute, which causes timeouts in automated tests.

This PR disables the pacing of the output, of software flow control is
disabled.  There can be no use cases for pacing without flow control, as
it could allow continuous output of characters at maximum speed,
assuming optimal timer behavior.  It cannot be "worse" than that without
pacing.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I224775a95c30abe0c2b5ff9ccbf72846199a11f6