[opentitantool] Improvements to host emulation GPIO
Allow setting GPIOs while DUT is off.  This is achieved by having the
"host emulation" transport backend keep track of how the simulated
environment drives each GPIO signal.  If the emulator is off, no attempt
will be made to immediately notify through a pipe.  When a sub-process
is later started, it will receive a dump of the initial state in a file
`he_init.json`, and from then on, updates to pins will be sent through
the usual pipes.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>

Change-Id: I5e1ad987e1357cbba9685b2bc5dd558aa458feba