[ottf] rename `ottf_flow_control` to `ottf_console` library
This renames the `ottf_flow_control` library to the `ottf_console`
library in preparation for moving all OTTF console related funtions to
the `ottf_console` library. This library will also grow in the near
future as we allow configuring different devices to act as the OTTF
console (besides just UART0).

Signed-off-by: Timothy Trippel <ttrippel@google.com>