[doc] Split dependency installation task in half
I got a bit confused when trying to follow this at first: I installed
the stuff through apt, but didn't notice that there was also a list of
Python dependencies.

Split them apart. This might be a bit clearer, but it also means that
the documentation build guide that points in can point at both
sections explicitly.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>