[opentitantool] Improved emulator startup
Rather than having the host emulation sub-process connect to a socket of
the parent process, as means of confirming successful startup, instead
rely on standard output from sub-process.

This has the advantage that if the sub-process crashes, and will never
acknowledge, no timeout is needed, as unlike with listening on a socket,
the parent process cannot avoid noticing the standard output pipe of the
child process being closed.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I97089e4c879867eb9515a74333cb47a448dd8517