[base/print] add framing protocol to spi_device console
Due to pitfalls in the hyperdebug SPI generic mode hw/firmware, a read
transaction must be performed before a write transaction, where all
communication is half-duplex. This means that naturally, some bytes will
be lost during transmission.

To account for this, and still enable the use of the spi_device in
generic mode for the OTTF console, a simple framing protocol was
developed to account for known pitfalls.

This partially addresses #17494.

Signed-off-by: Tim Trippel <ttrippel@google.com>