[bazel] Disable ccache under Bazel
The combination doesn't really work, mainly because ccache is trying
to write to a cache file on a read-only file system. This isn't
unreasonable though: the two programs are trying to do the same job!

Passing a "don't actually cache anything" flag like this to ccache
seems to be the easiest way to tell things to work under Bazel without
requiring the user to uninstall ccache from their computer.

Note that efdc97d58 is also to do with ccache interaction. But that's
specifically for OpenOCD. I *think* that this change probably means
that the efdc97d58 won't be necessary any more, but I haven't actually
checked.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>