[entropy_src] Don't predict counter clearing in FW_OV mode, align doc
Previously, the scoreboard was assuming that even in FW_OV mode (extract
and insert) the health test failure counters would be cleared upon full
passing test sequences. But this is not right. In FW_OV mode those
counters are not cleared as firmware may want to observe those values
when working on the extracted entropy bits (during that time further
entropy might be collected which is discarded and should not lead to
a clearing of the failure counters).

This hasn't been explicitly documented yet but the main FSM
(entropy_src_main_sm.sv) contains a comment on that:

"Even when stalled we keep monitoring for alerts and maintaining alert
 statistics. However, we don't signal alerts or clear HT stats in FW_OV
 mode."

Therefore, this commit aligns DV and doc with the RTL implementation.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>