[top_earlgrey/dv] Remove `chip_sw_entropy_src_cs_aes_halt` testpoint
Due to the current implementation of `cs_aes_halt` in Entropy Source, a
test to cover this testpoint cannot be implemented.  Details and
discussion are in #14193.  DV to verify the current implementation at
the block level is being added in PR #17943 for Entropy Source and in
PR #17950 for CSRNG.

This closes #14193.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>