[dif/pinmux] add spin wait to `*_pad_write_attrs()`
When enabling/disabling pull up/down on a pad, it takes some time for
the action to propagate to physical pads. This updates the pad attribute
configuration DIF to check if the pull up/down enablement is toggled,
and wait 5us to allow time for the change to propagate to the physical
pads.

This partially addresses #17862.

Signed-off-by: Tim Trippel <ttrippel@google.com>