Update clk_Div4 contraints with Xilinx recommended settings
Xilinx found only 1 or 3 Div4 clock trees were constrained correctly in
the original design constraints, and recommended new clock constraint in
clocks.xdc which hits all three branches of the Div4 clock.

1. Added Div4 clock group constraints recommended by Xilinx
2. Div4 clock group set async

NOTE regarding change 2: Div4 clock is asynchronous to other clocks as
designed, and the change is required for vivado to be able to build
the design without timing errors.

Signed-off-by: Cindy Liu <hcindyl@google.com>