[bazel] add custom rule to gen raw unlock token rust src
This adds a custom rule to generate a rust source file with the RAW
unlock token. We use a custom rule over a genrule becase we need to rely
on a `int_flag` for the `lc_seed`, so that it can be overridden on the
cmd line during dvsim runs.

Signed-off-by: Tim Trippel <ttrippel@google.com>