[test] Gate SPI upload processing on IRQ, not FIFO
Change the spi_passthrough_test software to gate command upload
processing on whether the cmdfifo_not_empty IRQ is pending.

Previously, the test used a nonempty command FIFO as the start signal,
but spi_device commits command FIFO entries and reports increased depth
before the transaction completes. Meanwhile, the status bits do not
update until the transaction is done. This would cause the test to
spuriously fail for uploaded commands with longer payloads, as the test
software checks that the WIP bit is high when an uploaded command is
received.

Signed-off-by: Alexander Williams <awill@opentitan.org>