[sw/silicon_creator] Add two optional prefixes to spx_verify()
Anti-rollback and usage constraints checks are currently integrated into
signature verification as an FI countermeasure. This requires storing a
portion of the verified message in the main SRAM.  This commit adds two
optional prefixes so that part of the message can live in the main SRAM
as opposed to the embedded flash.

Signed-off-by: Alphan Ulusoy <alphan@opentitan.org>