[sw/silicon_creator] Add structured kChipInfo to ROM
This commit adds a new `chip_info_t` struct that contains info about the
ROM's provenance. The ROM now prints the Git commit hash on shutdown
after the "VER:" prefix.

The ROM linker script was already configured to place the .chip_info
section at the top of ROM, but I'm not confident that it worked as
expected. The autogenerated header contained a static constant, so I
think a copy would be inlined at some arbitrary location into any
compilation unit that used it, rather than being placed by the linker
script.

To verify that the ROM prints the Git commit hash on shutdown:

    ./bazelisk.sh test --test_output=streamed \
      //sw/device/silicon_creator/rom/e2e:shutdown_output_dev_fpga_cw310_rom

To verify that the .chip_info section contains data from chip_info.o:

    ./bazelisk.sh build-then 'less %s' --config riscv32 \
      //sw/device/silicon_creator/rom:rom_with_fake_keys_fpga_cw310_map

To see the value of `kChipInfo` at the end of ROM:

    ./bazelisk.sh build-then 'xxd %s | less' --config riscv32 \
      //sw/device/silicon_creator/rom:rom_with_fake_keys_fpga_cw310_bin

For example, here's the end of the ROM from the previous command. At
0x7f80, there are 64 bits of a Git commit hash (e20871e1cf102bf6).

    00007f70: 0000 0000 0000 0000 0000 0000 0000 0000  ................
    00007f80: f62b 10cf e171 08e2                      .+...q..

Issue #14892
Signed-off-by: Dan McArdle <dmcardle@opentitan.org>