[dv] Make SW strap thread more flexible
The SW strap setting thread previously assumed the progression of all
tests would proceed to being "InTest" before any need for bootstrapping
on a subsequent boot. However, this is not true for all tests.

Change the thread to evaluate any time the SW status changes, and
instead of requiring a particular set of paths to the "InBootRom" state,
set up the straps regardless of how we reach the boot ROM.

Continue disconnecting the straps only upon reaching "InTest".

Signed-off-by: Alexander Williams <awill@opentitan.org>