[test] Remove incorrect wakeup check in chip_power_sleep_load
The prepare_to_exit() function contained a superfluous and invalid check
for wakeup reasons. Under all conditions in which it is called, the
aon_timer's wakeup is expected to be active. However, prepare_to_exit()
asserted that no wakeups should have happened. This was incorrect.

Meanwhile, the correct check is already done in the test_main()
function.

Signed-off-by: Alexander Williams <awill@opentitan.org>