[manuf,opentitanlib] add option to clear bitstream when loading one
On the FPGA, OTP is modeled using memories and are initialized from portions
of the bitstream. The current Bazel/opentitanlib test infrastructure checks
if the ROM has changed by reading an FPGA-only CSR, and triggers a bitstream
load if it has. However, this mechanism does not take into account OTP
changes that could impact test behavior.

Since the provisioning_functest modifies OTP and is designed to be
reentrant, test behavior will change between consecutive runs if we do
not first clear the bitstream. This adds a option to the test_utils
`init` function to clear the bitstream if reqested in the test commands.

Signed-off-by: Tim Trippel <ttrippel@google.com>