[lc_ctrl] SVA fix in kmac interface
This removes the first prim_sync_reqack_data instance
and just directly connects the token data bus to the
kmac domain.

This is done to fix a simulation issue with CDC instrumentation
enabled. I.e., due to the use of two prim_sync_reqack_data
instances, it can happen that the handshake has already
completed on the first one, but not on the second one.

This causes these two primitives to not operate in tandem anymore,
and can trip the SyncReqAckHoldReq assertion in the first primitive.

We can safely remove the first instance, since that instance does
not contain a data register (the data is guaranteed to remain stable
throughout the kmac operation).

Signed-off-by: Michael Schaffner <msf@opentitan.org>