[otbn, sca] Keep shares in distant registers
In FPGA implementation of the earlgrey, blanking is not used
for Bignum RF due to high resource consumption and this causes
leakage with FPGA experiments. This is an intermediate solution.
We are keeping shares in registers with higher Hamming distance
between addresses so that we are reducing the probability of
reading out both shares when address switches.

Co-authored-by: Jade Philipoom <jadep@google.com>
Signed-off-by: Abdullah Varici <abdullah.varici@lowrisc.org>