[lc_ctrl] Add raw unlock bypass for test chips
This adds a volatile RAW -> TEST_UNLOCKED0 mechanism for test chips.

This solution does NOT use KMAC token hashing, hence the RTL diff is smaller.
An alternative implementation with KMAC hashing is drafted in #18238 for
reference. This has been discussed with @cdgori and we decided not to
pursue that since the RTL diff is too big and would add significant risk.
The direct comparison without hashing is acceptable since the mechanism
will not be compiled into PROD silicon.

Note that the modifications in pinmux are needed so that the hardware
straps can be re-sampled in order to ungate the DFT and RISC-V TAPs.

Addresses #18246

Signed-off-by: Michael Schaffner <msf@opentitan.org>