Add HARDENED_TRAP to avoid coalesced unimp instructions
This is an alternative to HARDENED_UNREACHABLE. It is meant to address the
issue of the unimp trap instructions being lumped at the end of functions.
That behavior is caused by an optimization of the branch folder compiler pass,
made possible by the __builtin_unreachable used in HARDENED_UNREACHABLE.
By avoiding that builtin the basic block containing the unimp instructions
will have successors, so the optimization won't apply.

Signed-off-by: Luís Marques <luismarques@lowrisc.org>