[manuf] use OTP with ROM exec disabled in `manuf_cp_yield_test`
The `manuf_cp_yield_test` was not loading an OTP image with ROM
execution disabled, nor bootstrapping anything into flash. As a result
there was a race condition between when the JTAG operations would
complete and when ROM would start resetting the chip due to a failed
boot operation.

This updates the `manuf_cp_yield_test` to use an OTP image with ROM
execution disabled to remove this race condition. This fixes #18403.

Signed-off-by: Tim Trippel <ttrippel@google.com>