[crypto] remove all-one vector from SHA-512 implementations
In both SHA-512 implementations we use a wreg cointaining only
ones in the shifi operations. The value of these bits
does not matter since they are all masked out eventually.
Since the all-one vector is not used anymore at other places
in these algorithm we can remove it's generation and just
use the all-zero vector as a source for additional bits in the
shift operations.
This affects the clobbering of unused bits in the wregs
containing the result. Therefore, the expected values are
updated to match this.

Signed-off-by: Felix Miller <felix.miller@lowrisc.org>