[site] Fixup pagetoc nav issues
- The initial PR assumed a single H1 element on the page (the first heading),
and if another H1 was encountered within the page, no further elements would be
added to the ToC. This is fixed by including the first H1 element within the
ToC, and starting with a first wrap-element wrap-W0.

- Adding the first H1 element to the pagetoc then caused a minor issue where the
scroll-margin-top failed to apply correctly when nav'ing to it, causing it to be
placed behind the mdbook header. Add some js to force-scroll into view if this
element is selected.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>