[verilator,opentitanlib] Add support for multiple flash banks
Add the second flash bank to the list of memories for the Earl Grey
Verilator-based simulator, and rename the memories to include a bank
suffix.

Adjust opentitantool's verilator transport to the new name. Also add an
ability to specify which bank with the same colon-delimited suffix
notation as DV (except where the index merely represents the bank number
instead of a larger enum shared by all memories).

In the future, we may want to generalize this further to merely pass
through generic meminit lines instead of having opentitantool have
rigid, named memory arguments. Not all integrations will have an
internal flash memory node.

Signed-off-by: Alexander Williams <awill@opentitan.org>