[spi_host,dv] Refactor status_stall_vseq and scb, add more comments
This commit is cleanups and refactoring to make the test and scoreboard more
linear. It also adds a number of comments as I went through learning how it
works, and refactors names to be more meaningful.

The status_stall_vseq test has been a source of failures for a long time. This
commit refactors it in an attempt to make it easier to debug, and tries to
address some of the outstanding issues.
Problems experienced by this test have been 1) timeouts, due to requiring lots of
stimulus to be generated to fill the RXFIFO and induce a rxstall condition, and
2) the scoreboard not checking all of the expected/observed transaction items
due to this test exercising a different codepath in the scoreboard by using
cfg.tx_stall_check.

- Increase timeout for spi_host_status_stall test
- Schedule a smoketest sequence after the main test to check recovery, and make
  use of read_rx_fifo() after this to ensure the scb is up-to-date.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>