Remove stale comment in pwrmgr.hjson
This file is supposed to be autogenerated from pwrmgr.hjson.tpl, but
it appears we don't always do that! Indeed, the comment went away from
the template file two years ago, in
5039def571371aad816273522ce985f61a2c77b5.

There's clearly some scope for clean-ups (enforce autogeneration, or
do away with it altogether!), but let's just get this line in sync for
now.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>