[release] Increase timeout of CI job
In f55867086b046eac4f282a69d236081ef7f66e4d (PR #17216), we increased
the timeout of the CI check that builds the CW310 bitstream from 3 h to
4 h due to timeouts (see #17187).  As part of the nightly Release job,
we also build a CW310 bitstream, but we have not increased its timeout
yet.  This is likely the reason that Release jobs have been timing out
for the last month (or so -- as far as traceable with the logs that have
been retained).

This commit increases the timeout of the Release job by 60 min, which
should fix its timeout issues.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>