[manuf] Add functest to verify CRC verification
On the device side, the test only prints a message.
On the host side, the program first loads the SRAM program unmodified
and makes sure it executes correctly and prints the expected message.
It then modifies one byte of the program and re-executes it to make
sure that reports a CRC mismatch error.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>