[usbdev] Remove data toggle clear in suspend
The link_active_i term was misplaced, and data toggles should not be
cleared in suspended state. In fact, they should be retained, and usbdev
will need to give software greater visibility and control, so it can
save / restore state in coordination with deep sleep.

For the purposes of an ECO, leave link_active_i unused. It was intended
to help prevent the packet engines from responding to packets and
dirtying FIFOs before software is ready, but that is left for a later
commit.

Signed-off-by: Alexander Williams <awill@opentitan.org>
Co-authored-by: Adrian Lees <a.lees@lowrisc.org>