[edn/dv] Enable Boot mode and Auto mode for `edn_err` test
Prior to this commit, the `edn_err` test was only run while EDN was in
SW mode -- that is, *not* while EDN was in Auto mode or in Boot mode.
Consequentially, this test had no chance of triggering FSM transitions
from the Auto or Boot states into the Error state.

This commit:
- Creates `edn_err_test`, which inherits from `edn_intr_test` and has
  probabilities of 40% for Auto mode, 40% for Boot mode, and 20% for SW
  mode.
- Modifies `edn_err_vseq` to write the commands necessary for Auto mode
  or Boot mode (if either of them is enabled) to EDN's registers.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>