[edn/dv] Reset `additional_data` in `edn_init`
The `wr_cmd` task uses the `additional_data` counter variable to select
between command headers (with `acmd`, `clen`, etc.) and additional data
when writing CSRs.  If `additional_data` does not get reset when EDN is
re-initialized, a value from a previous call to `wr_cmd` can lead to the
writing of additional data instead of a command header.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>