[otbn,dv] Remove unnecessary TODO comments from bad_deep_loop.py
The second comment was probably a reminder to me when I was originally
writing the code. No, it can't fail, so we can just drop the comment.

The first comment is more interesting. The reason for it was that the
generated binary will look rather synthetic, because the nested LOOP*
instructions will all occur back-to-back in the trace.

We know that OTBN is supposed to explode even if we do some other
instructions in between, so the TODO comment was suggesting that we
could think harder and interleave some. If we're really worried about
this, an easier way to get the same result would just be to branch
from the middle of a loop and land on the start of the loop again.
Something like this:

  foo:
    nop
      loopi 10, 2
        addi x2, x2, 1
        beq x2, x2, foo

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>