[aes, dv] Comment and fix usage of status_fsm() task inside send_msg()
Previously, the motivation for choosing the return_to_idle argument was
not fully clear and the status_fsm() task always returned after the
first status read for invalid mode configurations (which are resolved
to AES_NONE). This commit better documents the usage of the status_fsm()
task inside send_msg() and fixes its handling of invalid mode configs.
More precisely, also if the mode field is set to AES_NONE, DV needs to
wait with configuring IV and initial key registers until the DUT is
idle. Otherwise, send_msg() may not de-assert the new_msg variable and
repeatedly try to provide the IV and initial key to the DUT which is
wrong.

This is required to enable configuration error testing with sideload
keys.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>