[tools] Use bazel-provided Python for fusesoc subprocesses
Add a py_binary shim that modifies the PATH environment variable to hold
the bazel-provided Python interpreter directory first, so /usr/bin/env
will point to it for the shebang method of running Python scripts. Add
the entire set of packages from ot_python_deps as deps for that shim, so
they're available to any generator fusesoc might call.

Signed-off-by: Alexander Williams <awill@opentitan.org>