Simplify paths in topgen's strong_random
Several tests in this file work in the same way: iterate over a list
of things and check that a property holds for each element. The
existing code collected up a list of booleans and then checked that
each was True. With this rewrite, we early-exit with False if anything
fails, and return True otherwise.

This should be a minor clean-up, following suggestions of @vrozic when
I fixed up some types in this code in ba9317a.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>