[otbn,dv] Wait more carefully for secure wipes in vseqs
The time taken for a secure wipe isn't constant, because it involves
waiting for randomness from the EDN. And this takes longer when CDC is
enabled, causing some tests to fail.

This change adds a "wait_secure_wipe()" task to otbn_base_vseq, which
makes sure to wait long enough for the secure wipe to finish. We can
probably replace some other waits with this, but I didn't want to
rewrite lots in this commit, so we just switch to it in
otbn_rf_base_intg_err_vseq (the test that was failing for me when I
wrote this).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>