[otbn,dv] Wait long enough for RF read in otbn_rf_base_intg_err_vseq
This vseq wants to start an OTBN run and then, at some time after the
run has started, corrupt the next read of the base register file. Of
course, this needs a bit of a wait: we need to wait until the base
register file is being read. This wait is usually very short (we read
from the RF with most instructions) but is rather longer if we happen
to be stalled reading from RND!

The stall for a RND read is longer when CDC is enabled, and pushed us
over the 2000 cycle max wait. Increasing by another multiple of 10
seems to fix things.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>