[alert_handler,dv] Tweak clocking block used in wait_alert_init_done
This task expects to see the alert_tx differential signal go high and
then low again. This is reasonable enough, but that signal is actually
driven in alert_receiver_driver.sv, where it's synchronised with the
async clock.

In alert_monitor.sv, we were reading it in a clocking block that is
synchronised with the main (possibly slower) clock. It turns out that
some seeds cause this short pulse of "on; then off" to be missed,
causing the test using alerts to get stuck while it waits for them to
initialise.

This commit switches things so that we wait for the signal without
using a clocking block at all. We then wait for a posedge on
the (main) clock so that everything stays with the same timing.

I've also shuffled the comments around, which I think makes the code
slightly easier to understand.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>