[crypto] Update the KMAC driver in preparation for aligned buffers.
The KMAC driver previously used the crypto_uint8_buf_t types internally,
which is different from the other drivers and created complications when
I tried to align more things within the cryptolib; this commit changes
it to match the other drivers and use separate buffer/length parameters.

Also, the driver previously used byte-reads internally to get the
digest. Since I am about to align the digest buffers, there are
performance and SCA benefits to swtiching to word-reads. Since we now
have word-wise KMAC code from the ROM driver, I updated the cryptolib
driver to match (with a memcpy that can be deleted if the output buffer
is aligned later).

Signed-off-by: Jade Philipoom <jadep@google.com>