[dv,top-level] Fix flakiness of chip_sw_data_integrity_escalation
Perform a single increment of flash counters, so the chances of attempting a new one
before the previous one is pending are restricted to separate interrupt or exception
types, which are much easier to control and seem to work very reliably.

Notice this test still has a potential flakiness related to failing error injections:
see #19268 for a proposed fix.

Fixes #18489

Signed-off-by: Guillermo Maturana <maturana@opentitan.org>