[bazel] Rework module ID check rules to make the output more readable
The way the current rule works, opentitantool is run on every elf files
and since it's a "build" step, the output will logged and printed by
bazel. This is not ideal since opentitantool likes to print messages
like a missing configuration files, which means that bazel will print
at least two lines *per test*. In case there is an actual error, the
output will be lost in a sea of useless messages. This commits reworks
the aspect rule to run a bash script that logs the opentitantool output
to a file. This output is not printed unless the module ID check fails.
This makes the output considerably smaller and will make CI output a lot
more useful.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>