[bazel,manuf] Integrate platform transition in opentitan_binary
At the moment, opentitan_binary produces a "<name>.elf" target
with a platform constraint but it uses cc_binary on which we
cannot define a transition. This means that any target that
depends on such an ELF file will not trigger a transition
and be built with the same platform (unless this rule defines
a transition on its attributes). This is problematic for tests
that needs to compile a tool with the host toolchain but build
some artefact with the platform toolchain, hence the use of
platform_target. This is not ideal though because the situation
protentially repeats with every binary built by opentitan_binary
so it makes more sense for this macro to always produce a target
that triggers a transition. This commit specifies and implements
that opentitan_binary now create an extra target named "<name>_elf"
(underscore instead of .elf) that triggers a transition.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>