[opentitantool] Improve UART on ti50 emulation
Host emulation works by compiling the GSC code into a Linux executable,
instead of an image for OpenTitan, and running this executable as a
sub-process to `opentitansession`.

The sub-process will create a number of unix pipes, for UARTs and
other peripherals, and will use these to communicate with the simulation
environment.

Prior to this CL, `opentitantool console` would not capture the first
serial console output, because of a race.  The sub-process would create
the unix pipes, tell opentitansession that was is ready, and then
immediately proceed to produce output to the unix pipe representing the
serial console.  This allowed opentitansession no time to connect to the
pipe, and any data sent before a reading process has connected will be
lost.

This CL introduces a handshake, such that the sub-process is instructed
to wait after having indicated that it is done creating Unix pipes.
Also, opentitansession will now connect all UART pipes on sub-process
startup, rather than only at first read operation.  Together, these
changes allow the emulator to guarantee that the `opentitantool console`
will show the very first serial console output as the emulated chip
comes out of reset.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I303ffe7f4f0ae39c1f6e26ba06a540b5898e063b