[dv/clkmgr] Increase wait cycles in clkmgr_lost_calib_ctrl_en_sva_if
When the clocks loose calibration (flagged by the AST) the cycle counters in clkmgr
get disabled. The delay between enabling counters and disabling them due to loss
of calibration needs to go through a prim_reg_cdc on the path to enable it and
another one to make the enable value visible to software. The previous number of
wait cycles still left room for some failures.

Signed-off-by: Guillermo Maturana <maturana@opentitan.org>