[opentitantool] Avoid required --help argument
A few subcommands used a --hexdata option with the short form -h,
conflicting with standard -h for --help.  The tpm.rs file used -d as the
short form of --hexdata.

This CL standardizes on -d as the short form of --hexdata, eliminating
the conflict.  The immediate need for this CL is that somehow, the
explicit declarations of long-form only --help cases errors about
"Required argument: help".  And rather than figuring our how to properly
declare --help as optional, I think it makes sense to standardize on a
short form that does not conflict with -h.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I78363d2e61708391852ae19a6463fd086a66439b