[otbn,dv] Wrap some secure wipe assertions in a named block
The RTL change changes the hierarchical IDs of some assertions so that
they all start with a particular scope. Once that is done, we can turn
them off with

    $assertoff(0, "tb.dut.g_secure_wipe_assertions");

This is needed to avoid some spurious test failures in the
otbn_ctrl_redun test.

Although this changes some RTL code (hw/ip/otbn/rtl/otbn.sv), it
should be completely safe: the only change is to the hierarchical IDs
of some assertion statements.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>