[util] Add test for libclang macro expansion quirk
The parent commit apparently broke the sec_mmio audit tool[^0] by using
a macro that creates a call site of `sec_mmio_write32()`.

This seems to be a deficiency of the `libclang` Python bindings[^1]
rather than our sec_mmio audit tool. To demonstrate the issue, this
commit adds a Python unit test that reproduces the issue while only
exercising libclang.

What happens is that we obtain a cursor to the macro-expanded call site,
but cannot obtain the "spelling" of preprocessor-generated tokens.
Rather, the Python libclang bindings return a span of tokens from the
original, non-preprocessed translation unit.

I've been unable to determine a workaround. Perhaps this and future code
auditing tools should be implemented as Clang plugins written in C++.

[^0]: To reproduce the failure, run this command:
      `./bazelisk.sh run //util/py/scripts:audit_sec_mmio_calls`

[^1]: <https://libclang.readthedocs.io/en/latest/_modules/clang/cindex.html>

Issue #19438

Signed-off-by: Dan McArdle <dmcardle@opentitan.org>