[otbn] Fix interpretation of lc_escalate_en signal values
According to the lc_ctrl spec, all values of escalation signals other
than OFF must be interpreted as ON. Also, no detection of invalid values
should be performed as these signals might experience staggered
transitions due to CDCs naturally leading to invalid values.

This commit changes the design accordingly to match this spec.

This is related to lowRISC/OpenTitan#19050.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>