[opentitantool] Allow configuration of SPI chip select
With this change, and accompanying modification to HyperDebug firmware,
OpenTitanTool is able to use named chip select pins in conjunction with
the hard-wired SPI data and clock pins.

This can either be achieved by command line arguments, such as:
`opentitantool spi --bus=S2 --chip-select=IOA7 sfdp`

or though declaring named spi target devices:
```
"spi": [
  {
    "name": "BOOTSTRAP",
    "bits_per_sec": 8000000,
    "chip_select": "CN10_6",
    "alias_of": "QSPI"
  },
  {
    "name": "TPM",
    "bits_per_sec": 250000,
    "chip_select": "IOA7",
    "alias_of": "QSPI"
  }
],
```

In either case, the operator must ensure that prior to issuing SPI
transactions, the chosen chip select pin is in `PushPull` mode and high
level, (and that the data and clock pins are in `Alternate` mode, but
that is the HyperDebug power-on default).

https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4739792

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Iccdffe8954567593d42883f225decd24a2e0e6a5