[opentitantool] Remove fallbacks for ancient HyperDebug firmware
OpenTitanTool now ships with "official" HyperDebug firmware, is able to
upgrade the device, and warns if used with non-official firmware.

Given the above, I do not think that we need to keep carrying fallback
code meant for cases in which opentitantool is working with older
HyperDebug firmware, which does not understand newly added commands.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I3f432c082ba8abb533a9d6d8ad4958fc4efbea2a