[bazel] Rework module ID check aspect
Currently, the module ID checker aspect runs on every cc_binary
without looking at whether it is relevant. This works because we
only run it on precisely the right binaries. We want to enable
this aspect on every build so this commit changes the aspect to
only run on binaries that contain module ID information. This is
done by tracking whether a cc_binary depends on the status
library that introduces the module ID information.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>