[manuf] Fix JTAG problems in manufacturing tests
Several manufacturing tests suffer from random JTAG issues because
they try to connect over JTAG while the device is rebooting in a
loop. Since a chip reset kills the JTAG, it can make the test fail
when a reboot occurs during the JTAG session. This is particular
problematic for the trigger_lc_transition() function that performs
an LC transition, a reboot and attempts to connect over JTAG: it
can happen that the initial LC state has CPU execution disabled
but that the target test has execution enabled. When this happens,
and unless a program was bootstrapped that can prevent a reboot
loop, connecting over JTAG is unsafe.

This commit solves this issue by
- make trigger_lc_transition() only optionally reconnect over JTAG,
- carefully documenting for each trigger_lc_transition() usage
  what the execution situation is after the transition,
- fix the problematic ones.

There are two problematic tests:
- manuf_cp_device_info_flash_wr gets into a reboot loop after the
  transition: since the host already bootstraps a program that
  check that the transition was successful, simply remove the JTAG
  based test to avoid the program.
- personalize_functest: the device program go into a reboot loop
  after personalizing the device and can cause the JTAG connection
  to break during the entire LC transition (it's unsafe even before
  the transition). Fix this by adding a loop in the device code after
  personalizing to wait for the transition, and after the transition
  when the RMA state is detected.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>