[site] Tweak margins / thresholds to prevent nav-button collisions
The main cause of these problems was the introduction of the pagetoc,
in combination with some hardcoded values in the theme.

The main change was to bump the @media (max-width: 1380px -> 1720px) value,
which means the left-right nav buttons only appear when the page is wider, which
prevents the collisions with the content.

We also slightly tweak the theme of the bottom-of-the-page left/right nav
buttons to be more visible.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>