[opentitantool] Bugfix for legacy Titan rescue protocol
The code has an optimization to skip over ranges of 0xFFFFFFFF words
when programming.  Previously a `position(...).unwrap()` expression was
guaranteed to not panic, because the range to be programmed had
previously been trimmed, such that the last word was guaranteed to not
be 0xFFFFFFFF.  However, at some point, another issue required rounding
up the size of the image to be flashed to a multiple of the Titan flash
page size.  This could result in the final frame to be transmitted to be
all 0xFFFFFFFF words.  In order to avoid triggering panic in that case,
this change makes it so that the code will not attempt to skip over a
blank range, if it is at the very end of the entire (trimmed then
padded) range to be programmed.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: If373756129cd2c0b3bc0924d62cbd68bfdc289b1