[otbn,dv] Make update of URND _state field slightly clearer
This is a representation of a 256-bit number as four 64-bit numbers.
It gets updated "without crossover", in that the limbs don't
interact, except that it "rolls" across by one limb each time.

So

    A,B,C,D

becomes

    f(B),f(C),f(D),f(A).

The existing Python did this by updating an array at position "i+1"
each time and then copying the last limb back to the start.

The new code updates at position "1 + i mod 4" which is hopefully a
little easier to understand, and avoids the Python model having an
extra element in the array.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>