[opentitantool] Updated HyperDebug firmware
Pack a new precompiled firmware for the HyperDebug device.

This adds support for SPI1, and makes the blue button act as manual
reset of the GSC under test.

Also, as slight fix/improvement is to reset all SPI and I2C busses to
power-on default speed and chip select settings, on "transport init".

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Ie086f30df7a353171e40265cea4c6817591df9bc