[otbn,rtl] Rework imem fetch logic for branches
When a branch is in the prefetch stage we don't want to prefetch either
the taken or not-taken targets to avoid data dependent branch timing.
Rather than directly suppress the imem request in this scenario instead
kill the incoming response valid the following cycle. It is done this
way primarily for X prop reasons (see comments in commit changes).

Signed-off-by: Greg Chadwick <gac@lowrisc.org>