[otbn,rtl] Refactor loop control so commit doesn't factor into prefetch
Previously when a loop start was requested but not committed (due to an
error) this effected the next loop information outputs used by the
prefetcher to determine what to prefetch. This can cause timing issues.

With this change the next loop information doesn't depend upon the
commit (it provides what will happen if the commit happens) and breaks
the paths causing timing issues.

Signed-off-by: Greg Chadwick <gac@lowrisc.org>