[crypto] Set two highest bits of p and q in RSA keygen.
Setting these bits is allowed by FIPS 186-5 and lets us skip the range
check on p and q. This commit also removes the tests that ensure we
check the range, since now it's just a postcondition of
`generate_prime_candidate` and we never expect an out-of-range value in
`check_p` or `check_q`.

Signed-off-by: Jade Philipoom <jadep@google.com>