[bitstreams] Redirect missing targets in the cache
Add a list of known / required designs to the bitstreams workspace
generator, and for cache entries that do not contain all the designs,
redirect the targets back to the Vivado-built ones.

This eliminates errors that may happen during the analysis phase due to
missing referenced targets. Previously, if the bitstreams cache was
missing an entry, the targets would not be created in the @bitstreams
repo. Now, those targets will always exist, and builds may continue even
when the cache entry is incomplete.

Signed-off-by: Alexander Williams <awill@opentitan.org>