[bazel] Reorganize and simplify the bitstream targets
Several bitstreams targets in hw/bitstreans/hyperdebug are wrong.
Also the code is full of duplicate code because it independently
splices the vivado and GCP bitstreams and then selects based on
the bitstream config. Instead, consolidate the code by selecting
one bitstream early and then splicing this selected bitstream.
The only downside of this approach is that for vivado test_rom
bitstream, it will splice the test_rom into the vivado bitstream
that already contains the test_rom.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>