[opentitantool] Wrapper cleanup
With the introduction of LogicalSpiWrapper and LogicalI2cWrapper, the
setting of backend interface speed and other properties is done lazily,
once the first operation is requested through a particular logical
wrapper.  Because of that, there is no need to eagerly get a handle to
each SPI and I2C port and set any properties.

This CL skips this step, which anyway had no useful effect.  Not loading
backend I2C and SPI port has the advantage that opentitantool can in
fact apply default configuration to the GPIO pins of HyperDebug by
"transport init", even if e.g. OpenOCD has claimed the CMSIS-DAP
interface of the USB device.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I036e974c03923bc24607bcc8e52f1d762cfe9566