[sram_ctrl,sival] Add explicit tests for remaining testable features
This doesn't include SRAM_CTRL.REGWEN or SRAM_CTRL.INTEGRITY.

For the REGWEN side, this feature is provided by some templated code,
so I'm not really convinced that it's worth writing an explicit test
for.

For INTEGRITY, I don't think this feature is easily testable in
hardware without a dedicated lab. That will be covered by the
penetration testing phase, so I'm not sure that it makes sense to list
a sival test for it.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>