[rom_ctrl,sival] Describe an extra sival test to read DIGEST CSRs
This test seems worth doing: if we'd messed things up so that the
computed and expected digest matched but both read as 0, we'd manage
to start executing things, but we still want to know about it!

Obviously, this would also appear in CreatorRootKey, but that's harder
to reason about. Checking the digests explicitly seems like the
easiest approach.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>