[e2e] Fix rom_e2e_shutdown_watchdog
The test does not make sense: it asks the watchdog to shutdown after
2 seconds but the waits only 1 second before returning. Therefore the
shutdown nevers happens while waiting and the tests fails. This commit
fixes the test by setting the watchdog bite threshold at 1 second and
change the test to wait 2 seconds. This way, and with some margin,
it will always reset during the wait.

I also change the flash program to display the value of the watchdog
count after the loop. The idea is for the test to verify that in
certain LC states, the watchdog was actually disabled. Ideally, a
further test should be added to verify that in PROD, if we wait
only half a second, the value will be printed before the shutdown
and show the expected value.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>