[edn/dv] move coverage sample call to wr_cmd task
This commit moves all the coverage sample cmd calls to a single
location in the wr_cmd task. This helps getting a better overview
over from where we are sampling. Furthermore, it guarantees that
we are always sampling when the wr_cmd task is called. This wasn't
always the case before this commit.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>