[edn/dv] add wr_cmd args mode and cmd_src
This commit adds an additional argument to the wr_cmd task.
Namely, the operational mode. The commit also adds a new enum type for
the source reg of the command.
This provides the foundation for an extension of the csrng command
coverage, where the operational mode and the command source will be
added as additional cover points.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>