[usbdev] Switch to single port packet buffer
Switch to SPRAM implementation of the usbdev packet buffer.
Collisions are very infrequent and we can just delay the TL-UL access.
Retain the read data on the usbdev side because the packet
engine operates on bytes and relies upon rdata remaining static
throughout the use of the 32-bit read word.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>
Signed-off-by: Michael Schaffner <msf@opentitan.org>