[manuf] halt CPU after FT SRAM program executes
After the FT SRAM program executes, ROM execution will be enabled. In
order to transition to the PROD LC state via JTAG, the CPU needs to be
halted so that upon the host program can switch TAPs and issue the LC
transtion command to mission mode. Otherwise, if the chip resets, the
ROM will attempt to boot what is in flash at this point.

Signed-off-by: Tim Trippel <ttrippel@google.com>