[chip-test/sival] Add edn testutils library
This commit adds a edn testutils library which for now includes
the random command generation for auto mode. So far this has only
been used in the csrng_edn_concurrency_test but the same code will
be reused in the sw_edn_sw_mode test. Therefore, it might make sense
to add it to the edn testutils library.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>