[manuf] complete Secret1&2 and keymgr flash page provisioning
This partially addresses #19455 by completing the secret provisioning
step of the chip personalization process. Additionally this completes
the CreatorSwCfg provisioning (activating scrambling of the flash data
region). The personalization process was split into two flash binaries,
as this code size limit for a personalization flash binary is 64k, the
size limit of the ROM_EXT slot, and the current personalization /
individualization libraries are too as the make use of the DIFs and
test utils which are bloated with debug strings.

Signed-off-by: Tim Trippel <ttrippel@google.com>