[opentitantool] add sanity check to manifest updater
SRAM programs don't have manifests, but we currently are still signing
them with opentitantool.

This commit adds a sanity check so that OT-tool will fail when it is
asked to update manifest of a binary that doesn't have one. The bazel
rule is changed so SRAM programs are not signed.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>