[test] Use correct common IDs for UART and SPI instances
Some tests were still using "0" as the identifier for UART and SPI
devices, but these are not supported across all transports. The correct
identifiers are "console" for the UART console and "BOOTSTRAP" for Earl
Grey's spi_device IP. The transport will convert these to the correct
internal devices that interface with them.

Fix that up in this commit.

Also, remove an accidental double-init on the transport in another test.

Signed-off-by: Alexander Williams <awill@opentitan.org>