[dv, rv_core_ibex] Refactor the rstmgr_cpu_info_test
This commit simplify the test relying on the fix below:

After #15219 was merged, the execution stops more predictably
once fetch_en is dropped due to a double fault.
The current pc should now correspond to the instruction after the
cause of the illegal load or that same instruction, depending on
whether the pipeline stalled.
The next pc is always the exception handler, because that's
where execution would have gone if it had not halted.

Signed-off-by: Douglas Reis <doreis@lowrisc.org>