[bazel] make tests runnable with `bazel run`
Currently if you try to run tests with `bazel run`, you'll end up in a
state where the execution doesn't stop and you can't use Ctrl-C to kill
it.

It turns out it's due to we use `console` command in OT-tool which, when
stdin is TTY, switch it to raw mode and forward stdin to FPGA uart. This
is useful for debugging but not what we want for the tests.

This commit changes all tests to run in non-interactive mode.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>