[csrng/dv] Add checks for the fips bit
This commit adds checks for the fips bit. Whenever a new CSRNG item is
complete, the fips bit is commpared to the expected output, which is stored
in the cfg.compliance variable. This comparison happens every time a new
set of genbits are produced at the respective outputs of the CSRNG.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>