[dv] Specialize dv_spinwait_* documentation comments
The stuff that was above DV_SPINWAIT_EXIT was actually the
documentation that was originally written for DV_SPINWAIT. This didn't
get reworded and split when DV_SPINWAIT_EXIT was defined and then used
to implement DV_SPINWAIT in b9f4a23b67e.

Split it to keep everything consistent.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>