[dvsim] Get rid of "mname" field in Modes.py
This was only being used for the __str__ function (which we've just
dropped). Let's keep the typing information in the type itself (and
avoid implementing it in the runtime Python).

The only remaining use is in the constructor, where it appears as a
"type_name" argument to improve error messages if someone gets an
hjson file wrong.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>