[ot_certs] Add ASN1 builder trait
Prepare the ground work for ot_certs being able to generate ASN1
documents from the template. This commit adds the basic ASN1
definitions and the Builder trait that will be used to build ASN1
document. Contrary to normal asn1 serializers, the basic building
blocks of the builder are Value<T>. This means that implementation
of the trait that are able to handle variables (such as the
uppcoming asn1 code generator) can do so.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>