[ottool] Add command to substitute variables in certificates
This command allows one to take a template and a data substitution
file and produce a new template where some of the variables have
been replaced by the provided values. An example of such a file
for the UDS certificate is provided.
Also add the possibility to directly substitute variables in a template
when generatin a DER.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>