[ot_certs] Add library support for generating random test data
This data can be used to exercise every piece of code that works
with data, such as asn1 der generator and the asn1 codegen. More
generally it can be used to create random certificates for testing.
At the moment it generates random data but this could be extended
to generate edges cases as well.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>