[opentitantool] Fix omission in SPI/I2C alias handling
The JSON configuration format allows named aliases of SPI or I2C ports,
such that e.g. "TPM" can be an alias for e.g. "SPI1".  The same hardware
port can have multiple aliases.  A key point is that each alias can have
its own default speed and other settings, and the `TransportWrapper`
layer will make sure that before a port alias is used for communication,
its settings are applied to the physical transport port.

It seems that a bug had gone un-noticed, because of which the settings
were repeatedly applied if the same alias was used in a sequence of
transactions.  This could seem to be merely the cause for some slowdown,
but Alex discovered that when SPI chip select is to be kept asserted
across a series of separate interactions, re-setting the speed and
choice of chip select pin will interfere, by deasserting the chip
select.

This CL adds code that was intentede from the start, to properly keep
track of which alias has most recently been used to access any given
physical port, and only apply speed and other settings if it was not the
alias that currently is about to run a transaction.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I57b24523dff1fbc32feef854ea3bd4f07c4e3926