[util] Fix fpvgen outdir argument handling
This fixes the behaviour of fpvgen.py when it isn't given an -o
argument. The code tries to do something if there's nothing supplied
by checking "if not args.outdir". Unfortunately this doesn't work: the
argument has type Path and the "" default gets turned into the path
"." (which is not falsy).

Change to a default of None and everything works the way you'd expect.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>