[primgen] Fix parameters in a primgen template
It turns out that the existing code would give something of this form:

    module prim_foo #(param1, param2) (arg1, arg2);
      parameter prim_pkg::impl_e Impl = `PRIM_DEFAULT_IMPL;
    endmodule

Unfortunately, SystemVerilog doesn't actually allow you to declare two
lists of parameters like this. The presence of the #(param1, param2)
list means that the Impl line gets turned into a localparam.

JasperGold prints out a warning (telling the user that what they
expected to be a parameter is now a localparam). This patch changes
the generated code to match: Impl is now declared as a localparam, so
JasperGold needn't tell us!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>