[aes, pre_sca] Enable masking evaluation of AES with PROLEAD
This commit adds a couple of support files and a how to for evaluating
the masking employed inside AES together with the instantiated PRNG
using the PROLEAD tool.

The library file as well as the PROLEAD config file have been derived
from files based on variants in the PROLEAD repository and kindly shared
by @AeinRezaeiShahmirzadi.

This is related to lowRISC/OpenTitan#19091.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>