[opentitantool] Add missed forwarding to I2C wrappers
When I added new methods to the I2C trait to support HyperDebug in
device mode, I declared default implementation, in order to not have to
touch every other backend implementation.

It seems however, that I accidentally forgot about the wrapper layer in
`src/app`, which should be able to pass along any request to switch to
device mode, for the case the underlying transport supports it.  This CL
adds that support.

Change-Id: Ie8e92c6d8769660b430362999405836522ea5ccd
Signed-off-by: Jes B. Klinke <jbk@chromium.org>