[sival] Add sysrst_ctrl_in_irq_test test
This tests follows the DV flow with two modifications: to avoid
the various stages interfering with each other, add a sync step
at the end of each stage so the host does not move forward until
after the device code has seen and disabled the interrupt.
The second problem is that on real device, the interrupt could
occur between the phase change and the call to `wait_for_interrupt`.
If that's the case, `wait_for_interrupt` will block forever.
Modify the code so that the safely checks if the interrupt has
already been handled before calling WFI.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>