[edn/rtl] Change sw_cmd_sts register
This commit changes the sw_cmd_sts register. The register now contains
the following fields: cmd_reg_rdy, cmd_rdy, cmd_sts and cmd_ack. An
explanation for these fields can be found in the contents of this commit.
The command reg ready signal should be high when the EDN is ready to accept
new SW command data and the output FIFO is ready to accept data.
This commit should make it easier for firmware to check whether a command
has been acked by CSRNG. Now software shouldn't need to check the intr signal
any more and it should be possible to find out whether the ack has happened
and the sts is succesful with a single read.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>