[aes,pre_sca] Modify evaluation parameters for PROLEAD
With the new Bivium-based masking PRNG, the evaluation with PROLEAD
consumes a lot more memory and the memory consumption keeps growing
with increasing number of simulations in normal mode.

Therefore, this commit switches to the compact evaluation mode and
instead increases the number of simulations by roughly a factor of 10x
as recommended in the PROLEAD wiki. Also, the reported results are
updated to match what's achievable with the Bivium-based PRNG.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>