[edn] Move prim_edn_req out of prim
This package doesn't fit all that well in prim. Apart from anything
else, it depends on edn_pkg (which doesn't live in prim).

Move the code out of hw/ip/prim and put it in hw/ip/edn instead. The
module that generates the requests is renamed from prim_edn_req to
edn_requester (to avoid the "prim" in the name).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>