[edn/dv] Fix failing edn_genbits tests
This commit fixes a failing genbits test.
The test sequence was waiting for the SM to reach the Idle state but
the SM was already in SWPortMode. This caused a timeout. Now the
sequence waits for SWPortMode instead, which is effectively the same
since the SM has to go into Idle before it can enter the SWPortMode.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>