[RV_DM]rv_dm_scoreboard
This commit contains a change which add a condition which check for signal 'lc_hw_debug_en'. This change is for sequence 'rv_dm_sba_debug_disabled_vseq'(#21042).If this signal is Off then it would not compare sba items and throw it away. Because no transactions were seen on tl interface, so we do not want to compare. Otherwise if signal is on then sba items will be compared.

Signed-off-by: Basit Mehmood <basit.impare@gmail.com>