Revert "[edn] Move prim_edn_req out of prim"
This reverts commit 0f7d7fd90f942805bd498c14080c8615a0dd175f. The
reason is that this will change some paths in the hierarchy and we're
worried that this will cause slightly pointless work for someone with
CDC/RDC waivers.

Honestly, it seems a bit rubbish that this is something we need to
worry about! But we can make things self-contained in a less intrusive
way, so that's probably the right approach.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>