[dv/csr_utils] Change `csr_peek` to function
This procedure does not have to be a task because it does not control
time.  Thus it can be a function, which is easier to use because it can
be called from functions and not only from tasks -- and functions are
easier to use because they can have a return value.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>