[spi_device] Relax stringent CSB pulse width requirements
Add a new edge-to-toggle synchronizer + filter module for SPI to remove
the stringent CSB pulse width requirements that fell out of using
prim_edge_detector to detect CSB de-assertion.

spid_csb_sync produces a pulse in the sys_clk domain when a CSB-clocked
flop produces a value change, compared against the last seen value in
the sys_clk domain. The toggling flop input signal gets filtered based
on (1) whether any SPI clocks were observed and (2) whether the
toggle_en_i signal was asserted for one of those SPI clocks.

In addition to merely providing the notification that a transaction has
completed, this module is also used in the spid_upload module for
qualifying whether an uploaded command was written (for the interrupt
event).

Signed-off-by: Alexander Williams <awill@opentitan.org>