[spi_device] Harmonize address mode syncs to byte beats
Simplify address mode synchronization by using a common commit point at
the end of the command phase. Add a sync pulse for the address mode to
commit at the 8th posedge of the SPI clock. This allows enough time for
software-originated changes to cross into the SPI domain and be picked
up for the next command. In addition, there is no need to reset the
entire spi_device IP when changing address modes after a period of SPI
activity.

Harmonize address calculations to align with this commit point. The
passthrough and upload modules have slightly different mechanisms, but
they now use the address mode from this point. The passthrough module
uses the registered value, but the upload module uses the D side of that
flop instead. The upload module may need this information for some
future enhancements related to metadata committed to the command FIFO.

In addition, inform software when a software-originated address mode
change is still pending. This may occur if no transactions occurred
since the last software write. When a write is pending, the value shown
is the pending value, not the current committed one in hardware.
However, that pending value will be picked up in time for the next
command. If the desired value is different from the prior write,
software may overwrite, as long as the SPI host is known to be inactive.

Signed-off-by: Alexander Williams <awill@opentitan.org>