[rv_dm,dv] Add checks to randomization in rv_dm_tap_fsm_vseq
No functional change intended, but I hope this simplifies stuff a bit
by avoiding some repetition.

I *think* the only change will be in how we randomize the last req,
where we throw in a couple of extra constraints. But they were only
missing (I think) because the author didn't need them: the test is to
make sure that we can reset an FSM, and that stays.

The reason that I found this at all is that Xcelium complained we
weren't checking the return value from calls to req.randomize(). Now
we are!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>