[ipgen,clkmgr] Turn a few more files into templates and fix paths
The files are almost identical to the previous ones, except they use
${topname} instead of a specific name and consecutive '#' are escaped
so they are not considered md comments. This escape replaces things like
`###` by `${"###"}`.

Adjust some relative paths to account for the change in the file locations.

Use {self_dir} to simplify some paths: it points to the location of the hjson
file used to process the contents of a given file.

Part of #8440

Signed-off-by: Guillermo Maturana <maturana@opentitan.org>