[lc_ctrl] Always allow transitions into scrap
We should always allow transitions into scrap, no matter whether the
transition counter is maxed out or not.

In order to prevent any abuse of this unconditional transition path,
we max out the counter and program the LC state to SCRAP already in the
transition counter increment phase instead of programming the counter
state and LC state separately in two phases as all other transitions.

This fixes #13617.

Signed-off-by: Michael Schaffner <msf@opentitan.org>