[dv] Tweak outstanding access check in cip_base_vseq.sv
This change won't really make any difference to the behaviour, but it
will hopefully make it a bit clearer what's going on. The test that is
being moved was added by d7ba5ec. I think the idea is that we want to
fail quickly if a reset happens in the middle of a CSR operation,
because it's rather painful to debug what happens when things get out
of sync.

This commit is moving the check a little earlier again. Now, we fail
just *before* issuing the reset, and it should hopefully be easy to
understand what has gone wrong if this fails.

Note that it's not exactly equivalent. Before this change, the check
would fail if seq_wo_reset completed before we issued a reset at all,
and it accidentally left a CSR item outstanding. The new code won't
notice that happen (but I don't think it's something that we're
worried about seeing).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>