[prim_sha2] Make digest writable from input while disabled
As a side effect, the digest is no longer constantly cleared while the
module is disabled (i.e., `sha_en_i != 1'b1`) but only when
transitioning from enabled to disabled.  This allows to read the digest
out before disabling, then clearing it by disabling and loading a
different digest, which is a prerequirement for context switching.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>