[rv_dm,dv] Shuffle tasks in smoke vseq
No real change in what gets run, except that we no longer guarantee to
run everything at least once.

The shuffle isn't quite a complete shuffle, because the changes to
dmactive are still done last. This is because setting check_dmactive()
might disable the debug module, after which the module will ignore the
writes that the other tasks expect to have an effect.

One possible question is whether this shuffle has made it likely we
don't run a particular check. This shouldn't be an issue: the
probability of not running a particular test when we go through the
randcase N times is going to be 3/4^N. This already drops below 0.3%
with N >= 20.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>