[kmac] Move DOM multiplier control from keccak_2share to keccak_round
Already before this change, keccak_round was basically responsible for
the fine-grained control of the DOM multiplier input/output muxes but
this wasn't obvious leading to a much more complicated design.

Moving all these control signals up to keccak_round allows to simplify
the code and more importantly, it paves the way for registering some
critical signals to avoid glitches on the input/output muxed signals
which is beneficial for SCA hardening.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>