[usbdev] usbdev_enable_vseq
Adds the sequence usbdev_enable_vseq, which serves as a test for the functionality of the USB device control register field 'enable.' Enabling this field activates the USB device, initiating communication with the device responding with a DP_PULLUP signal. After setting this bit, transmitted a packet, and the device successfully received and responded with an ACK handshake packet.

Signed-off-by: Mubashir Saleem <mubashir.impare@gmail.com>