[csr_seq_lib] Avoid slicing a queue
Here, all_csrs is a queue and SystemVerilog doesn't technically allow
the part-select that we were using when we created test_csrs.

Fortunately, it does allow access by index, so we can sort of build
the subsequence by hand, which is what this commit does.

This avoids the *W,RTSVQR warning from Xcelium.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>