[rom_ctrl,dv] Start CSR accesses later in passthru_mem_tl_intg_err
I suspect that there will be several sequences which could use this
change. This is just meant to be an easy framework to implement the
same thing for other sequences.

The change in this commit avoids two timeouts:

(1) If you try to start a CSR access right after coming out of reset,
    you crash into a timeout in csr_spinwait. This timeout is
    default_spinwait_timeout_ns and is configured in csr_utils_pkg.sv
    with an initial value of 10ms.

(2) If you wait a little after coming out of reset to avoid *that*

    timeout, you crash into a timeout for the run phase (20ms). One
    solution would be to increase that timeout but that seems a bit
    tricky and I suspect it shows that we're using UVM in a slightly
    unconventional way here.

    A simpler bodge is to run fewer iterations of the test, so that's
    what I've done in this commit.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>