[top/dv] Make volatile RAW unlock test parameteric
Besides being able to test the volatile RAW unlock mechanism,
we should also be able to test whether the case where that
mechanism has been disabled in HW.

This patch adds a plusarg that allows to parameterize the test with our
expectation.

In production silicon this mechanism should be disabled, for example.

Signed-off-by: Michael Schaffner <msf@opentitan.org>