[ott] Remove UUCP-style locking for serial ports
We supported this for compatability with `minicom` which only uses UUCP
locks and not `flock`. This lockfile is left behind very often by OTT,
though we're not sure why. On shared machines, OTT is unable to clean up
stale locks because `/var/lock/` has a stick bit set which prevents
other users deleting lock files.

We think that the hassle this causes is not worth it to get locking
compatability with `minicom`. Other tools which support `flock`s are
recommended instead, such as `picocom`.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>