[entropy_src/dv] Fix a failing rng_vseq test
This commit fixes the rng_vseq test which was failing because the
ht_threshold_scope was set to an invalid mubi value in the config.
Whenever the entropy_src enters an alert state due to the configuration,
it reconfigures the DUT in a safe way to run the test. However, for the
ht_threshold_scope the same value is kept. The issue here is, that when
the initialization fails because of the ht_threshold_scope we still have
to set it to a safe value.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>