[cryptolib, kmac] Add functest to KMAC driver sideloading
Add a functest for KMAC sideloading. Because we do not
know the key that is being sideloaded, we are not able
to compute the digest as part of testvectors. Therefore,
this test runs the same operation twice and compares the
result of these two operations. In between the two runs,
we also run a SHA-3 operation to make sure that the
result of the second operation is actually computed
by KMAC engine.

Signed-off-by: Fatih Balli <fatihballi@google.com>