[dv,random_reset] Enhance handling of random resets
Declare the intention to issue a random reset in the near future in order to halt
transaction generators. This prevents occasional failures in stress_all_with_rand_reset
because of timing out waiting for no outstanding accesses, either because the generators
have too many pending transactions, or because of a single cycle with no outstanding
accesses followed by more accesses.

Fixes #21531

Signed-off-by: Guillermo Maturana <maturana@opentitan.org>