[sival, rv_dm] implement rv_dm_csr_rw test
The original testplan requires testing of every single register, which
requires big changes to reggen, will be slow to run over JTAG, and isn't
really necessary.

Instead, the test is implemented to test one register for each IP block.
This needs to be a RW register so that both read and write over debug
module is tested. I picked the IRQ enable register for most IP blocks.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>