[bazel] Rework JTAG handling
This commit unifies cw310_params and cw310/hyper310_jtag_params by
integrating JTAG support to cw310_params as extra parameters.
The OpenOCD adapter config file (if any) is moved to execution
environments.

The intention is that one can choose to enable JTAG without having
to clear the bitstream. For backward compatibility of the rules,
*_jtag_params are set to clear the bitstream.

If test_harness is specified, then `--bitstream={bitstream}` is automatically
added to test_cmd. This was the behaviour for cw310_jtag_params, but is
generally useful for all other tests.

Co-authored-by: Amaury Pouly <amaury.pouly@lowrisc.org>
Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>
Signed-off-by: Gary Guo <gary.guo@lowrisc.org>