[dv] Make mem_model's compare_byte function less chatty
This spits out quite an impressive ream of text, making
--verbosity=medium less useful than it might be.

Looking at the history of the file, I think this message got promoted
from UVM_HIGH to UVM_MEDIUM as part of commit 6a4c51b. I suspect this
was a debug print that survived by accident.

Demote it again!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>