[hmac] Bump version to 2.0.0 and reset development stages
With the addition of *save & restore* (PR #21307) and *wider digest and
configurable key length* (PR #21604), HMAC will see major changes.  For
this reason, this commit gives HMAC a major version bump to 2.0.0 and
resets its development stages to L1 (Development), D1 (Functional), V0
(*Initial Work* because the DV environment may not be completely
functional for some time), and S0 (*Initial Work* because the DIF may
not be completely functional for some time).

The previous version 1.0.0 got assigned a commit ID for future reference
even though it had not reached L2, V3, and S3.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>