[entropy_src] Fix CS AES Halt interface
Now that the SHA3 core features an interface to directly implement the
functionality to correctly sync with the AES core inside CSRNG in order
to avoid both of them becoming active at the same time, most of the
custom logic for this feature inside ENTROPY_SRC can be removed.

This resolves lowRISC/OpenTitan#17941.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>