[entropy_src/dv] Fix usage of rng_max_delay plusarg
Previously, upon reading the plusarg, the specified value was
immediately written to the configuration object of the RNG agent in the
environment config. But as it turns out, the entropy_src_env_cfg holding
the RNG agent which got the plusarg value is created before
entropy_src_env. During the creation of the latter, a new RNG agent
object is generated. This means the plusarg value got previously
overwritten.

This commit solves this issue by storing the plusarg value into a new
configuration variable that can be used in entropy_src_env to configure
the newly created RNG agent.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>