[dv,sram_ctrl] Improve detection of init_done and key_scr_valid
The older scheme for this detection used csr_spinwait for the duration of
init and key request. This only uses csr_spinwait once the request is
satisfied, to detect the actual CSR updates. This is just as accurate as
the older scheme, and by avoiding csr_spinwait for the full time it
simplifies handling of random reset.

Signed-off-by: Guillermo Maturana <maturana@opentitan.org>