[spi_device,dv] Split sim config into separate 1r1w and 2p versions
This follows the scheme used for the aes block (where we select
between masked/unmasked implementations).

The "implementation" is that we pass a define (SRAM_TYPE). This
essentially does the same job as the SRAM_TYPE parameter that was in
spi_device_env_pkg.

But now we have a pair of _sim_cfg.hjson files that implement the two
modes. This is equivalent to aes_masked_sim_cfg.hjson /
aes_unmasked_sim_cfg.hjson.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>