[entropy_src/dv] Test whether noise source symbols are not dropped
This commit adds changes to checks for comparing entropy in the observe FIFO to
the actual noise from the noise source. The scoreboard now knows when pushed
into the observe FIFO is dropped and when it isnt.
This commit also revives a vseq that was unused and changes some settings.
Furthermore, the task reading entropy from the observe FIFO now has optional
checks to make sure that the entropy is contiguous.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>