[usbdev,dv] Sort sequences (mostly) alphabetically
There is no functional change here: we're just listing the sequences
alphabetically in usbdev_vseq_list.sv and in usbdev_sim_cfg.hjson.
This avoids an annoying situation where every PR is trying to add
something at the bottom of the list and thus conflicts.

Now, we can insert a new PR at its point in the alphabetical list and
should hopefully have simple merges most of the time.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>