[opentitantool] Update hyperdebug_teacup.json to match hardware
It seems that on the Teacup board IOB11 is routed to CN9.21 and IOB12 is
routed to CN9.19, the opposite of how there were declared.  As far as I
recall, the two signals were swapped at some point during Teacup design
review, in order for HyperDebug I2C capabilities to match how ChromeOS
intends to use the pins of OpenTitan.

Change-Id: I6292dedfccbb5f69c15c7a81afd75d3d74080724
Signed-off-by: Jes B. Klinke <jbk@chromium.org>