[dv,usbdev] Smoke test improvements for V1
Ensure that SETUP, OUT and IN data packets are transmitted
and checked at the receiver for a random endpoint, exercising
both transmission and reception of packets at both the CSR
side and the USB side.
Crude initial implementation of full packet decoding to check
the contents of IN data packets; further validation is required
and in particular the calculation and checking of CRC16.
Implement the removal of bit stuffed '0' bits.
Introduce workaround for the sampling issues that arise from
operating on a clock frequency that matches the bit rate; the
DV should preferably be employing an oversampled clock and not
stealing access to the USB device driver enable output; this
would then permit testing of frequency and phase mismatches
and operation using just the true USB signals, for example.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>