[usbdev,dv] Change name of rand_or_not arg to be more descriptive
This controls whether randomisation of m_data_pkt obeys the
num_of_bytes argument. We're randomising the contents either way, so
the existing name was a little unclear.

This also moves the if/else block inside the constraint and uses the
DV_CHECK_RANDOMIZE_WITH_FATAL macro instead of calling assert()
directly.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>