[i2c,rtl] Count number of NACKs sent
This is useful for software to know how many NACKs have happened since
the ACQ FIFO ended up being full. Otherwise there is no way for software
to know whether it was on time with emptying the FIFO or whether it just
missed the one byte or whether it missed multiple transactions.

Signed-off-by: Marno van der Maas <mvdmaas+git@lowrisc.org>