[usbdev,dv] Remove num_of_bytes from base vseq
It turns out that we almost never need the sequence to have an opinion
about num_of_bytes. The only time it matters is when we send an OUT
packet to populate a buffer and then try to read the data back out
again with an IN packet.

In those situations, we pick up the correct num_of_bytes value for
configuring the IN packet by looking at the data packet that we just
sent (see usbdev_pkt_sent_vseq for an example).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>