[csrng/rtl] Change and assign cmd_sts signal
This commit changes the sw_cmd_sts and hw_cmd_sts signals to multiple bits.
0 represents a successful acknowledgement, whereas any value other than 0
represents an error status.
The sts signal is now also able to anounce to the EDN that the CSRNG main
SM is hanging due to an invalid acmd.
This commit also aligns the EDN rtl with the new sts response.
Furthermore a new csrng interface pkg is created to avoid circular dependencies.

This commit also aligns the EDN and CSRNG DV, documentation and SW with
the new STS signal.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>