[uart] Reduce TX FIFO depth from 128B to 32B
The UART FIFOs had been increased from 32B to 128B for Earlgrey ES,
predominantly due to the fact that there was a requirement for 128B deep
RX FIFOs. Such a requirement does however not exist for the TX FIFO and
it can be just 32B deep.

Since the design only supported a homogeneous arrangement, both RX and
TX FIFOs had been set to the same depth. This patch reworks that and
makes both depths individually configurable.

Signed-off-by: Michael Schaffner <msf@opentitan.org>