[entropy_src] Rework the swread packer FIFO to reduce area
Previously, the swread packer FIFO for unpacking the 384-bit outputs of
the esfinal FIFO into 32-bit chunks readable by software was implemented
using a prim_packer_fifo module instantiating 384 FFs internally for
buffering the seed.

However, as software can only read that output when hardware doesn't get
any entropy from the esfinal FIFO, there is no need for this big buffer
and the design can be changed to read the output of the esfinal FIFO
directly. This is what this commit changes.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>