[i2c,rtl] Reduce NACK count to 8 bits
Originally this NACK count was 32 bits which is not really necessary.
With 8 bits the software can see up to the depth of the ACQ FIFO worth
of NACKs, beyond that they lose their meaning. The counter should
saturate so that when it reads 0xFF that the software knows it has
exceeded the counter.

Signed-off-by: Marno van der Maas <mvdmaas+git@lowrisc.org>