[otbn, rtl] Squash call stack read data to 0 when stack is empty
Previous to this change we were seeing X prop issues where a branch
happened using a call stack as one or both of the compare operands when
the stack was empty. Whilst OTBN raises an error due to the use of the
empty stack the branch result goes to X (as the comparison inputs are
X) which produces Xs on the imem request interface.

With this change output from the stack gets squashed to 0 when its
empty.

Signed-off-by: Greg Chadwick <gac@lowrisc.org>