Revert "[dv] Remove phase argument from monitor's collect_trans"
This reverts commit 6a7357f44cc74ed47b2bfe423ff301fb90eec1a7.

The change was reasonable, but it touched some vendored code and some
unrelated CI breakage meant that I didn't spot the CI check that would
have told me about it. Oops!

The easiest way to fix this properly is probably to implement the same
change on the Ibex side first, so let's back out the change at this
end for now.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>