[python] Fix FuseSoC version
Prior to this commit, the FuseSoC version downloaded when running

    pip install -r python-requirements.txt

would identify itself as **version `0.4.dev0`** to Pip as well as in
`fusesoc --version`.  This is problematic in at least two ways:

1. Because Pip gets told that it downloaded version 0.4, it can prefer
   to use a locally cached archive (and I've seen it doing that!) that
   is *really* version 0.4, causing the update to become a NOP even
   though the user thinks they just updated FuseSoC.

2. Users have no way of distinguishing between versions 0.4 and 0.5 from
   the command line, which makes debugging issues that got fixed in
   version 0.5 difficult.

To work around this problem, the PR lowRISC/fusesoc#6 changed
`fallback_version`, from which our FuseSoC gets its version, to
`0.5.dev0`, and this version was subsequently tagged as such.  This
commit now updates our Python requirements to load the fixed version.

Run `pip install -r python-requirements.txt` after this commit has
landed to update FuseSoC to version `0.5.dev0` for good.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>