[i2c] Split the target and controller FSMs
Split the target and controller FSMs both to simplify the code for
better maintainability / readability and to lay the groundwork for
multi-controller support. Note that there will likely be some increase
in area for duplicated structures here, but it turns out that much of
the state was already independent.

Also fix a bug where the controller-mode stretch counter would continue
counting cycles even when the target had already released SCL.

Signed-off-by: Alexander Williams <awill@opentitan.org>