[opentitantool] Improvements to C2D2 debugger support
The C2D2 debugger is used by ChromeOS for communicating with the
security chip among other things.  It offers basically only access to
the RESET signal and the two UART lines.  This is enough to perform
rescue firmware updating, or other operations.

It turns out that the default configuration file for C2D2 had not been
updated with proper RESET strapping, and also instead declared empty
strappings, which is confusing as it leads to silent errors.

This change properly declares the RESET strapping, and removes the
ROM_BOOTSTRAP strapping (as C2D2 does not have wires to enter bootstrap
mode.)

Change-Id: Ibe271e92ce378d70e822e9a01a3469e6f763e572
Signed-off-by: Jes B. Klinke <jbk@chromium.org>