[dvsim] Split and rename Modes.py
The capitalised filenames were named after the classes they
defined (confusingly, I renamed "Modes" to "Mode" in 57090470d0).
Also, this file defined lots of different classes.

This commit pulls out the more interesting subclasses, keeping with
the CapitalisedName style and then renames the remaining file to
modes.py so that we're no longer claiming that it contains a class
called Modes.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>