[prim,fpv] Tidy up and document some FPV macros
No functional change. The only change to the code is that
ASSERT_ERROR_TRIGGER_ALERT can be defined in terms of
ASSERT_ERROR_TRIGGER_ERR, so I've done that to avoid the duplicated
code.

The main change is a few lines of documentation comment that explain
exactly what the macros are actually doing.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>