[sram_ctrl,fpv] Fix incorrect assertion
This assertion is supposed to claim that if the regs FSM generates an
error then this error will make it to top-level in sram_ctrl.

The code had a bug: it actually asserted that alert_tx_o[0] would be
nonzero. This is definitely true (because alert_tx_o[0] is
differentially encoded) but isn't very interesting. Switch to use the
correct macro.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>