[sival] Fix `entropy_src_csrng_test`
The way the test handles EDN1 do not really make sense: it constantly
generates new entropy on both EDN0 and EDN1 but OTBN is only the user
of EDN1 so after the last OTBN request for entropy on the EDN1, a new
generate command will be queued but the last command's generated
entropy (generated after the end of the test) will never be used so
the command will queue forever and the test will hang when waiting
for an IRQ.

This commit modifies the test so that it generates *exactly* the
required amount of entropy on the EDN1. I think this is what the test
intended from the beginning but the implementation did not match.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>
(cherry picked from commit 93164c981d3f0cac3292c41371c0dc2279539118)