[prim_ram_1p_scr] Align documentation with actual implementation
The S&P network on the data got removed in lowRISC/OpenTitan#20855 as
it negatively impacts the end-to-end integrity scheme but the PR forgot
to update the documentation of the scrambling primitive (The SRAM_CTRL
documentation got updated though).

Since the additional diffusion might still be relevant for other
implementations, it's worth to keep the documentation but explain that
it is not enabled by default.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>