[otbn,dv] Shorten wait for core errors in otbn_ctrl_redun_vseq
This is intended to be a short wait before we inject the error into
the core. It turns out that many of our test binaries complete in far
fewer than 1000 cycles, which means that the test fails for silly
reasons (we don't inject any errors before the end).

There's no real reason to wait so long, so cut the random wait down by
a factor of 10.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>