[i2c] Add clock low / SMBus bus timeout feature
Add an option to switch the main TIMEOUT_CTRL from a "stretch timeout"
to a general "bus timeout." Since they are the same value in SMBus,
these timeouts do not need to be active simultaneously.

A stretch timeout only triggers when this IP's controller has released
SCL, but it doesn't reach a high state within the duration specified by
the CSR. By contrast, a bus timeout triggers whenever SCL is held low
continuously for the duration specified by the CSR, no matter the
source. A bus timeout affects both the target module and the controller
module, and transactions are aborted with error/NACK reporting.

Signed-off-by: Alexander Williams <awill@opentitan.org>