[fpv] Add an "after load" hook
This is a TCL file that should run after we have loaded up the design
adn the testbench. The idea is that we can supply this in the hjson
file (e.g. top_earlgrey_fpv_prim_cfgs.hjson).

This is a bit ugly: it would be much nicer to do this through fusesoc!
Unfortunately, in the FPV flow we're just using fusesoc as a tool to
collect up a bunch of .sv files (and it thinks it's using Icarus
Verilog). I couldn't see a nice way to add a TCL file there, nor could
I see a way to give that path to the TCL code that is getting run in
fpv.tcl.

There's no doubt a clean solution, but I suspect that will have to
wait until we jump to upstream fusesoc, and that won't happen until
after we've branched for the A1 chip.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>