[i2c] Require nonzero mask for address match
Adjust the address CSR so a given field can be disabled with an address
mask of zero. Only accept target address designations if at least one
bit in the mask is set.

If a user wishes to be the target of *all* I2C traffic, that user can
set the same single bit for each mask, then use all zeroes for one
address and all ones for the other. (Technically, only that single bit
is meaningful, but "all ones" covers whichever bit is selected in the
mask.)

Signed-off-by: Alexander Williams <awill@opentitan.org>