[prim_sha2,rtl/dv] Fix secret value wiping
This fixes wiping of internal values such that they
are overwritten with wipe_v value, rather than XORed
with it. This makes this feature consistent with how
the wiping of the HMAC secret key works. It also updates
the description and documentation of the
secure wiping feature and the wipe_secret register,
and fixes a TLT with the corresponding change.

Signed-off-by: Ghada Dessouky <gdessouky@google.com>