[kmac/rtl] Move `err_processed` from CFG_SHADOWED to CMD CSR
The `CFG_SHADOWED` CSR gets write-locked while the SHA3 module is
operating.  This makes sense for configuration values but it is
problematic for the `err_processed` bit, through which SW acknowledges
that it has handled an error occurring while KMAC/SHA3 is operating,
because that acknowledgment would then not propagate to KMAC and it
would get stuck.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>