[csrng/dv] Fix forcing of ack_sts signals in csrng_intr vseq
lowRISC/OpenTitan#21981 reworked the status signaling of CSRNG and
changed the definition of the ack_sts signal from 1-bit to a multi-bit
enum which broke the csrng_intr vseq trying to force the ack and ack_sts
signals. This commit gets the sequence working again until it's reworked
to properly test the error signaling without forcing signals.

This is related to lowRISC/OpenTitan#22869.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>