[hmac,dv] DV synchronization fixes for do_read_check
This fixes DV synchronization checks for SHA-2 256
such that all CSRs and predictions match. This also fixes
FIFO checks according to the FIFO changes implemented to adapt to
multiple digest sizes, and disables do_read_check to be able to
release the constraints on digest size and key length
until synchronization issues are fixed for 384 and 512.

Signed-off-by: Ghada Dessouky <gdessouky@google.com>