[opentitantool] Claim USB interface before using
When establihsing connection to a HyperDebug device, opentitantool will
send a request on the CMSIS-DAP USB interface to ask about the
capabilities of HyperDebug.  If the USB interface is later used for GPIO
bitbanging or similar, opentitantool will "claim" it before doing actual
communication.  We want to let openocd have the option of claiming the
interface as well, for JTAG debugging.

The querying of capabilities leads to a kernel warning about
opentitantool accessing an interface without claiming it first.  This CL
adds code to briefly claim then release the interface, on startup.  This
still allows openocd to later establish connection for JTAG, but it
could interfere with the ability to invoke opentitantool while JTAG
debugging is ongoing.  (I do not think it could have worked before, as
the kernel would probably have refused opentitantool's request to send
bulk data to an interface already claimed by another process.  And even
if not, the USB protocol has no way of multiplexing requests from
multiple sources.)

Change-Id: I5c4ec0f1aa09b668a7fd5597b4f801dbaf137dd4
Signed-off-by: Jes B. Klinke <jbk@chromium.org>