[keymgr/rtl] Abort into Invalid when Creator Root Key not valid
Prior to this commit, keymgr could be initialized and would open itself
to operations even if the Creator Root Key shares provided by OTP in the
`root_key_i` input were invalid (although keymgr did write all-zeros and
all-ones instead of the input from OTP into its key shares if that was
the case).  However, the specification dictates that kemygr must signal
an invalid operation, though it did not specify the next state.

This commit fixes the problem by letting keymgr raise an *invalid
operation* error in this case and then transition to the terminal
Invalid state (going through a wipe state to clear any secrets with
entropy).  The rationale for this implementation is that the Creator
Root Key is meant to be static after reset, so SW doesn't have to be
able to retry the initialization.  Letting keymgr abort into a terminal
state thus is an additional HW security countermeasure to deter FI
attacks aimed at loading a manipulated Creator Root Key.

This closes #17776.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>