[rom] Change the size of the key pair field in SPHINCS+ addresses.
This change does not affect the sphincsplus-shake-128s parameter set we
use, since the key pair field is less than two bytes for all of the
parameter sets that were submitted to the NIST competition. However,
some choices of parameters need a larger field, so for maximum
flexibility it makes sense to adjust. This change does not have any
significant effect on performance, since these copies/writes are not on
the hot path (checked this locally as well).

Corresponds to commit 7ec789a from the SPHINCS+ reference
implementation (PR 60):
https://github.com/sphincs/sphincsplus/commit/7ec789ace6874d875f4bb84cb61b81155398167e

Signed-off-by: Jade Philipoom <jadep@zerorisc.com>