[kmac/rtl] Fix assertions on key stability
When `kmac_app` aborts an operation as soon as the sideload key becomes
invalid (which it will do in a couple of commits) the current stability
assertions on `key_len_i` and `key_data_i` in `kmac_core` would be
violated.  This is not a problem: `kmac_core` *should* use an invalid,
useless key in that case.  To prevent these assertions from failing,
this commit introduces a signal with which `kmac_app` indicates to
`kmac_core` whether `key_len_i` and `key_data_i` are valid and rewrites
the assertions to take that valid signal into account.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>