[rv_dm,dv] Rewrite dtm_idle_hint_vseq to use DMI helpers
This seems less complicated than the code that was there before, which
had compiled the various DMI operations into their underlying JTAG
writes without comments. Undoing this means I'm getting steadily
better with a hex calculator...

There shouldn't be any functional change here, except that we're now
using the DMI frontdoor driver, which is probably a bit more robust
than what was there before.

Rather confusingly, this is somewhat reverting the change that landed
in this file as a394a85. That change had an identical commit message
to the previous change to the file and there wasn't really a proper
discussion on the pull request (#20349). I'm not entirely sure why
Basit thought that change was required, but I propose to undo things
and revisit it if necessary.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>