[sram_ctrl] Add readback feature
When enabled with the SRAM_CTRL.READBACK CSR, the SRAM readback mode
checks each read and write to the SRAM.

On a read, the readback mode issues a second read to the same address
to check, whether the correct data was fetched from memory. On a write,
the readback mode issues a read to check, whether the data was actually
written into the memory. To avoid that the holding register is read,
the readback is delayed by one cycle.

On a mismatch, a fatal error is triggered and the STATUS.READBACK_ERROR
register is set.

To avoid RTL changes as much as possible, the readback mode is
implemented in the tlul_sram_byte module. In this module, after the
initial read or write was executed, the bus interface to the host is
stalled and the readback is performed. Due to the bus stalling, a
performance impact is expected.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>
Co-authored-by: Greg Chadwick <gac@lowrisc.org>