[prim_diff_decode] Use `prim_xnor2` to detect integrity issue
The differential `diff_pi` and `diff_ni` should not have equal values;
otherwise a signal integrity issue exists.  Such issues can be detected
with an XNOR.  Prior to this commit, this was implemented with
behavioral RTL code, which could allow some synthesis tools to perform
undesired optimizations for the complementary differential signals.

This commit changes the implementation to a `prim_xnor2` instance, so
that synthesis tools cannot change that logic.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>