[manuf] reduce UDS cert size and cleanup all cert sizes
This removes the HW_CFG0 OTP partition measurement from the UDS cert, as
this partition only holds the device ID, which is already mixed into the
keyladder directly be hardware channels.

Additionally, this updates the sizes of all the DICE/TPM certs and their
UJSON buffers.

Lastly, this fixed a bug in country portion of the subject field in the
TPM CIK certificate ("USA" --> "US").

Signed-off-by: Tim Trippel <ttrippel@google.com>