[topgen] Fix width of `dio_in_raw`
Instead of being hardcoded, the width of `dio_in_raw` is now generated
based on the number of dedicated pads.  This will mainly become relevant
in the next commit, which removes a dedicated pad.  It already fixes the
width of the signal on the CW310 and the CW340, although this width
mismatch was not problematic because the signal is an unused output of
`padring` anyway.  (It's correct but also unused in
`chip_earlgrey_asic`, but there it's relevant because the width mismatch
causes a lint error.)

Signed-off-by: Andreas Kurth <adk@lowrisc.org>