[usb,sival] Make the USB harness more robust to error
Previously, the harness would only try to open each device once and
ignore them if they fail. However, in CI it can happen that the
permissions initially prevent the harness from opening the device
and are later changed to allow it. In this case, it is essential
to retry failed devices.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>