[jtag,dv] Move host_cb back to negedge of tck
This clocking block was moved to the negedge of tck in 2022 with
commit d4e84bafaa5. The reasoning there was it fixed an rv_dm issue on
the closed source side.

Two days later, commit 299daf2accb moved back to the posedge. The
message for that commit doesn't discuss the change, so I have a strong
suspicion that it was a rebasing issue. This commit also didn't remove
the documentation comment on the line above that still claims we're
using the negedge!

This commit moves us back to negedge. The logic is that we can then
deal better with a situation where we want to drive a value to TDI
when the clock is currently stopped. The jtag driver might have
something like:

    `HOST_CB.tdi <= tdi;
    @(`HOST_CB);

Suppose the JTAG clock is currently paused (with tck_en=0) and tck is
high. When host_cb is on the posedge of tck, the write gets missed by
the design code because the write sits waiting in the clocking block
until the posedge and the design ends up sampling the previous value.
The sequence of events is:

  - Clock pauses for a bit (tck=1)
  - Clock is enabled.
  - We write to TDI through the clocking block
  - tck=0
  - tck=1
  - Design samples the value that's currently on TDI
  - The write makes it through the clocking block

Moving the driver to the negedge ensures that this "write just after
unpausing TCK" gets seen correctly by the design, because we end up
with the following sequence of events:

  - Clock pauses for a bit (tck=1)
  - Clock is enabled.
  - We write to TDI through the clocking block
  - tck=0
  - The write makes it through the clocking block
  - tck=1
  - Design samples the value that's currently on TDI

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>