[sival] Clarify SiVal steps of flash_ctrl LC tests
In DV we can transition the chip and check these signals. This is not
possible in SiVal. We will still have a set of chips in these LC states
that we check the accessibilities of the OTP partitions from without
having to do the transition.

Also updates the `lc_states` to the steps specified by the plan as
needing to be checked.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>
(cherry picked from commit cf5c321ecd6b554d9a211b452a21d285048a106e)