[provisioning] support signing certificates with Cloud KMS
Add a certificate singing option using the private key stored in Google
Cloud KMS. In this case TBS hashing ans signing is done by openssl. A
certain environment setup is required for the openssl to be able to
communicate with Cloud KMS, which is left beyond the scope of this
patch.

By default the provisioning test flow would still be using the fake
key and CA certificate, to enable Cloud KMS signing and verification
during the test the operator needs to add '--ckms_cert_endorsement' to
the test invocation command line.

Tested by running end to end test on the FPGA board for both cases,
fake and Cloud KMS based keys endorsement.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>