[edn/dv] Fix some edge cases for edn_disable_auto_req_mode_vseq
edn_disable_auto_req_mode_vseq was failing due to some edge cases.
One of the edge cases is when the sw_cmd_sts is read right after disabling
and the hasn't had time yet to update the register to all 0s.

Another edge case is when EDN is disabled right after a cmd_req_done interrupt
happened and the vseq hadn't had the time yet to clear the interrupt.

This commit fixes both of those edge cases.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>