[pinmux,rtl] Tweak pinmux.hjson to match RTL
The RTL code was copied from Earlgrey with commit 531ce0c and the
hjson file no longer matched. I think the original intention with this
bit of the hw/ip vs. hw/top_earlgrey divide was that the stuff in
hw/ip/pinmux was a sort of template from which we derived versions for
the different tops.

There's a reasonably long-term effort to switch to generating this
block with ipgen (tracked as issue 8440), and doing so will realise
that dream! In the meantime, we've got some tests that we want to do
for Earlgrey that are currently run from the code in hw/ip/pinmux.

Tweak the numbers in the hjson so that these tests are testing the
only chip we are currently building.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>