[hmac/rtl] Fix idle_o when told to stop
Prior to this commit, `hmac_core` and thus the entire `hmac` module
would not raise the `idle_o` output when it got told to stop (#23036).
This is problematic because then HMAC cannot be clock-gated during
context switch (save & restore feature).  Clock gating would only work
when a message stream is fully processed (after the *process* command).

This commit fixes the problem by keeping track of idleness inside
`hmac_core`.  In addition to the conditions that existed before, this
tracker goes to idle when a *stop* command is received and the current
message block is completely processed.

This fixes #23036.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>