[spi_device] Add CSB reset sync for outclk domain
For flops that latch on the falling edge of SCK, an unsafe reset timing
can occur if SCK is permitted to toggle while CSB toggles. SPI NOR flash
datasheets typically only specify setup and hold requirements around the
rising edge of SCK, so these flops are not protected.

Add a synchronizing flop that releases the CSB reset on the rising edge
of SCK, and create a separate reset network for the outclk domain to fix
the RDC issue.

Signed-off-by: Alexander Williams <awill@opentitan.org>