[entropy_testutils] Add function to stop CSRNG and EDNs only
For some test scenarios, it's better to leave ENTROPY_SRC turned on
and only re-configure CSRNG and the EDNs. Leaving it turned on, allows
to produced seeds waiting in the esfinal FIFO to be used in the
following test iterations. This speeds up the testing allows reseeding
multiple CSRNG instances within a couple of cycles, generating higher
CSRNG load in general.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>